% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundances.R
\name{abundances}
\alias{abundances}
\alias{ab,}
\alias{otu}
\title{Abundance Matrix from Phyloseq}
\usage{
abundances(x, transform = "identity", assay.type = NULL)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{transform}{Transformation to apply. The options include:
'compositional' (ie relative abundance), 'Z', 'log10', 'log10p',
'hellinger', 'identity', 'clr', 'alr', 'rclr' or any method from the
vegan::decostand function.}

\item{assay.type}{Name of the assay to pick when \code{x} is a
SummarizedExperiment-derived object. Defaults to \code{NULL}, which
selects the \code{counts} assay when present and otherwise the first
assay. Ignored for phyloseq objects.}
}
\value{
Abundance matrix (OTU x samples).
}
\description{
Retrieves the taxon abundance table from a
phyloseq-class or SummarizedExperiment-derived object (including
TreeSummarizedExperiment) and ensures it is systematically returned as
taxa x samples matrix.
}
\examples{
data(dietswap)
a <- abundances(dietswap)
# b <- abundances(dietswap, transform='compositional')
}
\references{
See citation('microbiome')
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
