\name{add.hclust.plots}
\alias{add.hclust.plots}
\title{Description: Calculate and plot hierarchical clustering for profiling script output}
\usage{
  add.hclust.plots(dat, data.dir, phylogeny)
}
\arguments{
  \item{dat}{oligoprofile data in original (non-log)
  domain}

  \item{data.dir}{data directory}

  \item{phylogeny}{phylogeny Returns:}
}
\value{
  hclust object
}
\description{
  Arguments:
}
\examples{
# dat <- read.profiling(params$wdir, "species", "rpa"); hc <- add.hclust.plots(dat)
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

