% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_refseq.R
\name{add_refseq}
\alias{add_refseq}
\title{Add \code{refseq} Slot for \code{dada2} based \code{phyloseq} Object}
\usage{
add_refseq(x, tag = "ASV")
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object with sequences as rownames.}

\item{tag}{Provide name for Ids, Default="ASV".}
}
\value{
\code{\link{phyloseq-class}} object
}
\description{
Utility to add refseq slot for \code{dada2} based 
             \code{phyloseq} Object. Here, the taxa_names which are unique 
             sequences, are stored in \code{refseq} slot of \code{phyloseq}.
             Sequence ids are converted to ids using tag option.
}
\examples{

# ps <- add_refseq(p0,tag="ASV")
# ps

}
\author{
Contact: Sudarshan A. Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{utilities}
