% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_taxa.R
\name{aggregate_taxa}
\alias{aggregate_taxa}
\title{Aggregate taxa}
\usage{
aggregate_taxa(pseq, level)
}
\arguments{
\item{pseq}{\code{\link{phyloseq-class}} object}

\item{level}{Aggregation level (from \code{colnames(tax_table(pseq))})}
}
\value{
Aggregated phyloseq object
}
\description{
Aggregate phyloseq data into higher phylogenetic level.
when the phylogenetic tree is missing. If the tree is available, 
uses the \code{\link{tax_glom}} function from the \pkg{phyloseq} package
}
\examples{
# aggregate_taxa(pseq, "L1")
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

