% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_top_taxa.R
\name{aggregate_top_taxa}
\alias{aggregate_top_taxa}
\title{Aggregate Top Taxa}
\usage{
aggregate_top_taxa(x, top, level)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{top}{Keep the top-n taxa, and merge the rest under the category
'Other'. Instead of top-n numeric this can also be a character vector
listing the groups to combine.}

\item{level}{Summarization level (from \code{rank_names(pseq)})}
}
\value{
\code{\link{phyloseq-class}} object
}
\description{
Summarize phyloseq: combine other than the most abundant taxa.
}
\references{
See citation('microbiome')
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
