% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancom.R
\name{ancom}
\alias{ancom}
\title{ANCOM Test for Abundance Differences}
\usage{
ancom(x, varname, multcorr_type = 2, sig = 0.05)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}}}

\item{varname}{Name of the metadata variable to be used for grouping}

\item{multcorr_type}{takes value 1 (stringent correction on the full pairwise matrix), OR 2 (within OTU/taxa multiple correction), OR 3 (no correction).}

\item{sig}{significance level (default 0.05)}
}
\value{
Vector of adjusted p-values, one for each taxon.
}
\description{
Perform ANCOM test on phyloseq data.
}
\details{
This function is directly modified from Weiss et al. (2017). The R code was obtained from the first author and included with permission in the microbiome package. For reference to the original ANCOM method by Mandal et al. (2015).
}
\examples{
  # Taxa that show significant differences between nationalities
  data(dietswap)
  s <- ancom(dietswap, "nationality")
  print(names(which(s < 0.05)))
}
\references{
Weiss et al. Normalization and microbial differential abundance strategies depend upon data characteristics. Microbiome 5:27, 2017.

Mandal S et al. Analysis of composition of microbiomes: a novel method for studying microbial composition. Microbial Ecology in Health and Disease, 26:1–7, 2015.

To cite the microbiome R package, see citation('microbiome')
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
