% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.documentation.R
\docType{data}
\name{atlas1006}
\alias{atlas1006}
\title{HITChip Atlas data (n = 1006)}
\format{The data set in \code{\link{phyloseq-class}} format.}
\usage{
data(atlas1006)
}
\description{
This data set contains genus-level microbiota profiling with HITChip for 1006 western adults with no reported health complications, reported in Lahti et al. (2014) \url{http://www.nature.com/ncomms/2014/140708/ncomms5344/full/ncomms5344.html}.
}
\details{
The data is also available for download from the Data Dryad repository \url{http://doi.org/10.5061/dryad.pk75d}.
}
\author{
Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
Lahti et al. Tipping elements of the human intestinal ecosystem. 
  Nature Communications 5:4344, 2014.
  To cite the microbiome R package, see citation('microbiome')
}
\keyword{data}

