% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rda.R
\name{Bagged.RDA.Feature.Selection}
\alias{Bagged.RDA.Feature.Selection}
\title{Description: BAGGED RDA feature selection}
\usage{
Bagged.RDA.Feature.Selection(X, Y, sig.thresh = 0.1, nboot = 1000,
  verbose = T)
}
\arguments{
\item{X}{a matrix, samples on columns, variables (bacteria) on rows.}

\item{Y}{vector with names(Y)=rownames(X), for example}

\item{sig.thresh}{signal p-value threshold, default 0.1}

\item{nboot}{Number of bootstrap iterations}

\item{verbose}{verbose

Returns:}
}
\value{
List with items:
  	     loadings: bagged loadings
	     scores: bagged scores
	     significance: significances of X variables
	     group.centers: group centers on latent space
	     bootstrapped: bootstrapped loadings
	     data: data set with non-significant components dropped out.
}
\description{
Arguments:
}
\examples{
#
#  library(microbiome)
#  data(peerj32)
#  x <- t(peerj32$microbes)
#  y <- factor(peerj32$meta$time); names(y) <- rownames(peerj32$meta)
# Bag.res <- Bagged.RDA.Feature.Selection(x, y, sig.thresh=0.05, nboot=100)
# PlotBaggedRDA(Bag.res, y)
}
\author{
Contact: Jarkko Salojarvi \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

