% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betadiversity.R
\name{betadiversity}
\alias{betadiversity}
\title{Beta Diversity}
\usage{
betadiversity(x, mode = "anticorrelation")
}
\arguments{
\item{x}{phyloseq object}

\item{mode}{Beta diversity method}
}
\value{
Vector with beta diversities; one for each sample, quantifying the
dissimilarity of the sample from the group-level mean
}
\description{
Quantification of beta diversity.
}
\details{
Beta diversity quantifies community similarity between
two samples. At the group level, it can be used to quantify group
variability (or spread). This function provides a wrapper that is handy
for group-level comparisons of beta diversity.

When the sample size n=2, this function calculates the beta diversity
between the two samples.

When the sample size n>2, it calculates beta
diversity of each sample against the group mean. This can be
compared between groups in order to compare differences in group
homogeneity. 

Note that this homogeneity measure is affected by sample size.
Subsampling or bootstrapping can be applied to equalize sample sizes
between comparisons.

More advanced beta correlation measures are available but not implemented
here yet.

The anticorrelation mode is a simple educational indicator that returns
average spearman correlation between samples of the input data and
the overall group-wise average.
}
\examples{
  # Example data
  library(microbiome)
  data(peerj32)
  # Assess beta diversity among the African samples
  # in a diet swap study
  b <- betadiversity(subset_samples(dietswap, group == "AFR"))
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
The inter- and intra-individual homogeneity in
Salonen et al. ISME J. 8:2218-30, 2014 are obtained as
1 - beta where beta is the "anticorrelation" beta diversity.
To cite this R package, see citation('microbiome')
}
\seealso{
the vegdist function from the \pkg{vegan} package provides many
standard beta diversity measures
}
\keyword{utilities}

