% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biomodality_scores.R
\name{bimodality_score}
\alias{bimodality_score}
\title{Calculate Biomodality scores}
\usage{
bimodality_score(x, method = ...)
}
\arguments{
\item{x}{Data (vector or matrix)}

\item{method}{Statistical test}
}
\value{
bimodality score
}
\description{
Calculate Biomodality scores
}
\details{
Calculate bimodality scores using, Hartigan's Dip Test Statistic,
Coeffefient of bimodality, Silvermans Test and Potential analysis.
}
\examples{
library(microbiome)
  bimodality_score(x, method= "bc")
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
Code modified from the original source:
  r-bloggers.com/measuring-associations-between-non-numeric-variables/
  To cite the microbiome R package, see citation('microbiome')
}
\keyword{bimodality}

