\name{bootstrap.microbes}
\alias{bootstrap.microbes}
\title{bootstrap.microbes}
\usage{
  bootstrap.microbes(D, Nsample, minPrev, Nboot = 10000,
    I.thr = 1.8, ncore = 22)
}
\arguments{
  \item{D}{data}

  \item{Nsample}{sample size}

  \item{minPrev}{minimum prevalence}

  \item{Nboot}{bootstrap sample size}

  \item{I.thr}{threshold}

  \item{ncore}{number of nodes for parallelization

  Returns:}
}
\value{
  TBA
}
\description{
  Description: bootstrap.microbes
}
\details{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

