% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calculate.hclust}
\alias{calculate.hclust}
\title{calculate.hclust}
\usage{
calculate.hclust(dat, method = "complete", metric = "pearson")
}
\arguments{
\item{dat}{data matrix (use log10 with pearson!)}

\item{method}{hierarchical clustering method (see ?hclust)}

\item{metric}{clustering metrics (spearman / pearson / euclidean)

Returns:}
}
\value{
hclust object for log10 and for absolute scale data
}
\description{
Description: Calculate hierarchical clustering for standard selections in profiling script
}
\details{
Arguments:
}
\examples{
data(peerj32);
  	     dat <- peerj32$microbes;
          hc <- calculate.hclust(dat, "complete", "pearson")
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

