\name{calculate.stability}
\alias{calculate.stability}
\title{Description: Stability analysis. Calculates average Pearson '}
\usage{
  calculate.stability(dat)
}
\arguments{
  \item{dat}{data matrix phylotypes vs. samples

  Returns:}
}
\value{
  List with correlations and astability estimate
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

