% Generated by roxygen2 (4.0.1): do not edit by hand
\name{centerData}
\alias{centerData}
\title{Center data matrix.}
\usage{
centerData(X, rm.na = TRUE, meanvalue = NULL)
}
\arguments{
\item{X}{The data set: samples x features. Each feature will be centered.}

\item{rm.na}{Ignore NAs.}

\item{meanvalue}{Can be used to set a desired center value. The default is
0.}
}
\value{
Centered data matrix.
}
\description{
Center data matrix to 0 for each variable by removing the means.
}
\note{
Note that the model assumes samples x features matrix, and centers
each feature.
}
\examples{
#centerData(X)
}
\author{
Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome").
}
\keyword{maths}
\keyword{utilities}

