\name{check.dependencies}
\alias{check.dependencies}
\title{Description: Load/install necessary packages and check OS}
\usage{
  check.dependencies()
}
\value{
  operating system string
}
\description{
  Arguments:
}
\details{
  Returns:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

