\name{check.wilcoxon}
\alias{check.wilcoxon}
\title{Description: TBA}
\usage{
  check.wilcoxon(...)
}
\arguments{
  \item{...}{TBA

  Returns:}
}
\value{
  TBA
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

