% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{check_anova}
\alias{check_anova}
\title{ANOVA for phyloseq}
\usage{
check_anova(x, group, p.adjust.method = "BH")
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{group}{Metadata field specifying the groups.}

\item{p.adjust.method}{p-value correction method for p.adjust function 
(default 'BH'). For other options, see ?p.adjust}
}
\value{
Corrected ANOVA p-values for multi-group comparison. Uncorrected post-hoc p-values for each pairwise comparison. Group-wise averages.
}
\description{
ANOVA test for multi-group comparison for phyloseq objects.
}
\examples{
  data(peerj32)
  pval <- check_anova(peerj32$phyloseq, "group")
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

