% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/foldchange.R
\name{check_foldchange}
\alias{check_foldchange}
\title{Calculate Log10 Fold Change for two-group comparison}
\usage{
check_foldchange(x, group, sort = FALSE, paired = FALSE)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object or a data matrix
(features x samples; eg. HITChip taxa vs. samples)}

\item{group}{Vector with specifying the groups}

\item{sort}{sort the results}

\item{paired}{Paired comparison (Default: FALSE)}
}
\value{
Fold change information for two-group comparison.
}
\description{
Calculate Log10 Fold Change for two-group comparison
}
\examples{
#pseq <- download_microbiome("peerj32")$physeq
  #fc <- check_foldchange(pseq, "gender")
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

