\name{chooseDir}
\alias{chooseDir}
\title{Choosing (and creating) a directory}
\usage{
  chooseDir(...)
}
\arguments{
  \item{...}{parameters to pass}
}
\value{
  choice
}
\description{
  Choosing (and creating) a directory
}
\examples{
# TBA
}
\author{
  Douwe Molenaar. Maintainer: Leo Lahti
  \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

