% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmat2table.R
\name{cmat2table}
\alias{cmat2table}
\title{Convert cross correlation results to a table}
\usage{
cmat2table(res, verbose = FALSE)
}
\arguments{
\item{res}{Output from cross_correlate}

\item{verbose}{verbose}
}
\value{
Correlation table
}
\description{
Arrange correlation matrices from cross_correlate into a table format.
}
\examples{
  data(peerj32)
  d1 <- peerj32$microbes[1:20, 1:10]
  d2 <- peerj32$lipids[1:20,1:10]
  cc <- cross_correlate(d1, d2, mode = 'matrix')
  cmat <- cmat2table(cc)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

