\name{completeClust}
\alias{completeClust}
\title{Description: heatmap function for complete linkage clustering}
\usage{
  completeClust(x)
}
\arguments{
  \item{x}{data matrix

  Returns:}
}
\value{
  hclust object
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

