\name{corDist}
\alias{corDist}
\title{Description: Correlation distance. Computes the correlation
distance between the _columns_ of the given matrix}
\usage{
  corDist(x)
}
\arguments{
  \item{x}{data matrix

  Returns:}
}
\value{
  distance object
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

