% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{core}
\alias{core}
\title{Core Taxa}
\usage{
core(x, detection.threshold = 1, prevalence.threshold = 95)
}
\arguments{
\item{x}{phyloseq object}

\item{detection.threshold}{Detection threshold (non-negative real)}

\item{prevalence.threshold}{Prevalence threshold (in [0, 100])}
}
\value{
Vector of core members
}
\description{
Determine members of the core microbiota with given abundance and prevalence thresholds.
}
\examples{
  data(dietswap)
  a <- core(dietswap, 1, 95)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

