% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/core.R
\name{core.sum}
\alias{core.sum}
\title{core.sum}
\usage{
core.sum(data, intTr, prevalenceTr)
}
\arguments{
\item{data}{data matrix; phylotypes vs. samples}

\item{intTr}{intTr}

\item{prevalenceTr}{prevalenceTr}
}
\value{
Number of OTUs
}
\description{
core.sum
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
A Salonen et al. The adult intestinal core microbiota is determined by
  analysis depth and health status. Clinical Microbiology and Infection
  18(S4):16 20, 2012.
  To cite the microbiome R package, see citation('microbiome')
}
\keyword{utilities}

