\name{Core2D}
\alias{Core2D}
\title{Core2D}
\usage{
  Core2D(coreMat, fname, writedir, colnum = NULL)
}
\arguments{
  \item{coreMat}{core matrix}

  \item{fname}{output file id}

  \item{writedir}{output directory}

  \item{colnum}{colnum

  Returns:}
}
\value{
  TBA
}
\description{
  Description: Core visualization 2D
}
\details{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

