\name{Core3D}
\alias{Core3D}
\title{Core3D}
\usage{
  Core3D(coreMat, title = "Abundance core")
}
\arguments{
  \item{coreMat}{core matrix}

  \item{title}{title

  Returns:}
}
\value{
  Used for its side effects
}
\description{
  Description: Core visualization 3D
}
\details{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

