% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_abundance.R
\name{core_abundance}
\alias{core_abundance}
\title{Core Abundance}
\usage{
core_abundance(x, detection = 1, prevalence = 95)
}
\arguments{
\item{x}{phyloseq object}

\item{detection}{Detection threshold (non-negative real)}

\item{prevalence}{Prevalence threshold (in [0, 100])}
}
\value{
Total core abunance vector.
}
\description{
Calculate total core abundance
}
\examples{
 data(dietswap)
 a <- core_abundance(dietswap)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

