% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{core_bootstrap}
\alias{core_bootstrap}
\title{Bootstrap Analysis of the Core Microbiota}
\usage{
core_bootstrap(x, Nsample = NULL, prevalence = 0.5, bs.iter = 1000,
  detection = 0, I.max = NULL)
}
\arguments{
\item{x}{OTUxSample data matrix}

\item{Nsample}{bootstrap sample size, default is the same size as data}

\item{prevalence}{Lower limit for number of samples where microbe needs 
to exceed the intensity threshold for a 'present' call.}

\item{bs.iter}{bootstrap iterations}

\item{detection}{Lower limit for intensity threshold}

\item{I.max}{Upper limit for intensity threshold. Later addition.
set to NULL (default) to replicate Salonen et al.}
}
\value{
data frame with microbes and their frequency of presence in 
  	     the core
}
\description{
Bootstrap analysis of the core microbiota.
}
\examples{
  data(peerj32)
  # In practice, use bs.iter = 1000 or more
  # Not exported:
  # bs <- core_bootstrap(peerj32$phyloseq, bs.iter = 5)
}
\references{
The core microbiota bootstrap method implemented with this function:
Salonen A, Salojarvi J, Lahti L, de Vos WM. The adult intestinal
core microbiota is determined by analysis depth and health
status. Clinical Microbiology and Infection 18(S4):16-20, 2012
To cite this R package, see citation("microbiome")
}
\author{
Contact: Jarkko Salojarvi \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
