% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_matrix.R
\name{core_matrix}
\alias{core_matrix}
\title{core_matrix}
\usage{
core_matrix(x, prevalence.intervals = seq(5, 100, 5),
  detection.thresholds = NULL, verbose = FALSE)
}
\arguments{
\item{x}{\code{\link{phyloseq}} object}

\item{prevalence.intervals}{a vector of prevalence percentages in [0,100]}

\item{detection.thresholds}{a vector of intensities around the data range}

\item{verbose}{verbose}
}
\value{
Estimated core microbiota
}
\description{
create core matrix
}
\examples{
  #pseq <- download_microbiome("peerj32")$physeq
  #core <- core_matrix(pseq)

}
\author{
Contact: Jarkko Salojarvi \email{microbiome-admin@googlegroups.com}
}
\references{
A Salonen et al. The adult intestinal core microbiota is determined by 
  analysis depth and health status. Clinical Microbiology and Infection 
  18(S4):16 20, 2012. 

  To cite the microbiome R package, see citation('microbiome')
}
\keyword{utilities}

