% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_richness.R
\name{core_richness}
\alias{core_richness}
\title{Core Richness}
\usage{
core_richness(x, detection = 1, prevalence = 95)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{detection}{Detection threshold (non-negative real)}

\item{prevalence}{Prevalence threshold (in [0, 100])}
}
\value{
Total core abundance vector.
}
\description{
Calculate total core richness
}
\examples{
 data(dietswap)
 a <- core_richness(dietswap)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

