% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_heatmap.R
\name{correlation_heatmap}
\alias{correlation_heatmap}
\title{Correlation Heatmap}
\usage{
correlation_heatmap(df, Xvar, Yvar, fill, star, p.adj.threshold = 1,
  correlation.threshold = 0, step = 0.2, colours = c("darkblue", "blue",
  "white", "red", "darkred"), limits = NULL, legend.text = "",
  order.rows = TRUE, order.cols = TRUE, text.size = 10,
  filter.significant = TRUE, star.size = NULL, plot.values = FALSE)
}
\arguments{
\item{df}{Data frame. Each row corresponds to a pair of correlated 
variables. The columns give variable names, correlations and 
significance estimates.}

\item{Xvar}{X axis variable column name. For instance 'X'.}

\item{Yvar}{Y axis variable column name. For instance 'Y'.}

\item{fill}{Column to be used for heatmap coloring. 
For instance 'correlation'.}

\item{star}{Column to be used for cell highlighting. For instance 'p.adj'.}

\item{p.adj.threshold}{Significance threshold for the stars.}

\item{correlation.threshold}{Include only elements that have absolute 
correlation higher than this value}

\item{step}{color interval}

\item{colours}{heatmap colours}

\item{limits}{colour scale limits}

\item{legend.text}{legend text}

\item{order.rows}{Order rows to enhance visualization interpretability}

\item{order.cols}{Order columns to enhance visualization interpretability}

\item{text.size}{Adjust text size}

\item{filter.significant}{Keep only the elements with at least one 
significant entry}

\item{star.size}{NULL Determine size of the highlight symbols}

\item{plot.values}{Show values as text}
}
\value{
ggplot2 object
}
\description{
Visualizes n x m correlation table as heatmap.
}
\examples{
  data(peerj32)
  d1 <- peerj32$lipids[, 1:10]
  d2 <- peerj32$microbes[, 1:10]
  cc <- cross_correlate(d1, d2) 
  p <- correlation_heatmap(cc, 'X1', 'X2', 'Correlation', star = "p.adj")
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

