% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage}
\alias{coverage}
\title{Coverage Index}
\usage{
coverage(x, threshold = 0.5, split = TRUE)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{threshold}{Indicates the fraction  of the ecosystem to be occupied by the N most abundant species (N is returned by this function).}

\item{split}{(Optional). Logical. Should a separate set of richness
estimates be performed for each sample? Or alternatively,
pool all samples and estimate richness of the entire set.}
}
\value{
A vector of coverage indices
}
\description{
Community coverage index.
}
\details{
The coverage index gives the number of groups needed to
  have a given proportion of the ecosystem occupied (by default 0.5 ie 50%).
}
\examples{
  data(dietswap)
  d <- coverage(dietswap, threshold = 0.5)
}
\seealso{
dominance, global
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
