\name{createCore}
\alias{createCore}
\title{createCore}
\usage{
  createCore(data, I.thr = 1.8)
}
\arguments{
  \item{data}{data matrix}

  \item{I.thr}{core threshold

  Returns:}
}
\value{
  TBA
}
\description{
  Description: create coreMatrix
}
\details{
  Arguments:
}
\examples{
# core <- createCore(L2.matrix)
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

