% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosshyb.R
\name{CrosshybTable}
\alias{CrosshybTable}
\title{Cross hybridization table}
\usage{
CrosshybTable(tax.level = "L1", selected.taxa = NULL, tax.table)
}
\arguments{
\item{tax.level}{Taxonomic level to investigate}

\item{selected.taxa}{Restrict cross-hyb analysis to the selected groups.}

\item{tax.table}{tax.table}
}
\value{
A list containing cross-hybridization table
}
\description{
Cross-hybridization table between multimodal taxa as percentages of shared probes. The number indicates how many percent of oligos for the row taxon are also hybridizing the corresponding column taxon.
}
\examples{
\dontrun{ch <- CrosshybTable(tax.level = 'L1')}
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

