\name{deconvolution.nonneg}
\alias{deconvolution.nonneg}
\title{Description: mixingMatrix}
\usage{
  deconvolution.nonneg(oligo.data, oligo.map, level,
    block.solution = T)
}
\arguments{
  \item{oligo.data}{oligo.data}

  \item{oligo.map}{oligo.map}

  \item{level}{taxonomic level}

  \item{block.solution}{block.solution

  Returns:}
}
\value{
  list
}
\description{
  For cross-hyb control
}
\details{
  Arguments:
}
\author{
  Contact: Jarkko Salojarvi
  \email{jarkko.salojarvi@helsinki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

