% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_colors.R
\name{default_colors}
\alias{default_colors}
\title{Default Colors}
\usage{
default_colors(x, v = NULL)
}
\arguments{
\item{x}{Name of the variable type ("Phylum")}

\item{v}{Optional. Vector of elements to color.}
}
\value{
Named character vector of default colors
}
\description{
Default colors for different variables.
}
\examples{
col <- default_colors("Phylum")
}
\references{
See citation("microbiome")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
