\name{density.plot}
\alias{density.plot}
\title{density.plot}
\usage{
  density.plot(mat, title = NULL)
}
\arguments{
  \item{mat}{Data matrix to plot. The first two columns
  will be visualized as a cross-plot.}

  \item{title}{title text

  Returns:}
}
\value{
  ggplot2 object
}
\description{
  Description: Plots densities of data points in addition
  to cross-plot points.
}
\details{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

