% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dietswap}
\alias{dietswap}
\title{Diet Swap Data}
\format{
The data set in \code{\link{phyloseq-class}} format.
}
\usage{
data(dietswap)
}
\value{
Loads the data set in R.
}
\description{
The diet swap data set represents a study with African
and African American groups undergoing a two-week diet swap.
For details, see
\url{dx.doi.org/10.1038/ncomms7342}.
}
\details{
The data is also available for download from the Data Dryad
repository \url{http://datadryad.org/resource/doi:10.5061/dryad.1mn1n}.
}
\references{
O'Keefe et al. Nature Communications 6:6342, 2015.
\url{dx.doi.org/10.1038/ncomms7342}
To cite the microbiome R package, see citation('microbiome')
}
\author{
Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{data}
