\name{dir.exists}
\alias{dir.exists}
\title{dir.exists}
\usage{
dir.exists(d)
}
\arguments{
  \item{d}{d}
}
\value{
logical
}
\description{
dir.exists
}
\examples{
#
}
\author{
Contact: Leo Lahti
\email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

