\name{diversity.boxplot}
\alias{diversity.boxplot}
\title{diversity.boxplot}
\usage{
  diversity.boxplot(div.table,
    diversity.index = "invsimpson", title = NULL, col.list,
    sample.groups)
}
\arguments{
  \item{div.table}{diversity table: output from the
  estimate.diversity function}

  \item{diversity.index}{specify the diversity index to
  plot (diversity / richness / evenness)}

  \item{title}{figure title}

  \item{n.groups}{max number of groups for boxplot

  Returns:}
}
\value{
  Sample group list corresponding to the boxplot groups.
}
\description{
  Description: diversity.boxplot
}
\details{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

