% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_table.R
\name{diversity_table}
\alias{diversity_table}
\title{Estimate Diversity}
\usage{
diversity_table(x, split = TRUE, measures = NULL, det.th = 0)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{split}{See help(phyloseq::estimate_richness)}

\item{measures}{See help(phyloseq::estimate_richness). In addition,
the measure "Evenness" is provided (Pielou's index).}

\item{det.th}{detection for observing taxa (absence / presence)}
}
\value{
Vector containing relative proportions for each phylotype in 
        each sample
}
\description{
Diversity estimation.
   Augments the estimate_richness function of the phyloseq package.
}
\examples{
  data(dietswap)
  d <- diversity_table(dietswap)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

