% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{download_atlas}
\alias{download_atlas}
\title{Download HITChip Atlas}
\usage{
download_atlas(format = "phyloseq")
}
\arguments{
\item{format}{"phyloseq" or "original"}
}
\value{
Data set
}
\description{
Download HITChip Atlas
}
\examples{
\dontrun{download_atlas()}
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
Lahti et al. Tipping elements of the human intestinal ecosystem. 
  Nature Communications 5:4344, 2014.
  To cite the microbiome R package, see citation('microbiome')
}
\keyword{internal}

