% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{download_dietswap}
\alias{download_dietswap}
\title{Download Diet Swap Study}
\usage{
download_dietswap(...)
}
\arguments{
\item{...}{Arguments to be passed}
}
\value{
Data set
}
\description{
Download Diet Swap Study
}
\examples{
\dontrun{download_dietswap()}
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
O'Keefe et al. Nature Communications 6:6342, 2015.
  To cite the microbiome R package, see citation('microbiome')
}
\keyword{internal}

