% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{download_microbiome}
\alias{download_microbiome}
\title{Download microbiome data sets}
\usage{
download_microbiome(id = "datasets", format = NULL)
}
\arguments{
\item{id}{Data set name. For options, see download_microbiome()}

\item{format}{Data format ("phyloseq" or "original")}
}
\value{
Data set
}
\description{
Download microbiome data sets
}
\examples{
# x <- download_microbiome("peerj32")
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
To cite the microbiome R package, see citation('microbiome')
}
\keyword{utilities}

