% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equality.R
\name{equality}
\alias{equality}
\title{Gini Index}
\usage{
equality(x, detection, prevalence, split = TRUE)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{detection}{Detection threshold for absence/presence (greater or equal).}

\item{prevalence}{Prevalence threshold (in [0, 1]; greater or equal)}

\item{split}{(Optional). Logical. Calculate separate indices
for each sample, or pool all samples and estimate the index
for the entire set.}
}
\value{
A vector of Gini indices
}
\description{
Calculate Gini indices for a phyloseq object.
}
\details{
Gini index is a common measure for inequality in economical income, but can also be used as a community diversity measure.
}
\examples{
  data(dietswap)
  d <- equality(dietswap)
}
\references{
Relative Distribution Methods in the Social Sciences. Mark S.
  Handcock and Martina Morris, Springer-Verlag, Inc., New York,
  1999. ISBN 0387987789.
}
\seealso{
diversity, reldist::gini (inspired by that implementation but independently written here to avoid external depedencies)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
