\name{estimate.min.threshold}
\alias{estimate.min.threshold}
\title{Description: determine threshold for bg correction}
\usage{
  estimate.min.threshold(dat)
}
\arguments{
  \item{dat}{data matrix (in approximately normal scale ie.
  logged)

  Returns:}
}
\value{
  threshold value
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

