% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_homogeneity.R
\name{estimate_homogeneity}
\alias{estimate_homogeneity}
\title{Homogeneity analysis}
\usage{
estimate_homogeneity(x, type = "interindividual", group_by = "group",
  method = "spearman")
}
\arguments{
\item{x}{phyloseq object with the OTU matrix and sample metadata. The sample metadata should contain the fields "sample" and "group" (or another grouping variable specified in the group_by argument) for the "interindividual" method. For the "intraindividual" method, also provide "time", and "subject" fields.}

\item{type}{Homogeneity type: 'interindividual' or 'intraindividual'}

\item{group_by}{variable to be used in grouping. By default: "group"}

\item{method}{correlation method (see ?cor)}
}
\value{
List with correlations, group-wise statistics, and ANOVA linear model p-value for group differences.
}
\description{
Estimate homogeneity within or between sample groups.
}
\details{
Average correlation between samples in the input data within each
group with the overall group-wise average. Picks the lower triangular
matrix to avoid duplicating the correlations. Returns correlations and
stability estimate (inter-individual; average of the
correlations). Can also be used to calculate temporal homogeneity
between two data sets (intra-individual), given appropriate sample
metadata.
}
\examples{
\dontrun{
# Example data
library(microbiome)
x <- download_microbiome("dietswap")
sample_data(x)$time <- sample_data(x)$timepoint.within.group
# Estimate inter-individual homogeneity
res <- estimate_homogeneity(x, "interindividual")
# Estimate intra-individual homogeneity
res <- estimate_homogeneity(x, "intraindividual")
}
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
The inter- and intra-individual homogeneity are calculated
as described in Salonen et al. ISME J. 8:2218-30, 2014.
To cite this R package, see citation('microbiome')
}
\keyword{utilities}

