% Generated by roxygen2 (4.0.2): do not edit by hand
\name{evenness}
\alias{evenness}
\title{evenness}
\usage{
evenness(dat, det.th = NULL)
}
\arguments{
\item{dat}{data matrix (phylotypes x samples) in original (non-log) scale}

\item{det.th}{detection threshold. Used for richness and evenness estimation. Not used in diversity estimation.

Returns:}
}
\value{
Vector containing evenness estimate for each sample
}
\description{
Description: Estimate evenness for each sample with a given threshold
}
\details{
Arguments:
}
\examples{
data(peerj32);
	     eve <- evenness(10^t(peerj32$microbes),
	     	    	     det.th = 100)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
This function builds on tools from the vegan R package.
   Jari Oksanen et al. (2013). vegan: Community Ecology.
   R Package. http://CRAN.R-project.org/package=vegan

   To cite the microbiome R package, see citation("microbiome")
}
\keyword{utilities}

