\name{expandCondition}
\alias{expandCondition}
\title{Expands a list of "field", "value" pair lists into an SQL condition}
\usage{
  expandCondition(condition)
}
\arguments{
  \item{condition}{condition

  Returns:}
}
\value{
  TBA
}
\description{
  Arguments:
}
\examples{
# TBA
}
\author{
  Douwe Molenaar. Maintainer: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

