\name{FetchData}
\alias{FetchData}
\title{Description: Fetch data from the database}
\usage{
  FetchData(params, con, scriptVersion, save.data, scaling,
    cmetrics)
}
\arguments{
  \item{params}{params}

  \item{con}{con}

  \item{scriptVersion}{scriptVersion}

  \item{save.data}{save.data}

  \item{scaling}{scaling}

  \item{cmetrics}{cmetrics

  Returns:}
}
\value{
  data
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

