% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_prevalent.R
\name{filter_prevalent}
\alias{filter_prevalent}
\title{Filter Prevalent Taxa}
\usage{
filter_prevalent(x, detection.threshold, prevalence.threshold)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{detection.threshold}{Detection threshold for absence/presence.}

\item{prevalence.threshold}{Prevalence threshold}
}
\value{
Filtered phyloseq object including only prevalent taxa
}
\description{
Filter the phyloseq object to include only prevalent taxa.
}
\examples{
  data("peerj32")
  filter_prevalent(peerj32$phyloseq, 200, 0.2)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
To cite the microbiome R package, see citation('microbiome')
}
\keyword{utilities}

