% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_optima.R
\name{find_optima}
\alias{find_optima}
\title{Find Optima}
\usage{
find_optima(f, detection = 0, bw = 1, detection.limit = 1)
}
\arguments{
\item{f}{density}

\item{detection}{detection for peaks}

\item{bw}{bandwidth}

\item{detection.limit}{Minimun accepted density for a maximum; 
as a multiple of kernel height}
}
\value{
A list with min (minima), max (maxima), and
        detection.density (minimum detection density)
}
\description{
Detect optima, excluding local optima below
             detection.
}
\examples{
find_optima(rnorm(100), bw = 1)
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

