\name{GenerateSimulatedData}
\alias{GenerateSimulatedData}
\title{Description: Generate toydata for the package}
\usage{
  GenerateSimulatedData(oligo.matrix.nolog.simulated,
    phylogeny.info, output.file = "toydata-hitchip.rda")
}
\arguments{
  \item{oligo.matrix.nolog.simulated}{}

  \item{phylogeny.info}{phylogeny.info}

  \item{output.file}{output file name

  Returns:}
}
\value{
  output file name
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

