\name{get.dd.order}
\alias{get.dd.order}
\title{Description: This feature for dendrograms took like an hour to find out
              because of unadequate documentation...}
\usage{
  get.dd.order(dd)
}
\arguments{
  \item{dd}{TBA

  Returns:}
}
\value{
  dd
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

