\name{get.doligo2}
\alias{get.doligo2}
\title{Description: Calculate d.oligo2}
\usage{
  get.doligo2(featuretab, d.scaled, oligo.ids)
}
\arguments{
  \item{featuretab}{featuretab}

  \item{d.scaled}{d.scaled}

  \item{oligo.ids}{oligo.ids Returns:}
}
\value{
  d.oligo2
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

