% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.probeset}
\alias{get.probeset}
\title{Description: get probeset data matrix}
\usage{
get.probeset(name, level, phylogeny.info, oligo.matrix, log10 = TRUE)
}
\arguments{
\item{name}{name}

\item{level}{taxonomic level}

\item{phylogeny.info}{phylogeny.info}

\item{oligo.matrix}{oligos vs. samples preprocessed data matrix;
                 absolute scale}

\item{log10}{Logical. Logarithmize the data TRUE/FALSE

Returns:}
}
\value{
probeset data matrix
}
\description{
Arguments:
}
\examples{
phylogeny.info <- GetPhylogeny('HITChip', 'filtered')
     # ps <- get.probeset('Vibrio', 'L2', phylogeny.info, oligo.matrix)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

