\name{get.sampleid}
\alias{get.sampleid}
\title{Description: Pick sampleIDs from d.oligo column names}
\usage{
  get.sampleid(d.oligo)
}
\arguments{
  \item{d.oligo}{d.oligo matrix Returns:}
}
\value{
  sampleID vector
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

