% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hitchip_taxonomy.R
\name{get_hitchip_taxonomy}
\alias{get_hitchip_taxonomy}
\title{HITChip Taxonomy Table}
\usage{
get_hitchip_taxonomy(chip, phylogeny.version = "full", data.dir = NULL)
}
\arguments{
\item{chip}{chip type (e.g. 'HITChip')}

\item{phylogeny.version}{'full' or 'filtered' 
(latter is the basis for species/L1/L2 summarization)}

\item{data.dir}{Data directory path}
}
\value{
phylogeny mapping table
}
\description{
Get taxonomy table for phylogenetic microarrays (mainly HITChip).
}
\examples{
# tax.table <- get_hitchip_taxonomy('HITChip', 'full')
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

