% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ordination.R
\name{get_ordination}
\alias{get_ordination}
\title{Get Ordination}
\usage{
get_ordination(x, method = "NMDS", distance = "bray")
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object or a data matrix 
(features x samples; eg. HITChip taxa vs. samples)}

\item{method}{Ordination method, see phyloseq::plot_ordination}

\item{distance}{Ordination distance, see phyloseq::plot_ordination}
}
\value{
data.frame with ordination coordinates and metadata
}
\description{
Ordinate phyloseq data and merge it with sample metadata
}
\details{
This is a wrapper for phyloseq ordination functions, providing
smooth access to ordinated data.frame with full info on the projection
and metadata necessary for further visualizations.
}
\examples{
data(dietswap)
fc <- get_ordination(dietswap)
}
\references{
See citation('microbiome')
}
\seealso{
phyloseq::plot_ordination
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
