\name{getPvalsFromLmMatrix}
\alias{getPvalsFromLmMatrix}
\title{Description: Get Pvals From Lm Matrix}
\usage{
  getPvalsFromLmMatrix(lmMatrix)
}
\arguments{
  \item{lmMatrix}{TBA

  Returns:}
}
\value{
  anova.pvals
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

