% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{gini}
\alias{gini}
\title{Gini Index}
\usage{
gini(x, detection, prevalence, split = TRUE)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{split}{(Optional). Logical. Should a separate set of richness
estimates be performed for each sample? Or alternatively,
pool all samples and estimate richness of the entire set.}
}
\value{
A vector of Gini indices
}
\description{
Calculates the Gini index for phyloseq object
}
\details{
Gini index is a common measure for inequality in economical income, but can also be used as a community diversity measure.
}
\examples{
  data(dietswap)
  d <- gini(dietswap)
}
\references{
Relative Distribution Methods in the Social Sciences. Mark S.
    Handcock and Martina Morris, Springer-Verlag, Inc., New York,
    1999. ISBN 0387987789.
}
\seealso{
diversity, reldist::gini (inspired by that implementation but independently written here to avoid external depedencies)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
