\name{hitchip.phylodistance}
\alias{hitchip.phylodistance}
\title{Description: phylogenetic distance matrix for HITChip}
\usage{
  hitchip.phylodistance(level, f.phylomap, f.phylodist)
}
\arguments{
  \item{level}{phylogeny level}

  \item{f.phylomap}{phylogenetic map with a column for each
  level, and one column for the ID which matches with the
  sequence similarity matrix in f. phylodist file}

  \item{f.phylodist}{probe sequence similarity matrix

  Returns:}
}
\value{
  list
}
\description{
  Arguments:
}
\author{
  Contact: Jarkko Salojarvi
  \email{jarkko.salojarvi@helsinki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

