\name{hitchip.sig.lme}
\alias{hitchip.sig.lme}
\title{Description: TBA}
\usage{
  hitchip.sig.lme(d.hitchip, annot, my.class, time = NULL)
}
\arguments{
  \item{d.hitchip}{hitchip data matrix: phylogroups x
  samples}

  \item{annot}{annotations for samples, containing
  my.class, subjectID, sampleID, and perhaps time}

  \item{my.class}{specify the inspected class;}

  \item{time}{if time is additionally used in modeling then
  annot needs to contain field named time

  Returns:}
}
\value{
  TBA
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

