% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hitchip.taxonomy}
\alias{hitchip.taxonomy}
\title{HITChip Taxonomy}
\format{List with the element 'filtered', including a simplified version
        of the HITChip taxonomy.}
\usage{
data(hitchip.taxonomy)
}
\description{
HITChip taxonomy table.
}
\author{
Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
Lahti et al. Tipping elements of the human intestinal ecosystem. 
  Nature Communications 5:4344, 2014.
  To cite the microbiome R package, see citation('microbiome')
}
\keyword{data}

