\name{HTMLReport}
\alias{HTMLReport}
\title{generic report part}
\usage{
  HTMLReport(x,
    file = paste(file.path(x$directory, x$filename), "html", sep = "."),
    CSSfile = "R2HTML.css")
}
\arguments{
  \item{x}{TBA}

  \item{file}{TBA}

  \item{CSSfile}{TBA}
}
\value{
  TBA
}
\description{
  generic report part
}
\examples{
# TBA
}
\author{
  Douwe Molenaar. Maintainer: Leo Lahti
  \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

