\name{HTMLReportBegin}
\alias{HTMLReportBegin}
\title{header part}
\usage{
  HTMLReportBegin(file = "report.html",
    title = "Report Title")
}
\arguments{
  \item{file}{filename}

  \item{title}{title}
}
\value{
  TBA
}
\description{
  header part
}
\examples{
# TBA
}
\author{
  Douwe Molenaar. Maintainer: Leo Lahti
  \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

