\name{HTMLReportEnd}
\alias{HTMLReportEnd}
\title{End html report part}
\usage{
  HTMLReportEnd(file = "report.html")
}
\arguments{
  \item{file}{filename}
}
\value{
  TBA
}
\description{
  End html report part
}
\examples{
# TBA
}
\author{
  Douwe Molenaar. Maintainer: Leo Lahti
  \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

