% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_hitchip.R
\name{import_hitchip}
\alias{import_hitchip}
\title{Import HITChip data}
\usage{
import_hitchip(data.dir, method = "frpa", detection.threshold = 0,
  verbose = F)
}
\arguments{
\item{data.dir}{Profiling script output directory for reading the data.}

\item{method}{Probe summarization method ("rpa" or "sum")}

\item{detection.threshold}{Taxon absence/presence thresholds (typically 10^1.8 for HITChip)}

\item{verbose}{verbose}
}
\value{
data matrix (phylo x samples)
}
\description{
Import HITChip output into phyloseq format.
}
\examples{
 \dontrun{
  data.dir <- system.file("extdata", package = "microbiome")
  dat <- import_hitchip(data.dir)
}
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

