\name{include.fig}
\alias{include.fig}
\title{Description: brew embedded documentation helper function}
\usage{
  include.fig(fname, captiontext, width = "5cm",
    height = "5cm", rotate = NULL, include.figname = TRUE)
}
\arguments{
  \item{fname}{Ouput file name}

  \item{captiontext}{Caption text}

  \item{width}{Image width}

  \item{height}{Image height}

  \item{rotate}{Optional. Image rotation angle.}

  \item{include.figname}{Include figure name Returns:}
}
\value{
  Saves image in the output file
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

