\name{include.figs}
\alias{include.figs}
\title{Description: brew embedded documentation helper function to include
multiple figures}
\usage{
  include.figs(fnames, captiontext, width = "5cm",
    height = "5cm", rotate = NULL)
}
\arguments{
  \item{fnames}{Ouput file name}

  \item{captiontext}{Caption text}

  \item{width}{Image width}

  \item{height}{Image height}

  \item{rotate}{Optional. Image rotation angle.

  Returns:}
}
\value{
  Saves image in the output file
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

