% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.compositional.R
\name{is.compositional}
\alias{is.compositional}
\title{Check Compositionality}
\usage{
is.compositional(x)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}
}
\value{
Logical TRUE/FALSE
}
\description{
Check if a phyloseq object is compositional.
}
\examples{
data(dietswap)
d <- is.compositional(dietswap)
d <- is.compositional(transform(dietswap, "compositional"))
}
\references{
See citation('microbiome')
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
