% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.phyloseq.R
\name{is.phyloseq}
\alias{is.phyloseq}
\title{Identify Phyloseq Objects}
\usage{
is.phyloseq(x)
}
\arguments{
\item{x}{object to test}
}
\value{
Logical
}
\description{
Identifies whether a given object is from the phyloseq class
}
\examples{
  library(microbiome)
  data(dietswap)
  is.phyloseq(dietswap)
}

