\name{L2.species}
\alias{L2.species}
\title{Description: L2-species mappings}
\usage{
  L2.species(l2, oligomap)
}
\arguments{
  \item{l2}{l2}

  \item{oligomap}{oligomap

  Returns:}
}
\value{
  L2-species mappings
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

