\name{level.species}
\alias{level.species}
\title{Description: L2/L1-species mappings. For L1/L2 groups, list corresponding species.}
\usage{
  level.species(level, oligomap, groups = NULL)
}
\arguments{
  \item{level}{level}

  \item{oligomap}{oligomap}

  \item{groups}{return selected groups only

  Returns:}
}
\value{
  level-species mappings
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

