% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internal.R
\name{level1TOlevel0}
\alias{level1TOlevel0}
\title{L1-L0 mappings}
\usage{
level1TOlevel0(l1, phylogeny.info)
}
\arguments{
\item{l1}{level1 phylotypes}

\item{phylogeny.info}{phylogeny.info}
}
\value{
mappings
}
\description{
L1-L0 mappings
}
\author{
Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation('microbiome')
}
\keyword{internal}

