% Generated by roxygen2 (4.0.2): do not edit by hand
\name{level2TOlevel1}
\alias{level2TOlevel1}
\title{level2-level1 mappings}
\usage{
level2TOlevel1(l2, phylogeny.info)
}
\arguments{
\item{l2}{level2 phylotypes}

\item{phylogeny.info}{phylogeny.info}
}
\value{
mappings
}
\description{
level2-level1 mappings
}
\author{
Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation('microbiome')
}
\keyword{internal}

