\name{levelmap}
\alias{levelmap}
\title{Description: map phylotypes to upper hierarchy level}
\usage{
  levelmap(phylotypes, level.from, level.to, oligomap)
}
\arguments{
  \item{phylotypes}{phylotypes}

  \item{level.from}{level.from}

  \item{level.to}{level.to}

  \item{oligomap}{oligomap

  Returns:}
}
\value{
  mappings
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

