\name{list.clustering.metrics}
\alias{list.clustering.metrics}
\title{Description: List clustering metrics}
\usage{
  list.clustering.metrics()
}
\value{
  list of clustering metrics
}
\description{
  Arguments:
}
\details{
  Returns:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

