\name{list.entities}
\alias{list.entities}
\title{Description: list corresponding entities}
\usage{
  list.entities(level, phylo)
}
\arguments{
  \item{level}{level}

  \item{phylo}{phylo

  Returns:}
}
\value{
  entities
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

