\name{list.mysql.projects}
\alias{list.mysql.projects}
\title{Description: List projects in MySQL database}
\usage{
  list.mysql.projects(dbuser, dbpwd, dbname)
}
\arguments{
  \item{dbuser}{MySQL user}

  \item{dbpwd}{MySQL password}

  \item{dbname}{MySqL database name Returns:}
}
\value{
  project names vector
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

