\name{list.scaling.methods}
\alias{list.scaling.methods}
\title{Description: List scaling methods}
\usage{
  list.scaling.methods()
}
\value{
  List of scaling methods
}
\description{
  Arguments:
}
\details{
  Returns:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

