\name{list2ascii}
\alias{list2ascii}
\title{Description: R or S function to write an R list to an ASCII file.
  This can be used to create files for those who want to use
  a spreadsheet or other program on the data.}
\usage{
  list2ascii(x,
    file = paste(deparse(substitute(x)), ".txt", sep = ""),
    append = F)
}
\arguments{
  \item{x}{TBA}

  \item{file}{TBA}

  \item{append}{TBA

  Returns:}
}
\value{
  TBA
}
\description{
  Arguments:
}
\note{
  Modified from version MHP July 7, 2004.
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

