% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_phyloseq.R
\name{lm_phyloseq}
\alias{lm_phyloseq}
\title{Limma test for phyloseq}
\usage{
lm_phyloseq(x, varname, transformation = "log10", p.adj.method = "BH")
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{varname}{Metadata field specifying the investigated metadata variable.}

\item{transformation}{Transformation for the original phyloseq values. By default a log10 transformation is done to better approximate the requirements of a linear model.}

\item{p.adj.method}{p-value correction method for p.adjust function 
(default 'BH'). For other options, see ?p.adjust}
}
\value{
Limma output table.
}
\description{
Limma for a single continuous variable vs. OTUs with a phyloseq object.
}
\examples{
 data("atlas1006")
 tab <- lm_phyloseq(atlas1006, "age")
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

