% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_abundance.R
\name{low_abundance}
\alias{low_abundance}
\title{Low_Abundance Index}
\usage{
low_abundance(x, detection, split = TRUE)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{detection}{Detection threshold for absence/presence (greater or equal).}

\item{split}{(Optional). Logical. Pool all samples and estimate index for the entire set.}
}
\value{
A vector of low_abundance indices
}
\description{
Calculates the community low_abundance index.
}
\details{
The low_abundance index gives the relative proportion of rare species in [0,1]. The species that are below the indicated detection threshold are considered rare. Note that population prevalence is not considered.
}
\examples{
  data(dietswap)
  d <- low_abundance(dietswap, detection = 0.1/100)
}
\seealso{
core_abundance, rarity, diversity
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
