% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{matrix.padjust}
\alias{matrix.padjust}
\title{matrix.padjust}
\usage{
matrix.padjust(pvals, p.adjust.method = "BH")
}
\arguments{
\item{pvals}{p-value matrix}

\item{p.adjust.method}{p-value adjustment method: for options, see ?p.adjust}
}
\value{
Adjusted p-value matrix
}
\description{
Calculate adjusted p-values for a matrix of pvalues
which may contain missing values.
}
\examples{
qvals <- matrix.padjust(matrix(runif(1000), nrow = 100))
}
\author{
Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
JD Storey 2003. Ann. Statist. 31(6):2013-2035. The positive false
  discovery rate: a Bayesian interpretation and the q-value.

  To cite the microbiome R package, see citation('microbiome')
}
\keyword{utilities}

