\name{mixingMatrix}
\alias{mixingMatrix}
\title{Description: mixingMatrix}
\usage{
  mixingMatrix(oligomap, level)
}
\arguments{
  \item{oligomap}{oligomap}

  \item{level}{taxonomic level

  Returns:}
}
\value{
  oligos x phylotypes mixing matrix
}
\description{
  For cross-hyb control
}
\details{
  Arguments:
}
\author{
  Contact: Jarkko Salojarvi
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

