% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimodality_score.R
\name{multimodality_score}
\alias{multimodality_score}
\title{Multimodality Score}
\usage{
multimodality_score(x, detection = 1, bw.adjust = 1, bs.iterations = 100,
  detection.limit = 1, verbose = TRUE)
}
\arguments{
\item{x}{A vector, or data matrix (variables x samples)}

\item{detection}{Mode detection}

\item{bw.adjust}{Bandwidth adjustment}

\item{bs.iterations}{Bootstrap iterations}

\item{detection.limit}{minimum accepted density for a maximum;
as a multiple of kernel height}

\item{verbose}{Verbose}
}
\value{
A list with following elements: 
  \itemize{
    \item{score}{Fraction of bootstrap samples with multiple observed modes}
    \item{nmodes}{The most frequently observed number of modes in bootstrap}
    \item{results}{Full results of potential_analysis_bootstrap for each
                     row of the input matrix.}
}
}
\description{
Multimodality score based on bootstrapped potential analysis.
}
\details{
Repeats potential analysis (Livina et al. 2010) multiple times
         with bootstrap sampling for each row of the input data
         (as in Lahti et al. 2014) and returns the specified results.
}
\examples{
  data(peerj32)
  s <- multimodality_score(
      	    t(peerj32$microbes[, c("Akkermansia", "Dialister")]))
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
\itemize{
  \item{}{Livina et al. (2010). Potential analysis reveals changing number
            of climate states during the last 60 kyr.
     \emph{Climate of the Past}, 6, 77-82.}
  \item{}{Lahti et al. (2014). Tipping elements of the human intestinal
    	     ecosystem. \emph{Nature Communications} 5:4344.}
 }
}
\keyword{utilities}

