\name{mysql.format}
\alias{mysql.format}
\title{Description: Format string vector to mysql query format}
\usage{
  mysql.format(s)
}
\arguments{
  \item{s}{string vector

  Returns:}
}
\value{
  mysql query version
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

