% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neatmap.R
\name{neat}
\alias{neat}
\title{Neatmap ordering for matrices}
\usage{
neat(x, arrange = "both", method = "NMDS", distance = "bray",
  first = NULL, ...)
}
\arguments{
\item{x}{A matrix.}

\item{arrange}{Order "rows" or "cols" or "both".}

\item{method}{Ordination method. Only NMDS implemented for now.}

\item{distance}{Distance method. See \code{\link{vegdist}} function from the \pkg{vegan} package.}

\item{first}{Optionally provide the name of the first sample/taxon to start the ordering (the ordering is cyclic so we can start at any point). The choice of the first sample may somewhat affect the overall ordering.}

\item{...}{Arguments to pass.}
}
\value{
Sorted matrix
}
\description{
Order matrix based on the neatmap approach.
}
\details{
This function borrows elements from the heatmap implementation in the \pkg{phyloseq} package. The row/column sorting is there
not available as a separate function at present, however, hindering reuse in other tools. Therefore I implemented this function to
provide an independent method for easy sample/taxon reordering for phyloseq objects.
}
\examples{
\dontrun{
   data(peerj32)
   x <- peerj32$microbes
   x <- neat(x, "both", method = "NMDS", distance = "bray") 
                  }
}
\references{
This function is partially based on code derived from the \pkg{phyloseq} package. However for the original
  neatmap approach for heatmap sorting, see (and cite):
  Rajaram, S., & Oono, Y. (2010). NeatMap--non-clustering heat map alternatives in R. BMC Bioinformatics, 11, 45.
}
\keyword{utilities}

