% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noncore_abundance.R
\name{noncore_abundance}
\alias{noncore_abundance}
\title{Non-Core Abundance Index}
\usage{
noncore_abundance(x, detection = 0.1/100, prevalence = 50/100)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{detection}{Detection threshold for absence/presence
(strictly greater by default).}

\item{prevalence}{Prevalence threshold (in [0, 1]). The
required prevalence is strictly greater by default. To include the
limit, set include.lowest to TRUE.}
}
\value{
A vector of indices
}
\description{
Calculates the noncore_abundance community index.
}
\details{
The noncore_abundance index gives the relative proportion of rare
species (ie. those that are not part of the core microbiota) in the
interval [0,1]. This is the complement (1-x) of the core abundance.
The rarity function provides the abundance of the least abundant taxa
within each sample, regardless of the population prevalence.
}
\examples{
data(dietswap)
d <- noncore_abundance(dietswap, detection=0.1/100, prevalence=50/100)
}
\seealso{
core_abundance, rarity, diversity
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
