% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_time.R
\name{normalize_time}
\alias{normalize_time}
\title{Normalize Phyloseq Metadata Time Field}
\usage{
normalize_time(x)
}
\arguments{
\item{x}{phyloseq object. The sample_data(x) should contain the following fields: subject, time}
}
\value{
Phyloseq object with a normalized time field
}
\description{
Shift the time field in phyloseq sample_data such that the first time point of each subject is always 0.
}
\examples{
  library(microbiome)
  data(atlas1006)
  atlas1006b <- normalize_time(atlas1006)
}

