\name{oligo.bg.correction}
\alias{oligo.bg.correction}
\title{Description: Calculate species summaries and possibly update d.oligo2}
\usage{
  oligo.bg.correction(d.oligo2, bgc.method)
}
\arguments{
  \item{d.oligo2}{d.oligo2}

  \item{bgc.method}{background correction method Returns:}
}
\value{
  Background-corrected data matrix
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

