\name{oligoTOspecies}
\alias{oligoTOspecies}
\title{Description: oligo-species mappings}
\usage{
oligoTOspecies(oligo, phylogeny.info)
}
\arguments{
  \item{oligo}{oligos}

  \item{phylogeny.info}{phylogeny.info

  Returns:}
}
\value{
mappings
}
\description{
Arguments:
}
\author{
Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("microbiome")
}
\keyword{internal}

