% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{overlap}
\alias{overlap}
\title{Overlap Measure}
\usage{
overlap(x, detection = 0)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{detection}{Detection threshold.}
}
\value{
Overlap matrix
}
\description{
Quantify microbiota 'overlap' between samples.
}
\examples{
data(atlas1006)
o <- overlap(atlas1006, detection = 0.1/100)
}
\references{
Bashan, A., Gibson, T., Friedman, J. et al.
Universality of human microbial dynamics.
Nature 534, 259–262 (2016). \url{https://doi.org/10.1038/nature18301}
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
