\name{panel.stability}
\alias{panel.stability}
\title{panel.stability}
\usage{
  panel.stability(x, y, scale = c("linear", "logarithmic"),
    ...)
}
\arguments{
  \item{x}{TBA}

  \item{y}{TBA}

  \item{scale}{TBA}

  \item{...}{parameters to pass}
}
\value{
  TBA
}
\description{
  panel.stability
}
\examples{
# TBA
}
\author{
  Douwe Molenaar. Maintainer: Leo Lahti
  \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

