\name{PCA.dialogue}
\alias{PCA.dialogue}
\title{Description: PCA dialogue window}
\usage{
  PCA.dialogue(d, d.i, write.dir)
}
\arguments{
  \item{d}{TBA}

  \item{d.i}{TBA}

  \item{write.dir}{Output directory

  Returns:}
}
\value{
  tkpack
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

