% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phylo.barplot.R
\name{phylo.barplot}
\alias{phylo.barplot}
\title{phylo.barplot}
\usage{
phylo.barplot(x, color.level = "L1", phylogeny.info = NULL, title = NULL,
  plot = TRUE, sort = TRUE)
}
\arguments{
\item{x}{Data vector across taxa (each element should be named by taxon)}

\item{color.level}{Higher-order phylogenetic level to indicate by colors}

\item{phylogeny.info}{oligo-phylotype mappings}

\item{title}{title}

\item{plot}{draw plot TRUE/FALSE}

\item{sort}{sort the effects by magnitude}
}
\value{
ggplot2 object
}
\description{
Barplot for *ITChip sample (across taxa) with higher-level
taxonomic groups indicated by colours.
}
\examples{
data(peerj32)
  phylogeny.info <- GetPhylogeny('HITChip', 'filtered')
  signal <- unlist(peerj32$microbes[1, 1:10])
  p <- phylo.barplot(signal, 'L1', phylogeny.info)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

