\docType{data}
\name{phylogeny}
\alias{phylogeny}
\title{phylogeny}
\description{
  The phylogeny contains the oligo-species-L2-L1 mappings
  for HITChip
}
\author{
  Leo Lahti \email{microbiome-devel@googlegroups.com}
}
\references{
  \url{microbiome.github.com}
}
\keyword{data}
\keyword{datasets}

