\name{phylotype.rm.list}
\alias{phylotype.rm.list}
\title{Description: Default list of removed phylotypes and oligos}
\usage{
  phylotype.rm.list(chip)
}
\arguments{
  \item{chip}{Chip name (HIT/MIT/PIT/Chick)Chip Returns:}
}
\value{
  List of removed oligos and phylotypes
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

