\name{PhylotypeRatios}
\alias{PhylotypeRatios}
\title{PhylotypeRatios}
\usage{
  PhylotypeRatios(dat)
}
\arguments{
  \item{dat}{phylotypes vs. samples data matrix in log10
  scale}
}
\value{
  phylotype pairs x samples matrix indicating the ratio (in
  log10 domain) between each unique pair
}
\description{
  Calculate phylotype ratios (eg. Bacteroides vs.
  Prevotella etc.) for a given phylotypes vs. samples
  matrix
}
\examples{
#
}
\author{
  Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

