% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_baseline.R
\name{pick_baseline}
\alias{pick_baseline}
\title{Pick Baseline Timepoint Samples}
\usage{
pick_baseline(x, na.omit = TRUE)
}
\arguments{
\item{x}{phyloseq object. Assuming that the sample_data(x) has the fields "time", "sample" and "subject"}

\item{na.omit}{Logical. Ignore samples with no time point information. If this is FALSE, the first sample for each subject is selected even when there is no time information.}
}
\value{
Phyloseq object with only baseline time point samples selected.
}
\description{
Identify and select the baseline timepoint samples in a phyloseq object.
}
\details{
Arranges the samples by time and picks the first sample for each subject.
}
\examples{
 data(atlas1006)
 a <- pick_baseline(atlas1006)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

