% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_abundances.R
\name{plot_abundances}
\alias{plot_abundances}
\title{Taxonomic Abundance Plot}
\usage{
plot_abundances(x, level = "Phylum", tax = "Bacteroidetes")
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{level}{Taxonomic level to show.}

\item{tax}{Taxonomic group to use in sorting.}
}
\value{
\code{\link{ggplot}} object
}
\description{
Plot taxon abundance for samples.
}
\examples{
data(dietswap)
pseq <- subset_samples(dietswap, group == 'DI' & nationality == 'AFR' &
   sex == "female")
p <- plot_abundances(pseq)
}
\keyword{utilities}
