% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rda.R
\name{plot_bagged_rda}
\alias{plot_bagged_rda}
\title{Description: Bagged RDA visualization}
\usage{
plot_bagged_rda(x, which.bac = 1:nrow(Bag.res$loadings), ptype = "spider",
  comp = 1:2, cex.bac = 0.5, plot.names = T,
  group.cols = as.numeric(unique(Y)), ...)
}
\arguments{
\item{x}{Output from bagged_rda}

\item{which.bac}{TBA}

\item{ptype}{TBA}

\item{comp}{TBA}

\item{cex.bac}{TBA}

\item{plot.names}{Plot names}

\item{group.cols}{TBA}

\item{...}{Other arguments to be passed


Returns:}
}
\value{
TBA
}
\description{
Arguments:
}
\examples{
#
  #library(microbiome)
  #data(peerj32)
  #x <- t(peerj32$microbes)
  #y <- factor(peerj32$meta$time); names(y) <- rownames(peerj32$meta)
  #res <- bagged_rda(x, y, sig.thresh=0.05, nboot=100)
  #plot_bagged_rda(res)
}
\author{
Contact: Jarkko Salojarvi \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

