% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{plot_density}
\usage{
plot_density(x, otu.name = NULL, log10 = FALSE, adjust = 1,
  kernel = "gaussian", trim = FALSE, na.rm = FALSE, fill = "gray")
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object or an OTU matrix (samples x phylotypes)}

\item{otu.name}{OTU name to visualize}

\item{log10}{Logical. Show log10 abundances or not.}

\item{adjust}{see stat_density}

\item{kernel}{see stat_density}

\item{trim}{see stat_density}

\item{na.rm}{see stat_density}

\item{fill}{Fill color}
}
\value{
A \code{\link{ggplot}} plot object.
}
\description{
Plot taxon density for samples
}
\examples{
# 
}
\keyword{utilities}

