% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diversity.R
\name{plot_diversity}
\alias{plot_diversity}
\title{plot_diversity}
\usage{
plot_diversity(physeq, x = "group", title = "", measures = "Shannon",
  nrow = 1, scales = "free_y", det.th = 0)
}
\arguments{
\item{physeq}{\code{\link{phyloseq-class}} object}

\item{x}{A variable to map to the horizontal axis. The vertical
axis will be mapped to the alpha diversity index/estimate
and have units of total taxa, and/or index value (dimensionless).
This parameter (\code{x}) is a character string indicating a
in the dataset (nsamples(physeq)).}

\item{title}{Optional title for the graphic.}

\item{measures}{Default is \code{NULL}. In this case
all available alpha-diversity measures will be included.
Alternatively, you can specify one or more measures
as a character vector. Values must be among those supported:
\code{c("Observed", "Chao1", "ACE", "Shannon", "Simpson", "InvSimpson", "Fisher")}.}

\item{nrow}{Number of rows for plot faceting.}

\item{scales}{scales for the plot}

\item{det.th}{Detection threshold for the diversity measure 'Observed' 
(ie. species richness). See \code{\link{estimate_diversity}}}
}
\value{
A \code{\link{ggplot}} plot object summarizing
 the richness estimates, and their standard error.
}
\description{
Plot alpha diversity with ggplot2

This function estimates a number of alpha-diversity metrics using the 
\code{\link{estimate_richness}} function,
and returns a \code{ggplot} object. 
The plot generated by this function will include every sample
in \code{physeq}, but they can be further grouped on the horizontal axis
through the argument to \code{x}, 
and shaded according to the argument to \code{color} (see below).
You must use untrimmed, non-normalized count data for meaningful results.
}
\examples{
# 
}
\seealso{
\code{\link{estimate_richness}}
 \code{\link{estimate_diversity}}
 \code{\link{plot_richness}}
 \code{\link[vegan]{estimateR}}
 \code{\link[vegan]{diversity}}
}
\keyword{utilities}

