% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_landscape.R
\name{plot_landscape}
\alias{plot_landscape}
\title{Landscape Plot}
\usage{
plot_landscape(x, method = "NMDS", distance = "bray", col = NULL,
  main = NULL, x.ticks = 10, rounding = 0, add.points = TRUE,
  adjust = 1, size = 1, legend = FALSE)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object or a data matrix 
(features x samples; eg. HITChip taxa vs. samples)}

\item{method}{Ordination method, see phyloseq::plot_ordination}

\item{distance}{Ordination distance, see phyloseq::plot_ordination}

\item{col}{Variable name to highlight samples (points) with colors}

\item{main}{title text}

\item{x.ticks}{Number of ticks on the X axis}

\item{rounding}{Rounding for X axis tick values}

\item{add.points}{Plot the data points as well}

\item{adjust}{Kernel width adjustment}

\item{size}{point size}

\item{legend}{plot legend TRUE/FALSE}
}
\value{
A \code{\link{ggplot}} plot object.
}
\description{
Plot abundance landscape ie. sample density in 2D
projection landscape
}
\details{
For consistent results, set random seet (set.seed) before
function call
}
\examples{
data(dietswap)
p <- plot_landscape(abundances(transform(dietswap, "log10"))[, 1:2])
}
\keyword{utilities}
