% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_longitudinal.R
\name{plot_longitudinal}
\alias{plot_longitudinal}
\title{Longitudinal time series plot}
\usage{
plot_longitudinal(x, taxon, subjectID = NULL, tipping.point = NULL,
  shift = 0.001)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{taxon}{Selected taxonomic group.}

\item{subjectID}{Selected subject.}

\item{tipping.point}{Optional. Indicate critical point for abundance variations to be highlighted.}

\item{shift}{Small constant to avoid problems with zeroes in log10}
}
\value{
\code{\link{ggplot}} object
}
\description{
Plot longitudinal time series.
}
\details{
Assuming the sample_data(x) has 'subject' field and
some subjects have multiple time points.
}
\examples{
\dontrun{
  #data("atlas1006")
  #pseq <- atlas1006
  #pseq <- transform_phyloseq(pseq, "relative.abundance")
  #p <- plot_longitudinal(pseq, "Dialister")}
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

