% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot_ordn}
\alias{plot_ordn}
\title{Ordination plot}
\usage{
plot_ordn(physeq, ordination, type = "samples", axes = 1:2, color = NULL,
  shape = NULL, label = NULL, title = NULL, show.density = TRUE)
}
\arguments{
\item{physeq}{(Required). \code{\link{phyloseq-class}}. 
The data about which you want to plot and annotate the ordination.}

\item{ordination}{(Required). An ordination object. Many different classes
of ordination are defined by \code{R} packages. Ordination classes
currently supported/created by the \code{\link{ordinate}} function are
supported here. There is no default, as the expectation is that the 
ordination will be performed and saved prior to calling this plot function.}

\item{type}{(Optional). The plot type. Default is \code{"samples"}. The
currently supported options are \code{c("samples", "sites", "species", "taxa", "biplot", "split", "scree")}. The option ``taxa'' is equivalent to ``species'' in this case, and similarly, ``samples'' is equivalent to ``sites''.  The options \code{"sites"} and \code{"species"} result in a single-plot of just the 
sites/samples or species/taxa of the ordination, respectively.
The \code{"biplot"} and \code{"split"} options result in a combined
plot with both taxa and samples, either combined into one plot (``biplot'')
or separated in two facet panels (``split''), respectively.
The \code{"scree"} option results in a call to \code{\link{plot_scree}},
which produces an ordered bar plot of the normalized eigenvalues
associated with each ordination axis.}

\item{axes}{(Optional). A 2-element vector indicating the axes of the 
ordination that should be used for plotting. 
Can be \code{\link{character-class}} or \code{\link{integer-class}},
naming the index name or index of the desired axis for the horizontal 
and vertical axes, respectively, in that order. The default value, 
\code{c(1, 2)}, specifies the first two axes of the provided ordination.}

\item{color}{(Optional). Default \code{NULL}. Character string.
The name of the variable to map to
colors in the plot. 
This can be a sample variable 
(among the set returned by \code{sample_variables(physeq)} )
or taxonomic rank (among the set returned by \code{rank_names(physeq)}).
The color scheme is chosen automatically
by \code{link{ggplot}},
but it can be modified afterward with an additional layer using
\code{\link[ggplot2]{scale_color_manual}}.}

\item{shape}{(Optional). Default \code{NULL}. Character string.
The name of the variable to map
to different shapes on the plot. 
Similar to \code{color} option, but for the shape if points.
The shape scale is chosen automatically by \code{link{ggplot}},
but it can be modified afterward with an additional layer using
\code{\link[ggplot2]{scale_shape_manual}}.}

\item{label}{(Optional). Default \code{NULL}. Character string.
The name of the variable to map to text labels on the plot.
Similar to \code{color} option, but for plotting text.}

\item{title}{(Optional). Default \code{NULL}. Character string.
The main title for the graphic.}

\item{show.density}{Show point density}
}
\value{
A \code{\link{ggplot}} plot object, graphically summarizing
 the ordination result for the specified axes.
}
\description{
General ordination plotting.
}
\details{
Modified from phyloseq::plot_ordination to add some features. For useful examples of phyloseq ordination graphics, see \href{http://joey711.github.io/phyloseq/plot_ordination-examples}{phyloseq online tutorials}.
}
\examples{
# See other examples at
# http://joey711.github.io/phyloseq/plot_ordination-examples
  data(GlobalPatterns)
  taxa <- names(sort(taxa_sums(GlobalPatterns), TRUE)[1:50])
  GP <- prune_taxa(taxa, GlobalPatterns)
  gp_bray_pcoa = ordinate(GP, "CCA", "bray")
  plot_ordn(GP, gp_bray_pcoa, "samples", color="SampleType")
}
\seealso{
Many related examples are included in the phyloseq plot_ordination page \href{http://joey711.github.io/phyloseq/plot_ordination-examples}{phyloseq online tutorials}.
}

