\name{PlotDiversity}
\alias{PlotDiversity}
\title{PlotDiversity}
\usage{
  PlotDiversity(dat, diversity.index, filename = NULL)
}
\arguments{
  \item{dat}{data matrix (phylotypes x samples) in original
  (non-log) scale}

  \item{diversity.index}{Diversity index: "shannon" /
  "invsimpson"}

  \item{filename}{output file name to store the diversity
  indices

  Returns:}
}
\value{
  diversity
}
\description{
  Description: Plot and save diversity indices Based on ST
  01-03-2011
}
\details{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

