\name{plotGroupPCA}
\alias{plotGroupPCA}
\title{Description:}
\usage{
  plotGroupPCA(dPCA, groups, main = "PCA plot", xPC = 1,
    yPC = 2, type = "n", legLoc = "topright", w = 11,
    h = 11, ...)
}
\arguments{
  \item{dPCA}{TBA}

  \item{groups}{TBA}

  \item{main}{TBA}

  \item{xPC}{TBA}

  \item{yPC}{TBA}

  \item{type}{TBA}

  \item{legLoc}{TBA}

  \item{w}{TBA}

  \item{h}{TBA}

  \item{...}{further parameters to be passed

  Returns:}
}
\value{
  plot
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

