\name{plotOligoP}
\alias{plotOligoP}
\title{Description: Plot oligoprofile figures without data retrieval from the db}
\usage{
  plotOligoP(d, tax.level = "L1", metric = "correlation",
    include.tree = T, fontsize = 11, figureratio = 15,
    phylogeny.info = phylogeny.infoJN, width = 12,
    height = 20, phylogeny.infoJN = NULL,
    plot.profile = NULL)
}
\arguments{
  \item{d}{data}

  \item{tax.level}{Taxonomic level}

  \item{metric}{metrics}

  \item{include.tree}{TBA}

  \item{fontsize}{font size}

  \item{figureratio}{figure ratio}

  \item{phylogeny.info}{phylogeny.info}

  \item{width}{width}

  \item{height}{height}

  \item{phylogeny.infoJN}{TBA}

  \item{plot.profile}{TBA

  Returns:}
}
\value{
  plot
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

