\name{plotPCA}
\alias{plotPCA}
\title{Description: Plot PCA with labels and explained variances}
\usage{
  plotPCA(dPCA, main = "PCA plot", xPC = 1, yPC = 2,
    showLabels = T, type = "n", ...)
}
\arguments{
  \item{dPCA}{TBA}

  \item{main}{TBA}

  \item{xPC}{TBA}

  \item{yPC}{TBA}

  \item{showLabels}{TBA}

  \item{type}{TBA}

  \item{...}{further parameters to be passed

  Returns:}
}
\value{
  TBA
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

