\name{plotScatter}
\alias{plotScatter}
\title{Description: plot scatter}
\usage{
  plotScatter(sampleA, sampleB, d, cex = 1)
}
\arguments{
  \item{sampleA}{TBA}

  \item{sampleB}{TBA}

  \item{d}{TBA}

  \item{cex}{TBA

  Returns:}
}
\value{
  scatterplot
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

