\name{polish.oligomap}
\alias{polish.oligomap}
\title{polish.oligomap}
\usage{
  polish.oligomap(oligomap)
}
\arguments{
  \item{oligomap}{oligomap data frame}
}
\value{
  polished oligomap
}
\description{
  Ensure oligomap is in correct format
}
\examples{
#polished.oligomap <- impute(oligomap)
}
\author{
  Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

