% Generated by roxygen2 (4.0.0): do not edit by hand
\name{polish.phylogeny.info}
\alias{polish.phylogeny.info}
\title{polish.phylogeny.info}
\usage{
polish.phylogeny.info(phylogeny.info)
}
\arguments{
\item{phylogeny.info}{phylogeny.info data frame}
}
\value{
polished phylogeny.info
}
\description{
Ensure phylogeny.info is in correct format
}
\examples{
#polished.phylogeny.info <- impute(phylogeny.info)
}
\author{
Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

