% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phylo.R
\name{polish.tax.table}
\alias{polish.tax.table}
\title{polish.tax.table}
\usage{
polish.tax.table(tax.table)
}
\arguments{
\item{tax.table}{tax.table data frame}
}
\value{
polished tax.table
}
\description{
Ensure tax.table is in correct format
}
\examples{
#tax.table <- GetPhylogeny('HITChip', 'filtered')
  #tax.table <- polish.tax.table(tax.table)
}
\author{
Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{internal}

