\name{populate.checkboxes}
\alias{populate.checkboxes}
\title{Description: low.level functions required by GUI}
\usage{
  populate.checkboxes(tt, title, var.names, var.init)
}
\arguments{
  \item{tt}{TBA}

  \item{title}{TBA}

  \item{var.names}{TBA}

  \item{var.init}{TBA

  Returns:}
}
\value{
  A list.
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

