\name{populate.radiobuttons}
\alias{populate.radiobuttons}
\title{Description: populate radiobuttons}
\usage{
  populate.radiobuttons(tt, title, var.names, var.values,
    var.init)
}
\arguments{
  \item{tt}{TBA}

  \item{title}{TBA}

  \item{var.names}{TBA}

  \item{var.values}{TBA}

  \item{var.init}{TBA

  Returns:}
}
\value{
  A list.
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

