% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential_bootstrap.R
\name{potential_bootstrap}
\alias{potential_bootstrap}
\title{Bootstrapped Potential Analysis}
\usage{
potential_bootstrap(x, detection, bw.adjust = 1, bs.iterations = 100,
  detection.limit = 1)
}
\arguments{
\item{x}{Input data vector}

\item{detection}{Mode detection}

\item{bw.adjust}{Bandwidth adjustment}

\item{bs.iterations}{Bootstrap iterations}

\item{detection.limit}{minimum accepted density for a maximum; as a multiple of kernel height}
}
\value{
List with following elements:
\itemize{
  \item{modes}{Number of modes for the input data vector (the most frequent number of modes from bootstrap)}
  \item{modes}{minima: Average of potential minima across the bootstrap samples (for the most frequent number of modes)}
  \item{modes}{maxima: Average of potential maxima across the bootstrap samples (for the most frequent number of modes)}
  \item{modes}{unimodality.support Fraction of bootstrap samples exhibiting unimodality}
}
}
\description{
Analysis of multimodality based on bootstrapped potential
   analysis of Livina et al. (2010) as described in Lahti et al. (2014).
}
\references{
\itemize{
  \item{}{Livina et al. (2010). Potential analysis reveals changing number of climate states during the last 60 kyr. \emph{Climate of the Past}, 6, 77-82.}
  \item{}{Lahti et al. (2014). Tipping elements of the human intestinal ecosystem. \emph{Nature Communications} 5:4344.}
 }
}

