\name{preproc.Hitchip}
\alias{preproc.Hitchip}
\title{Description: HITChip preprocessing function}
\usage{
  preproc.Hitchip(x)
}
\arguments{
  \item{x}{TBA

  Returns:}
}
\value{
  x
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

