% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalent.R
\name{prevalent_taxa}
\alias{prevalent_taxa}
\title{prevalent_taxa}
\usage{
prevalent_taxa(x, detection.threshold, prevalence.threshold)
}
\arguments{
\item{x}{A matrix or a x \code{\link{phyloseq}} object}

\item{detection.threshold}{Detection threshold for absence/presence.}

\item{prevalence.threshold}{Detection threshold for prevalence}
}
\description{
List prevalent groups
}
\details{
For phyloseq object, lists taxa that are more prevalent with the given detection
         threshold. For matrix, lists columns that satisfy these criteria.

  @return Vector of prevalent taxa names
}
\examples{
  #peerj32 <- download_microbiome("peerj32")
  #prevalent_taxa(peerj32$data$microbes, 10^1.8 + 100, 0.2) # matrix
  #prevalent_taxa(peerj32$physeq, 100, 0.2) # phyloseq object

}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
A Salonen et al. The adult intestinal core microbiota is determined by 
  analysis depth and health status. Clinical Microbiology and Infection 
  18(S4):16 20, 2012. 
  To cite the microbiome R package, see citation('microbiome')
}
\keyword{utilities}

