\name{project.data}
\alias{project.data}
\title{Description: Project high-dimensional data on two-dimensional plane by various methods}
\usage{
  project.data(amat, type = "PCA")
}
\arguments{
  \item{amat}{data matrix (samples x features)}

  \item{type}{projection type Returns:}
}
\value{
  projected data matrix
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

