% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiet.R
\name{quiet}
\alias{quiet}
\title{Quiet Output}
\usage{
quiet(expr, all = TRUE)
}
\arguments{
\item{expr}{Expression to run.}

\item{all}{If \code{TRUE} then suppress warnings and messages as well;
otherwise, only suppress printed output (such as from \code{print} or
\code{cat}).}
}
\value{
Used for its side effects.
}
\description{
Suppress all output from an expression. Works cross-platform.
}
\examples{
quiet(1 + 1)
}
\author{
Adapted from \url{https://gist.github.com/daattali/6ab55aee6b50e8929d89}
}
\keyword{internal}
