% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neatsort.R
\name{radial_theta}
\alias{radial_theta}
\title{Radial Theta Function}
\usage{
radial_theta(x)
}
\arguments{
\item{x}{position parameter}
}
\value{
theta
}
\description{
Adapted from \pkg{NeatMap} and \pkg{phyloseq} packages
  but not exported and hence not available via phyloseq. Completely
  rewritten to avoid license conflicts. Vectorized to gain efficiency;
  only calculates theta and omits r.
}
\keyword{internal}
