% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadialTheta.R
\name{RadialTheta}
\alias{RadialTheta}
\title{Radial theta function.}
\usage{
RadialTheta(pos)
}
\description{
Adapted from \pkg{NeatMap} and \pkg{phyloseq} packages. In phyloseq this was vectorized for speed and simplicity, also only calculates theta and not r. Copied here as it is not exported and hence not available via phyloseq
}
\keyword{internal}

