% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare_abundance.R
\name{rare_abundance}
\alias{rare_abundance}
\title{Rare Abundance}
\usage{
rare_abundance(x, detection = 1, prevalence = 95)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{detection}{Detection threshold (non-negative real)}

\item{prevalence}{Prevalence threshold (in [0, 100])}
}
\value{
Total rare abunance vector.
}
\description{
Calculate total rare abundance
}
\examples{
 data(dietswap)
 a <- rare_abundance(dietswap)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

