% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare_members.R
\name{rare_members}
\alias{rare_members}
\title{Rare Taxa}
\usage{
rare_members(x, detection = 1, prevalence = 95, sort = TRUE)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{detection}{Detection threshold (non-negative real)}

\item{prevalence}{Prevalence threshold (in [0, 100])}

\item{sort}{Logical. Sort the taxa.}
}
\value{
Vector of rare members
}
\description{
Determine members of the rare microbiota with given abundance
'             and prevalences.
}
\details{
For phyloseq object, lists taxa that are more prevalent with the
  given detection. For matrix, lists columns that satisfy
  these criteria.
}
\examples{
  data(dietswap)
  a <- rare_members(dietswap, 1, 95)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
To cite the microbiome R package, see citation('microbiome')
}
\keyword{utilities}

