% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare_richness.R
\name{rare_richness}
\alias{rare_richness}
\title{Rare Richness}
\usage{
rare_richness(x, detection = 1, prevalence = 95)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{detection}{Detection threshold (non-negative real)}

\item{prevalence}{Prevalence threshold (in [0, 100])}
}
\value{
Total rare abundance vector.
}
\description{
Calculate total rare richness
}
\examples{
 data(dietswap)
 a <- rare_richness(dietswap)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

