% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rda.R
\name{rda_physeq}
\alias{rda_physeq}
\title{rda_physeq}
\usage{
rda_physeq(x, varname, scale = FALSE, na.action = na.fail)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{varname}{Variable to apply in RDA visualization.}

\item{scale}{See help(rda)}

\item{na.action}{See help(rda)

Returns:}
}
\value{
rda result. See help(vegan::rda)
}
\description{
RDA for phyloseq objects. Based on the \code{\link{rda}} function
from the \pkg{vegan} package.
}
\details{
Arguments:
}
\examples{
#
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

