\name{read.hitchip}
\alias{read.hitchip}
\title{Description: read hitchip}
\usage{
  read.hitchip(path = "../data/", dataID = "")
}
\arguments{
  \item{path}{Source path}

  \item{dataID}{data ID

  Returns:}
}
\value{
  d
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

