% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\name{read_hitchip}
\alias{read_hitchip}
\title{Read run.profiling.script output into R and preprocess into phyloseq format}
\usage{
read_hitchip(data.dir, output = "all", method = "rpa")
}
\arguments{
\item{data.dir}{Profiling script output directory for reading the data.
If not given, GUI will ask to specify the file and
   overruns the possible level / method arguments in the
   function call.}

\item{output}{Specify the desired output.}

\item{method}{Probe summarization method ("rpa" or "sum")}
}
\value{
data matrix (phylo x samples)
}
\description{
Read run.profiling.script output into R and preprocess into phyloseq format
}
\examples{
# data.dir <- system.file("extdata", package = "microbiome")
  # dat <- read_profiling(data.dir)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

