% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mothur2phyloseq.R
\name{read_mothur2phyloseq}
\alias{read_mothur2phyloseq}
\title{Read Mothur Output into a Phyloseq Object}
\usage{
read_mothur2phyloseq(shared.file, consensus.taxonomy.file,
  mapping.file = NULL)
}
\arguments{
\item{shared.file}{A \href{http://www.mothur.org/wiki/Shared_file}{shared file} produced by \emph{mothur}. Identified from the .shared extension}

\item{consensus.taxonomy.file}{\href{http://www.mothur.org/wiki/ConTaxonomy_file}{consensus taxonomy file} produced by \emph{mothur}. Identified from with the .taxonomy extension.}

\item{mapping.file}{Metadata/mapping file with .csv extension}
}
\value{
\code{\link{phyloseq-class}} object.
}
\description{
Read mothur shared and consensus taxonomy files into a \code{\link{phyloseq-class}} object.
}
\details{
Mothur shared and consensus taxonomy files will be converted to \code{\link{phyloseq-class}}.
}
\examples{
  \dontrun{
    # Example data
    library(microbiome)
    p0 <- read_mothur2phyloseq(
            shared.file = "mothur_shared.shared",
            consensus.taxonomy.file = "mothur_taxonomy.taxonomy",
       mapping.file = "mothur_mapping.csv")
  }
}
\author{
Sudarshan A. Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{utilities}
