% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\name{read_profiling}
\alias{read_profiling}
\title{read_profiling}
\usage{
read_profiling(data.dir, method)
}
\arguments{
\item{data.dir}{Profiling script output directory for reading the data.
If not given, GUI will ask to specify the file and
   overruns the possible level / method arguments in the
   function call.}

\item{method}{Select the preprocessing method that you like to check}
}
\value{
data matrix (phylo x samples)
}
\description{
Read run.profiling.script output into R
}
\examples{
# data.dir <- system.file("extdata", package = "microbiome")
  # dat <- read_profiling(data.dir)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

