% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readcount.R
\name{readcount}
\alias{readcount}
\title{Total Read Count}
\usage{
readcount(x)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}
}
\value{
Vector of read counts.
}
\description{
Total Read Count
}
\examples{
data(dietswap)
d <- readcount(dietswap)
}
\references{
See citation('microbiome')
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
