\name{ReadHclustParameters}
\alias{ReadHclustParameters}
\title{Description: Read parameters for PROFILE PLOT FUNCTION}
\usage{
  ReadHclustParameters(dat, data.dir, ppcm = 150,
    hclust.method = "complete", pal = "white/blue",
    lev = "level 2",
    clmet = "Pearsons correlation coefficient",
    tree.display = "yes", figureratio = 12, fontsize = 12)
}
\arguments{
  \item{dat}{data matrix}

  \item{data.dir}{data directory}

  \item{ppcm}{plotting parameter}

  \item{hclust.method}{hierarchical clustering method}

  \item{pal}{color palette}

  \item{lev}{phylogeny level}

  \item{clmet}{clustering metrics}

  \item{tree.display}{tree.display}

  \item{figureratio}{figureratio}

  \item{fontsize}{fontsize

  Returns:}
}
\value{
  plots the profile
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

