\name{ReadParameters}
\alias{ReadParameters}
\title{Description: Define parameters in select box}
\usage{
  ReadParameters(con)
}
\arguments{
  \item{con}{Output from dbConnect(dbDriver("MySQL"),
  username = dbuser, password = dbpwd, dbname =
  'PhyloArray')

  Returns:}
}
\value{
  list with defined parameters
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

