\name{relative.abundance}
\alias{relative.abundance}
\title{relative.abundance}
\usage{
  relative.abundance(dat, det.th = 0)
}
\arguments{
  \item{dat}{data matrix (phylotypes x samples) in original
  (non-log) scale}

  \item{det.th}{detection threshold.

  Returns:}
}
\value{
  Vector containing relative proportions for each phylotype
  in each sample
}
\description{
  Description: Estimate relative abundance for each
  phylotype in each sample with a given threshold
}
\details{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

