\name{remove.background}
\alias{remove.background}
\title{Description: simple background correction}
\usage{
  remove.background(dat)
}
\arguments{
  \item{dat}{data matrix

  Returns:}
}
\value{
  background corrected data matrix
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

