% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_taxa.R
\name{remove_taxa}
\alias{remove_taxa}
\title{Exclude Taxa}
\usage{
remove_taxa(taxa = NULL, x)
}
\arguments{
\item{taxa}{Names of taxa to be removed.}

\item{x}{\code{\link{phyloseq-class}} object}
}
\value{
Filtered phyloseq object.
}
\description{
Filter out selected taxa from a phyloseq object.
}
\details{
This complements the phyloseq function prune_taxa by providing
a way to exclude given groups from a phyloseq object.
}
\examples{
data(dietswap)
pseq <- remove_taxa(c("Akkermansia", "Dialister"), dietswap)
}
\references{
To cite the microbiome R package, see citation('microbiome')
}
\seealso{
phyloseq::prune_taxa, phyloseq::subset_taxa
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
