\name{roc.auc}
\alias{roc.auc}
\title{Description: ROC AUC calculation}
\usage{
  roc.auc(ordered.results, true.positives)
}
\arguments{
  \item{ordered.results}{Items ordered from best to worst
  according to the test score.}

  \item{true.positives}{known true positives

  Returns:}
}
\value{
  ROC AUC value
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

