\name{roc.plot}
\alias{roc.plot}
\title{Description: Plot ROC curve}
\usage{
  roc.plot(ordered.results, true.positives, line = F,
    title = "")
}
\arguments{
  \item{ordered.results}{Items ordered from best to worst
  according to the test score.}

  \item{true.positives}{known true positives}

  \item{line}{Draw 45 angle line}

  \item{title}{Title text Returns:}
}
\value{
  Used for its side effects (plot)
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

