\name{roc}
\alias{roc}
\title{Description: ROC analysis returning true and false positive rates
along an ordered list}
\usage{
  roc(ordered.results, true.positives)
}
\arguments{
  \item{ordered.results}{Items ordered from best to worst
  according to the test score.}

  \item{true.positives}{known true positives

  Returns:}
}
\value{
  List: true positive rate (tpr) and false positive rate
  (fpr)
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

