\name{run.profiling.script}
\alias{run.profiling.script}
\title{run.profiling.script}
\usage{
  run.profiling.script(dbuser, dbpwd, dbname,
    verbose = TRUE)
}
\arguments{
  \item{dbuser}{MySQL username}

  \item{dbpwd}{MySQL password}

  \item{dbname}{MySQL database name}

  \item{verbose}{verbose

  Returns:}
}
\value{
  Profiling parameters. Also writes output to the
  user-specified directory.
}
\description{
  Description: Profiling main script
}
\details{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

