\name{runPCA}
\alias{runPCA}
\title{Description: PCA function}
\usage{
  runPCA(d, level, c1, c2, data.labels = NULL, write.dir,
    group.by = NULL, ...)
}
\arguments{
  \item{d}{TBA}

  \item{level}{TBA}

  \item{c1}{TBA}

  \item{c2}{TBA}

  \item{data.labels}{Optional}

  \item{write.dir}{Output directory path}

  \item{group.by}{Optional}

  \item{...}{Other parameters to be passed to the function

  Returns:}
}
\value{
  A list.
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

