\name{sampleReproducibility}
\alias{sampleReproducibility}
\title{sampleReproducibility function}
\usage{
  sampleReproducibility(con, sampleID)
}
\arguments{
  \item{con}{MySQL connection}

  \item{sampleID}{sampleID}
}
\value{
  TBA
}
\description{
  sampleReproducibility function
}
\examples{
# TBA
}
\author{
  Douwe Molenaar. Maintainer: Leo Lahti
  \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

