% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tables.R
\name{save_tables}
\alias{save_tables}
\title{Saving Various Files as Comma Seperated Files}
\usage{
save_tables(x, type)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{type}{"OTU" or "TAXA" or "METADATA"}
}
\value{
Output file path (a string)
}
\description{
This command will save the otu,taxonomy and meta data as *.csv file for manual inspection.
}
\examples{
\dontrun{
  library(microbiome)
  data(dietswap)
  pseq <- dietswap
  save_tables(pseq, "OTU")
  save_tables(pseq, "TAXA")
  save_tables(pseq, "METADATA")
 }
}
\keyword{utilities}
