\name{SplitTrainTest}
\alias{SplitTrainTest}
\title{Description: Split the data in training and test samples}
\usage{
  SplitTrainTest(x, y, spec)
}
\arguments{
  \item{x}{x}

  \item{y}{y}

  \item{spec}{species

  Returns:}
}
\value{
  splitted data
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

