\name{submitDB}
\alias{submitDB}
\title{Generic method for submission of calculated data to the database}
\usage{
  submitDB(x)
}
\arguments{
  \item{x}{TBA}
}
\value{
  TBA
}
\description{
  Generic method for submission of calculated data to the
  database
}
\examples{
# TBA
}
\author{
  Douwe Molenaar. Maintainer: Leo Lahti
  \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

