\name{summarize.oligos}
\alias{summarize.oligos}
\title{Description: summarize.oligos EXPERIMENTAL}
\usage{
  summarize.oligos(oligo.matrix, oligomap, level = "L2")
}
\arguments{
  \item{Data}{Data}

  \item{PH.i}{PH.i}

  \item{level}{taxonomic level

  Returns:}
}
\value{
  list
}
\description{
  For cross-hyb control
}
\details{
  Arguments:
}
\author{
  Contact: Jarkko Salojarvi
  \email{jarkko.salojarvi@helsinki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

