% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.sum.R
\name{summarize.sum}
\alias{summarize.sum}
\title{Sum-based probe summarization}
\usage{
summarize.sum(taxonomy, level, probedata, verbose = TRUE,
  downweight.ambiguous.probes = TRUE)
}
\arguments{
\item{taxonomy}{oligo - phylotype matching data.frame}

\item{level}{taxonomic level for the summarization.}

\item{probedata}{preprocessed probes x samples data matrix in absolute domain}

\item{verbose}{print intermediate messages}

\item{downweight.ambiguous.probes}{Downweight probes with multiple targets}
}
\value{
List with two elements: abundance.table (summarized data matrix in absolute scale) and probe.parameters used in the calculations
}
\description{
Probeset summarization with SUM
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

