% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_phyloseq.R
\name{summarize_phyloseq}
\alias{summarize_phyloseq}
\title{Summarize phyloseq object}
\usage{
summarize_phyloseq(x)
}
\arguments{
\item{x}{Input is a \code{\link{phyloseq-class}} object.}
}
\value{
Prints basic information of \code{\link{phyloseq-class}} object.
}
\description{
Prints basic information of data.
}
\details{
The summarize_phyloseq function will give information on weather
   data is compositional or not, reads (min. max, median, average),
   sparsity, 
presence of singletons and sample variables.
}
\examples{
data(dietswap)
summarize_phyloseq(dietswap)
}
\author{
Contact: Sudarshan A. Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{utilities}
