% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarize_probedata.R
\name{summarize_probedata}
\alias{summarize_probedata}
\title{Summarize phylogenetic microarray probe-level data from given input folder.}
\usage{
summarize_probedata(data.dir = NULL, probedata = NULL, taxonomy = NULL,
  level, method)
}
\arguments{
\item{data.dir}{Data folder.}

\item{probedata}{probe-level data matrix}

\item{taxonomy}{probe taxonomy}

\item{level}{Summarization level}

\item{method}{Summarization method}
}
\value{
data matrix (taxa x samples)
}
\description{
Summarize phylogenetic microarray probe-level data from given input folder.
}
\examples{
#
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

