% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_taxa.R
\name{summarize_taxa}
\alias{summarize_taxa}
\title{Summarize taxa}
\usage{
summarize_taxa(pseq, level)
}
\arguments{
\item{pseq}{\code{\link{phyloseq-class}} object}

\item{level}{Summarization level (from \code{rank_names(pseq)})}
}
\value{
Summarized phyloseq object
}
\description{
Summarize phyloseq data into a higher phylogenetic level.
}
\details{
This provides a convenient way to aggregate phyloseq OTUs (or other taxa) when the phylogenetic tree is missing. Calculates the sum of OTU abundances over all OTUs that map to the same higher-level group. Removes ambiguous levels from the taxonomy table. Returns a phyloseq object with the summarized abundances.
}
\examples{
# summarize_taxa(pseq, "Genus")
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

