% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{taxa}
\alias{taxa}
\title{Taxa Names}
\usage{
taxa(x)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}
}
\value{
A vector with taxon names.
}
\description{
List the names of taxonomic groups in a phyloseq or
SummarizedExperiment-derived object.
}
\details{
A handy shortcut for phyloseq::taxa_names, with a potential to add
to add some extra tweaks later.
}
\examples{
data(dietswap)
x <- taxa(dietswap)
}
\references{
To cite the microbiome R package, see citation('microbiome')
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
