\name{theme_bottom_border}
\alias{theme_bottom_border}
\title{Description: ggplot2::ggplot2 theme}
\usage{
  theme_bottom_border(colour = "black", size = 1,
    linetype = 1)
}
\arguments{
  \item{colour}{colour}

  \item{size}{size}

  \item{linetype}{linetype Returns:}
}
\value{
  set theme
}
\description{
  Arguments:
}
\examples{
theme_bottom_border()
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

