% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_normalize.R
\name{time_normalize}
\alias{time_normalize}
\title{Normalize Phyloseq Metadata Time Field}
\usage{
time_normalize(x)
}
\arguments{
\item{x}{phyloseq object. The sample_data(x) should contain the following
fields: subject, time}
}
\value{
Phyloseq object with a normalized time field
}
\description{
Shift the time field in phyloseq sample_data such that the
first time point of each subject is always 0.
}
\examples{
data(peerj32)
pseq <- time_normalize(peerj32$phyloseq)
}
