% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesort_subjects.R
\name{timesort_subjects}
\alias{timesort_subjects}
\title{Temporal Sorting Within Subjects}
\usage{
timesort_subjects(x)
}
\arguments{
\item{x}{A metadata data.frame including the following columns: time, subject, sample, signal}
}
\value{
A list with metadata (data.frame) for each subject.
}
\description{
Within each subject, sort samples by time and calculate distance from the baseline point (minimum time).
}
\examples{
\dontrun{timesort_subjects(x) }
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

