% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_abundance.R
\name{top_abundance}
\alias{top_abundance}
\title{Top Abundance Index}
\usage{
top_abundance(x, rank = 1, aggregate = TRUE, split = TRUE)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{rank}{Optional. The rank of the dominant taxa to consider.}

\item{aggregate}{Aggregate (TRUE; default) the top members or not (FALSE) ?}

\item{split}{(Optional). Logical. Should a separate set of richness
estimates be performed for each sample? Or alternatively,
pool all samples and estimate richness of the entire set.}
}
\value{
A vector of top_abundance indices
}
\description{
Calculates the community top_abundance index.
}
\details{
The top_abundance index gives the abundance of the most abundant species in [0,1]. This simple diversity index is occasionally used in ecological literature, and sometimes also called dominance. However, note that the microbiome::dominance function uses a different definition. With rank = 2, the sum of abundances for the two most abundant taxa are returned etc. However, if aggregate=FALSE, the abundance for the single n'th most dominant taxa (n = rank) is returnde instead the sum of abundances up to that rank.
}
\examples{
  data(dietswap)
  d <- top_abundance(dietswap)
}
\seealso{
dominance, diversity
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
