% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_taxa.R
\name{top_taxa}
\alias{top_taxa}
\title{Top Taxa}
\usage{
top_taxa(x, n = .ntaxa(x))
}
\arguments{
\item{x}{phyloseq object}

\item{n}{Number of top taxa to return (default: all)}
}
\value{
Character vector listing the top taxa
}
\description{
Return n most abundant taxa (based on total abundance over
all samples), sorted by abundance
}
\examples{
data(dietswap)
topx <- top_taxa(dietswap, n=10)
}
