\docType{data}
\name{toydata-hitchip}
\alias{toydata-hitchip}
\title{Toydata-hitchip data set (phylotype composition)}
\format{Contains the following objects: 
  \describe{ 
    \item{species.matrix}{species-like phylotype data matrix: species vs. samples}
    \item{genus.matrix}{genus-like (level2) phylotype data matrix: species vs. samples}
    \item{phylum.matrix}{phylum-like (level1) phylotype data matrix: species vs. samples}
  }
}
\source{Simulations}
\description{Toy data sets for species-, genus- and phylum-level phylotype matrices (phylotypes vs. samples). Intended to test analysis functions.}
\examples{
data(toydata-hitchip)
}
\references{See citation("microbiome")}
\keyword{datasets}

