\name{trans}
\alias{trans}
\title{Description: transparent color}
\usage{
  trans(color, tr = 150)
}
\arguments{
  \item{color}{color}

  \item{tr}{transparency level

  Returns:}
}
\value{
  TBA
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

