% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate Phyloseq}
\usage{
validate(x)
}
\arguments{
\item{x}{phyloseq object}
}
\value{
A validated and polished phyloseq object
}
\description{
Validate phyloseq object.
}
\details{
Checks that the abundances and sample_data have exactly same samples.
}
\examples{
  library(microbiome)
  data(atlas1006)
  validate(atlas1006)
}
