% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variability_individual.R
\name{variability_individual}
\alias{variability_individual}
\title{Individual Variability}
\usage{
variability_individual(x, method = "CoV")
}
\arguments{
\item{x}{phyloseq object}

\item{method}{"CoV" (coef. of variation ie. std/mean) or "std" (standard deviation) or "timevar" (average shift normalized by time)}
}
\value{
A vector of average variability within individuals for each taxa.
}
\description{
Average coefficient of variation (CoV) within individuals for each taxa.
}
\details{
The sample metadata should contain the subject and time fields
}
\examples{
  library(microbiome)
  data(atlas1006)
  cv <- variability_individual(atlas1006)
}

