% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_members.R
\name{variable_members}
\alias{variable_members}
\title{Variable Microbiota}
\usage{
variable_members(x, detection, prevalence, include.lowest = FALSE)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{detection}{Detection threshold for absence/presence
(strictly greater by default).}

\item{prevalence}{Prevalence threshold (in [0, 1];
strictly greater by default)}

\item{include.lowest}{Include the lower boundary of the detection and
prevalence cutoffs. FALSE by default.}
}
\value{
Filtered phyloseq object including only variable taxa.
}
\description{
Filter the phyloseq object to include only variable taxa.
}
\details{
Variable taxa are members of the microbiota that are not rare, and
not members of the core microbiota. Such taxa reach notable abundance in a
notable fraction of the population. The variable members are calculated by
using the indicated prevalence threshold to define rare taxa; and its
complement 1-prevalence to define core taxa. The remaining taxa are
considered variable. The same detection threshold is used in all cases.
}
\examples{
  data(dietswap)
  # Detection threshold 0 (strictly greater by default);
  # Prevalence threshold 50 percent (strictly greater by default)
  pseq.rel <- transform(dietswap, 'compositional')
  pseq <- variable_members(pseq.rel, 0.5/100, 20/100)

}
\references{
To cite the microbiome R package, see citation('microbiome')
}
\seealso{
core_members, core, rare
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
