\name{wardClust}
\alias{wardClust}
\title{Description: wardClust is for heatmap function for Ward-clustering}
\usage{
  wardClust(x)
}
\arguments{
  \item{x}{data matrix

  Returns:}
}
\value{
  hclust object
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

