% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_phyloseq.R
\name{write_phyloseq}
\alias{write_phyloseq}
\title{Saving Various Files as Comma Seperated Files}
\usage{
write_phyloseq(x, type)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{type}{"OTU" or "TAXA" or "METADATA"}
}
\value{
Output file path (a string)
}
\description{
This command will save the otu,taxonomy and meta data as *.csv file for manual inspection.
}
\examples{
\dontrun{
  library(microbiome)
  data(dietswap)
  pseq <- dietswap
  write_phyloseq(pseq, "OTU")
  write_phyloseq(pseq, "TAXA")
  write_phyloseq(pseq, "METADATA")
 }
}
\keyword{utilities}
