% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{ztransform}
\alias{ztransform}
\title{Phyloseq Z Transformation}
\usage{
ztransform(x, which)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{which}{Specify Z transform for "sample" or "OTU"}
}
\value{
Z-transformed phyloseq object
}
\description{
Z transform phyloseq objects.
}
\details{
Performs centering (to zero) and scaling (to unit
  variance) across samples for each taxa.
}
\examples{
\dontrun{
  data(peerj32)
  pseqz <- ztransform(peerj32$phyloseq)
}
}
\references{
See citation('microbiome')
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{internal}
