% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otu_table.R
\name{abundances}
\alias{ab}
\alias{abundances}
\title{Abundance matrix from phyloseq object}
\usage{
abundances(x)
}
\arguments{
\item{x}{\code{\link{phyloseq}} object}
}
\value{
Abundance matrix.
}
\description{
Retrieves the taxon abundance table from
   \code{\link{phyloseq-class}} object and ensures it is returned as
   taxa x samples matrix.
}
\examples{
  data(dietswap)
  abundances(dietswap)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

