% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundances.R
\name{abundances}
\alias{ab}
\alias{abundances}
\title{Abundance matrix from phyloseq object}
\usage{
abundances(x, transform = "identity")
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{transform}{Transformation to apply. The options include:
'compositional' (ie relative abundance), 'Z', 'log10', 'hellinger',
'identity', 'clr', 'ilr',
 or any method from the vegan::decostand function.}
}
\value{
Abundance matrix.
}
\description{
Retrieves the taxon abundance table from
   \code{\link{phyloseq-class}} object and ensures it is returned as
   taxa x samples matrix.
}
\examples{
  data(dietswap)
  abundances(dietswap)
  abundances(dietswap, transform = "identity")
  abundances(dietswap, transform = "compositional")
  abundances(dietswap, transform = "clr")
  abundances(dietswap, transform = "Z")
  abundances(dietswap, transform = "log10")
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

