% Generated by roxygen2 (4.0.1): do not edit by hand
\name{add.heatmap}
\alias{add.heatmap}
\title{add.heatmap
Description: Add oligprofile heatmap into output directory}
\usage{
add.heatmap(dat, output.dir, output.file = NULL, phylogeny.info, ppcm = 150,
  hclust.method = "complete", palette = "white/black", level = "L1",
  metric = "pearson", figureratio = 10, fontsize = 40,
  tree.display = TRUE)
}
\arguments{
\item{dat}{oligoprofile data in original (non-log) domain}

\item{output.dir}{output data directory}

\item{output.file}{output file name}

\item{phylogeny.info}{oligo-phylotype mappings}

\item{ppcm}{figure size}

\item{hclust.method}{hierarchical clustering method}

\item{palette}{color palette ("white/black" / "white/blue" / "black/yellow/white")}

\item{level}{taxonomic level to show}

\item{metric}{clustering metric}

\item{figureratio}{figure ratio}

\item{fontsize}{font size}

\item{tree.display}{tree.display

Returns:}
}
\value{
Plotting parameters
}
\description{
Arguments:
}
\examples{
# data(peerj32); hc <- add.heatmap(peerj32$microbes[, 1:4])
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

