% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bagged.RDA.R
\name{Bagged.RDA}
\alias{Bagged.RDA}
\title{Bagged RDA}
\usage{
Bagged.RDA(X, Y, boot = 1000)
}
\arguments{
\item{X}{a matrix, samples on columns, variables (bacteria) on rows.}

\item{Y}{vector with names(Y)=rownames(X), for example}

\item{boot}{Number of bootstrap iterations}
}
\value{
List with items:
  	     loadings: bagged loadings
	     scores: bagged scores
	     significance: significances of X variables
	     etc. TBA.
}
\description{
Bootstrap solutions that follows the Jack-knife estimation of PLS by Martens and Martens, 2000.  Solves rotational invariance of latent space by orthogonal procrustes rotations
}
\examples{
\dontrun{data(peerj32); x <- as.matrix(peerj32$microbes)[1:20, 1:6]; y <- rnorm(nrow(x)); names(y) <- rownames(x); res <- Bagged.RDA(x, y , boot = 5)}
}
\author{
Contact: Jarkko Salojarvi \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

