% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda_bagged.R
\name{Bagged.RDA}
\alias{Bagged.RDA}
\title{Bagged RDA Auxiliary Function}
\usage{
Bagged.RDA(X, Y, bs.iter = 100)
}
\arguments{
\item{X}{a matrix, samples on columns, variables (bacteria) on rows.}

\item{Y}{vector with names(Y)=rownames(X), for example}

\item{bs.iter}{Number of bootstrap iterations}
}
\value{
List with elements:
\itemize{
\item{loadings }{bagged loadings}
\item{scores }{bagged scores}
\item{significance }{significances of X variables}
}
}
\description{
Bootstrap solutions that follows the Jack-knife estimation of
PLS by Martens and Martens, 2000.  Solves rotational
invariance of latent space by orthogonal procrustes rotations.
}
\examples{
# Not exported
\dontrun{
data(peerj32)
x <- as.matrix(peerj32$microbes)[1:20, 1:6]
y <- rnorm(nrow(x))
names(y) <- rownames(x)
res <- Bagged.RDA(x, y , bs.iter=5)
}
}
\references{
See citation('microbiome')
}
\author{
Jarkko Salojarvi \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
