% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rda.R
\name{bagged_rda}
\alias{bagged_rda}
\title{Description: Bagged RDA feature selection}
\usage{
bagged_rda(x, y, sig.thresh = 0.1, nboot = 1000, verbose = T)
}
\arguments{
\item{x}{a matrix, samples on columns, variables (bacteria) on rows.
       Or a \code{\link{phyloseq-class}} object}

\item{y}{vector with names(y)=rownames(X).
         Or name of phyloseq sample data variable name.}

\item{sig.thresh}{signal p-value threshold, default 0.1}

\item{nboot}{Number of bootstrap iterations}

\item{verbose}{verbose

Returns:}
}
\value{
List with items:
  	     loadings: bagged loadings
	     scores: bagged scores
	     significance: significances of X variables
	     group.centers: group centers on latent space
	     bootstrapped: bootstrapped loadings
	     data: data set with non-significant components dropped out.
}
\description{
Arguments:
}
\examples{
#
#  library(microbiome)
#  data(peerj32)
#  x <- t(peerj32$microbes)
#  y <- factor(peerj32$meta$time); names(y) <- rownames(peerj32$meta)
#  res <- bagged_rda(x, y, sig.thresh=0.05, nboot=100)
#  PlotBaggedRDA(res, y)
}
\author{
Contact: Jarkko Salojarvi \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

