% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimodality.R
\name{bimodality}
\alias{bimodality}
\title{Bimodality Analysis}
\usage{
bimodality(x, method = "potential_analysis", detection = 1, bw.adjust = 1,
  bs.iter = 100, min.density = 1, verbose = TRUE)
}
\arguments{
\item{x}{A vector, matrix, or a phyloseq object}

\item{method}{bimodality quantification method ('potential_analysis'
or one of the methods in the function \code{\link{bimodality_sarle}})}

\item{detection}{Mode detection}

\item{bw.adjust}{Bandwidth adjustment}

\item{bs.iter}{Bootstrap iterations}

\item{min.density}{minimum accepted density for a maximum; as a multiple of kernel height}

\item{verbose}{Verbose}
}
\value{
A list with following elements:
  \itemize{
    \item{score}{Fraction of bootstrap samples where multiple modes are observed}
    \item{nmodes}{The most frequently observed number of modes in bootrstrap sampling results}
    \item{results}{Full results of potential_analysis for each row of the input matrix.}
  }
}
\description{
A wrapper to calculate bimodality scores.
}
\details{
\itemize{
    \item{Sarle.finite.sample}{Coefficient of bimodality for finite sample. See SAS 2012.}
    \item{Sarle.asymptotic}{Coefficient of bimodality, used and described in Shade et al. (2014) and Ellison AM (1987).}
    \item{potential_analysis}{Repeats potential analysis (Livina et al. 2010) multiple times with bootstrap sampling for each row of the input data (as in Lahti et al. 2014) and returns the bootstrap score.}
  }
}
\examples{
  
  bimodality(c(rnorm(100, mean = 0), rnorm(100, mean = 5)))
 
 # The classical DIP test:
 # quantifies unimodality. Values range between 0 to 1. 
 # dip.test(x, simulate.p.value = TRUE, B = 200)$statistic
 # Values less than 0.05 indicate significant deviation from unimodality. 
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
\itemize{
  \item{}{Livina et al. (2010). Potential analysis 
        reveals changing number of climate states during the last 60
	   kyr. \emph{Climate of the Past}, 6, 77-82.}
  \item{}{Lahti et al. (2014). Tipping elements of the human intestinal
        ecosystem. \emph{Nature Communications} 5:4344.}
  \item{}{Shade et al. mBio 5(4):e01371-14, 2014.}
  \item{}{AM Ellison, Am. J. Bot 74:1280-8, 1987.}
  \item{}{SAS Institute Inc. (2012). SAS/STAT 12.1 user's guide. Cary, NC.}
}
}
\seealso{
A classical test of multimodality is provided by dip.test in the \pkg{DIP} package.
}
\keyword{utilities}

