% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bimodality_score.R
\name{bimodality_score}
\alias{bimodality_score}
\title{Calculate a score based on a given statistical method for detecting bimodality.
The score can be computed using the Hartigan's Dip Test Statistic, coefficient of bimodality,
Silverman's Test or potential analysis. A low score (min = 0) infers unimodality, whereas a
high score (max = 1) infers multimodality (or bimodality when using coefficient of bimodality).}
\usage{
bimodality_score(x, method = "diptest")
}
\arguments{
\item{x}{Data vector or matrix (samples x phylotypes)}

\item{method}{Hartigan's Dip Test Statistic (diptest), coefficient of bimodality (cob),
Silverman's Test (silverman) or potential analysis (potential)}
}
\value{
Bimodality score
}
\description{
Calculate a score based on a given statistical method for detecting bimodality.
The score can be computed using the Hartigan's Dip Test Statistic, coefficient of bimodality,
Silverman's Test or potential analysis. A low score (min = 0) infers unimodality, whereas a
high score (max = 1) infers multimodality (or bimodality when using coefficient of bimodality).
}
\details{
Hartigan's dip test statistic as described in Hartigan PM
         App. Stat, 320-325, 1985.

         Coefficient of bimodality used in Shade et al. mBio 5(4):e01371-14.
         and picked from Ellison AM Am. J. Bot 74:1280-8, 1987

         Silverman's bootstrap test for multimodality adopted from
         Silverman BW J. Ro. St. Soc 97-99, 1981.

         Potential analysis from R Early Warning Signals Toolbox
         based on Dakos et al. (2012)PL. ONE 7(7):e41010, 2012.
}
\examples{
x <- rnorm(100)
bimodality_score(x, method="cob")
}
\author{
Contact: Tineka Blake \email{tineka.blake@aalto.fi}
}
\references{
PM Hartigan, App. Stat, 320-325, 1985.
  Shade et al. mBio 5(4):e01371-14, 2014.
  AM Ellison, Am. J. Bot 74:1280-8, 1987.
  BW Silverman, Jo. R. St. Soc 97-99, 1981.
  Dakos et al. PL. ONE 7(7):e41010, 2012.

  To cite the microbiome R package, see citation('microbiome')
}
\keyword{utilities}

