\name{bootstrap.microbecount}
\alias{bootstrap.microbecount}
\title{bootstrap.microbecount}
\usage{
  bootstrap.microbecount(D, Nsample = NULL, minprev = 1,
    Nboot = 100, I.thr = 1.8, ncore = 1)
}
\arguments{
  \item{D}{data}

  \item{Nsample}{sample size}

  \item{minprev}{minimum prevalence}

  \item{Nboot}{bootstrap sample size}

  \item{I.thr}{threshold}

  \item{ncore}{number of nodes for parallelization

  Returns:}
}
\value{
  TBA
}
\description{
  Description: bootstrap.microbecount
}
\details{
  Arguments:
}
\examples{
# tmp <- bootstrap.microbecount(phylum.matrix)
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

