% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.microbes.R
\name{bootstrap.microbes}
\alias{bootstrap.microbes}
\title{Bootstrap microbes}
\usage{
bootstrap.microbes(D, Nsample = NULL, minPrev = 2, Nboot = 1000,
  I.thr = 1.8, ncore = 1, I.max = NULL)
}
\arguments{
\item{D}{data (phylotypes x samples)}

\item{Nsample}{bootstrap sample size, default is the same size as data}

\item{minPrev}{Lower limit for number of samples where microbe needs 
to exceed the intensity threshold for a 'present' call.}

\item{Nboot}{bootstrap sample size}

\item{I.thr}{Lower limit for intensity threshold}

\item{ncore}{number of nodes for parallelization - default 1}

\item{I.max}{Upper limit for intensity threshold. Later addition.
set to NULL (default) to replicate Salonen et al.}
}
\value{
data frame with microbes and their frequency of presence in 
  	     the core
}
\description{
Bootstrap method for core microbiota estimation from Salonen et al. (2012)
}
\examples{
data(peerj32); 
	     bs <- bootstrap.microbes(t(peerj32$microbes), Nboot = 5)
}
\author{
Contact: Jarkko Salojarvi \email{microbiome-admin@googlegroups.com}
}
\references{
The core microbiota bootstrap method implemented with this function:
Salonen A, Salojarvi J, Lahti L, de Vos WM. The adult intestinal
core microbiota is determined by analysis depth and health
status. Clinical Microbiology and Infection 18(S4):16-20, 2012
To cite this R package, see citation("microbiome")
}
\keyword{utilities}

