% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{bootstrap_microbecount}
\alias{bootstrap_microbecount}
\title{Bootstrap Microbe Count Data}
\usage{
bootstrap_microbecount(D, Nsample = NULL, minprev = 1, bs.iter = 1000,
  detection = 1.8, I.max = NULL)
}
\arguments{
\item{D}{data}

\item{Nsample}{sample size}

\item{minprev}{minimum prevalence}

\item{bs.iter}{bootstrap sample size}

\item{detection}{intensity threshold}

\item{I.max}{max intensity threshold}
}
\value{
median microbe count in bootstrapped cores
}
\description{
Auxiliary function for bootstrap_microbes.
}
\examples{
  \dontrun{
    data(peerj32)
    tmp <- bootstrap_microbecount(t(peerj32$microbes), bs.iter = 5)
 }
}
\references{
The core microbiota bootstrap method implemented with this function:
Salonen A, Salojarvi J, Lahti L, de Vos WM. The adult intestinal
core microbiota is determined by analysis depth and health
status. Clinical Microbiology and Infection 18(S4):16-20, 2012
To cite this R package, see citation("microbiome")
}
\author{
Contact: Jarkko Salojarvi \email{microbiome-admin@googlegroups.com}
}
\keyword{internal}
