\name{calculate.rpa}
\alias{calculate.rpa}
\title{Description: RPA for HITChip}
\usage{
  calculate.rpa(level, phylo, oligo.data)
}
\arguments{
  \item{level}{level}

  \item{phylo}{phylo}

  \item{oligo.data}{oligo.data

  Returns:}
}
\value{
  RPA preprocessed data
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

