\name{check.wilcoxon}
\alias{check.wilcoxon}
\title{Description: TBA}
\usage{
  check.wilcoxon(...)
}
\arguments{
  \item{...}{TBA

  Returns:}
}
\value{
  TBA
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

