% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{check_anova}
\alias{check_anova}
\title{check_anova}
\usage{
check_anova(x, group, p.adjust.method = "BH", sort = FALSE)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object or a data matrix 
(features x samples; eg. HITChip taxa vs. samples)}

\item{group}{Vector with specifying the groups}

\item{p.adjust.method}{p-value correction method for p.adjust function 
(default 'BH'). For other options, see ?p.adjust}

\item{sort}{sort the results}
}
\value{
Corrected p-values for multi-group comparison.
}
\description{
Calculate ANOVA test (BH correction) for multi-group comparison
}
\examples{
  data(peerj32)
  pval <- check_anova(t(peerj32$microbes), peerj32$meta$time)

}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

