% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check_wilcoxon}
\alias{check.wilcoxon}
\alias{check_wilcoxon}
\title{Calculate Wilcoxon test (unpaired; BH correction) for the
specified sample groups.}
\usage{
check_wilcoxon(dat, G1, G2, p.adjust.method = "BH", sort = FALSE,
  paired = FALSE)
}
\arguments{
\item{dat}{data matrix (features x samples)}

\item{G1}{Sample group 1 (for comparison)}

\item{G2}{Sample group 2 (for comparison)}

\item{p.adjust.method}{p-value correction method for p.adjust function
            (default 'BH'). If NULL, no correction will be performed.}

\item{sort}{sort the results}

\item{paired}{paired Wilcoxon test}
}
\value{
(Corrected) p-values for two-group comparison.
}
\description{
Calculate Wilcoxon test (unpaired; BH correction) for the
specified sample groups.
}
\examples{
data(peerj32)
 pval <- check_wilcoxon(t(peerj32$microbes), G1 = 1:22, G2 = 23:44)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

