% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon.R
\name{check_wilcoxon}
\alias{check_wilcoxon}
\title{Check Wilcoxon}
\usage{
check_wilcoxon(x, group, p.adjust.method = "BH", sort = FALSE,
  paired = FALSE)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object or a data matrix 
(features x samples; eg. HITChip taxa vs. samples)}

\item{group}{Vector with specifying the groups}

\item{p.adjust.method}{p-value correction method for p.adjust function 
(default 'BH'). For other options, see ?p.adjust}

\item{sort}{sort the results}

\item{paired}{Paired comparison (Default: FALSE)}
}
\value{
Corrected p-values for two-group comparison.
}
\description{
Calculate Wilcoxon test (with BH correction) for two-group comparison for phyloseq objects.
}
\examples{
  #pseq <- download_microbiome("peerj32")$physeq
  #pval <- check_wilcoxon(pseq, "gender")
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

