\name{choose.projects}
\alias{choose.projects}
\title{Description: Select projects to analyze}
\usage{
  choose.projects(con, multi = TRUE, condition = NULL)
}
\arguments{
  \item{con}{valid MySQL connection}

  \item{multi}{enable selection of multiple options}

  \item{condition}{TBA Returns:}
}
\value{
  vector of project names
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

