\name{choose.samples}
\alias{choose.samples}
\title{choose.samples}
\usage{
  choose.samples(con, multi = TRUE,
    title = "Select samples:", condition = NULL)
}
\arguments{
  \item{con}{MySQL connection}

  \item{multi}{multiple selections allowed}

  \item{title}{title}

  \item{condition}{TBA}
}
\value{
  sample vector
}
\description{
  choose.samples
}
\examples{
# TBA
}
\author{
  Douwe Molenaar. Maintainer: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

