% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cmat2table}
\alias{cmat2table}
\title{Description: Arrange correlation matrices from cross.correlate into
        a table format}
\usage{
cmat2table(res, verbose = FALSE)
}
\arguments{
\item{res}{Output from cross.correlate}

\item{verbose}{verbose

Returns:}
}
\value{
Correlation table
}
\description{
Arguments:
}
\examples{
data(peerj32);
         cc <- cross.correlate(peerj32$microbes[1:20, 1:10],
                              peerj32$lipids[1:20,1:10],
                  mode = 'matrix');
                   cmat <- cmat2table(cc)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

