\name{cmat2table}
\alias{cmat2table}
\title{Description: Arrange correlation matrices from cross.correlate into a table format}
\usage{
  cmat2table(res)
}
\arguments{
  \item{res}{Output from cross.correlate

  Returns:}
}
\value{
  Correlation table
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

