% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associate.R
\name{cmat2table}
\alias{cmat2table}
\title{Convert Correlation Matrix into a Table}
\usage{
cmat2table(res, verbose = FALSE)
}
\arguments{
\item{res}{Output from associate}

\item{verbose}{verbose}
}
\value{
Correlation table
}
\description{
Arrange correlation matrices from associate into a table format.
}
\examples{
data(peerj32)
d1 <- peerj32$microbes[1:20, 1:10]
d2 <- peerj32$lipids[1:20,1:10]
cc <- associate(d1, d2, mode='matrix', method='pearson')
cmat <- associate(d1, d2, mode='table', method='spearman')
}
\references{
See citation('microbiome')
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
