% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bimodality.R
\name{coefficient.of.bimodality}
\alias{coefficient.of.bimodality}
\title{Coefficient of bimodality}
\usage{
coefficient.of.bimodality(x, bs.iter = 1)
}
\arguments{
\item{x}{Data vector for which bimodality will be quantified}

\item{bs.iter}{Bootstrap iterations}
}
\value{
Bimodality score
}
\description{
Coefficient of bimodality
}
\details{
Coefficient of bimodality used in Shade et al. mBio 5(4):e01371-14.
         and picked from Ellison AM Am. J. Bot 74:1280-8, 1987
}
\examples{
coefficient.of.bimodality(rnorm(100))
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
Shade et al. mBio 5(4):e01371-14, 2014.
  AM Ellison, Am. J. Bot 74:1280-8, 1987.

  To cite the microbiome R package, see citation('microbiome')
}
\keyword{utilities}

