% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/d3.R
\name{convert4corrwscatter}
\alias{convert4corrwscatter}
\title{Interactive scatter plot}
\usage{
convert4corrwscatter(dat, group, reorder = TRUE)
}
\arguments{
\item{dat}{data matrix}

\item{group}{sample groups (vector or factor)}

\item{reorder}{reorder the data}
}
\value{
JSON
}
\description{
Interactive scatter plot
}
\examples{
print("Check
https://github.com/microbiome/d3/tree/master/corr_w_scatter
for an example of the convert4corrwscatter function")
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
Based on a similar example originally suggested by K. Broman:
          https://www.biostat.wisc.edu/~kbroman/D3/corr_w_scatter
   To cite microbiome R package, see citation("microbiome")
}
\keyword{utilities}

