\name{convert4corrwscatter}
\alias{convert4corrwscatter}
\title{Description:}
\usage{
  convert4corrwscatter(dat, group, reorder = TRUE)
}
\arguments{
  \item{dat}{data matrix}

  \item{group}{sample groups (vector or factor)}

  \item{reorder}{reorder the data

  Returns:}
}
\value{
  JSON
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

