% Generated by roxygen2 (4.0.1): do not edit by hand
\name{convert4corrwscatter}
\alias{convert4corrwscatter}
\title{Description:
Arguments:}
\usage{
convert4corrwscatter(dat, group, reorder = TRUE)
}
\arguments{
\item{dat}{data matrix}

\item{group}{sample groups (vector or factor)}

\item{reorder}{reorder the data

Returns:}
}
\value{
JSON
}
\description{
Description:
Arguments:
}
\examples{
# NOT RUN See https://github.com/microbiome/d3/tree/master/corr_w_scatter"
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

