% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{core}
\alias{core}
\title{Filter Prevalent Taxa}
\usage{
core(x, detection, prevalence)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{detection}{Detection threshold (non-negative real)}

\item{prevalence}{Prevalence threshold (in [0, 100])}
}
\value{
Filtered phyloseq object including only prevalent taxa
}
\description{
Filter the phyloseq object to include only prevalent taxa.
}
\examples{
  data(peerj32)
  core(peerj32$phyloseq, 200, 20)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
To cite the microbiome R package, see citation('microbiome')
}
\keyword{utilities}

