% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.microbes.R
\name{core.which}
\alias{core.which}
\title{Core which}
\usage{
core.which(data, intTr, prevalenceTr)
}
\arguments{
\item{data}{phylotypes vs. samples data matrix}

\item{intTr}{intTr}

\item{prevalenceTr}{prevalenceTr}
}
\value{
Number of OTUs.
}
\description{
Auxiliary function
}
\author{
Contact: Jarkko Salojarvi \email{microbiome-admin@googlegroups.com}
}
\references{
A Salonen et al. The adult intestinal core microbiota is determined by 
  analysis depth and health status. Clinical Microbiology and Infection 
  18(S4):16 20, 2012. 
  To cite the microbiome R package, see citation('microbiome')
}
\keyword{utilities}

