\name{core.which}
\alias{core.which}
\title{Description: core.which}
\usage{
  core.which(data, intTr, prevalenceTr)
}
\arguments{
  \item{data}{data matrix; phylotypes vs. samples}

  \item{intTr}{intTr}

  \item{prevalenceTr}{prevalenceTr

  Returns:}
}
\value{
  TBA
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

