% Generated by roxygen2 (4.0.2): do not edit by hand
\name{core.which}
\alias{core.which}
\title{core.which}
\usage{
core.which(data, intTr, prevalenceTr)
}
\arguments{
\item{data}{data matrix; phylotypes vs. samples}

\item{intTr}{intTr}

\item{prevalenceTr}{prevalenceTr}
}
\value{
Number of OTUs.
}
\description{
core.which
}
\author{
Contact: Jarkko Salojarvi \email{microbiome-admin@googlegroups.com}
}
\references{
A Salonen et al. The adult intestinal core microbiota is determined by
  analysis depth and health status. Clinical Microbiology and Infection
  18(S4):16 20, 2012.

  To cite the microbiome R package, see citation('microbiome')
}
\keyword{utilities}

