% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.which.R
\name{core.which}
\alias{core.which}
\title{Core which}
\usage{
core.which(data, intTr, prevalenceTr)
}
\arguments{
\item{data}{phylotypes vs. samples data matrix}

\item{intTr}{intTr}

\item{prevalenceTr}{prevalenceTr}
}
\value{
Number of OTUs.
}
\description{
Auxiliary function
}
\keyword{internal}

