% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Core2D}
\alias{Core2D}
\title{Core2D}
\usage{
Core2D(coreMat, title = "Common core", plot = TRUE, xlabel = "Abundance",
  ylabel = "Core size (number of taxa)")
}
\arguments{
\item{coreMat}{core matrix}

\item{title}{title}

\item{plot}{plot the figure}

\item{xlabel}{X axis label}

\item{ylabel}{Y axis label


Returns:}
}
\value{
Used for its side effects
}
\description{
Description: Core visualization 2D
}
\details{
Arguments:
}
\examples{
data(peerj32);
	     c2d <- Core2D(createCore(t(peerj32$microbes)))
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

