% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Core2D}
\alias{Core2D}
\title{Core2D}
\usage{
Core2D(coreMat, colnum = NULL, title = "Common core")
}
\arguments{
\item{coreMat}{core matrix}

\item{colnum}{colnum}

\item{title}{title

Returns:}
}
\value{
Used for its side effects
}
\description{
Description: Core visualization 2D
}
\details{
Arguments:
}
\examples{
data(peerj32); c2d <- Core2D(createCore(t(peerj32$microbes)))
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

