\name{Core3D}
\alias{Core3D}
\title{Core3D}
\usage{
  Core3D(coreMat, fname, writedir)
}
\arguments{
  \item{coreMat}{core matrix}

  \item{fname}{output file id}

  \item{writedir}{output directory

  Returns:}
}
\value{
  TBA
}
\description{
  Description: Core visualization 3D
}
\details{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

