% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Core3D}
\alias{Core3D}
\title{Core3D}
\usage{
Core3D(coreMat, title = "Core microbiota", xlab = "Minimum Intensity",
  cex.axis = 0.7)
}
\arguments{
\item{coreMat}{core matrix}

\item{title}{title}

\item{xlab}{X axis label}

\item{cex.axis}{axis text size

Returns:}
}
\value{
Used for its side effects
}
\description{
Description: Core visualization 3D
}
\details{
Arguments:
}
\examples{
data(peerj32);
         c3d <- Core3D(createCore(t(peerj32$microbes)))
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
A Salonen et al. The adult intestinal core microbiota is determined by
  analysis depth and health status. Clinical Microbiology and Infection
  18(S4):16 20, 2012.
  To cite the microbiome R package, see citation('microbiome')
}
\keyword{utilities}

