% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_abundance.R
\name{core_abundance}
\alias{core_abundance}
\title{Core Abundance Index}
\usage{
core_abundance(x, detection, prevalence, split = TRUE)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{detection}{Detection threshold for absence/presence (greater or equal).}

\item{prevalence}{Prevalence threshold (in [0, 1]; greater or equal)}

\item{split}{(Optional). Logical. Should a separate set of richness
estimates be performed for each sample? Or alternatively,
pool all samples and estimate richness of the entire set.}
}
\value{
A vector of core abundance indices
}
\description{
Calculates the community core abundance index.
}
\details{
The core abundance index gives the relative proportion of core species in [0,1]. This is exactly the complement (1 - x) of the rarity function; provided for convenience.
}
\examples{
  data(dietswap)
  d <- core_abundance(dietswap, detection = 0.1/100, prevalence = 50/100)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
