% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core.R
\name{core_heatmap}
\alias{core_heatmap}
\title{core_heatmap}
\usage{
core_heatmap(x, detection.thresholds = NULL, palette = "bw")
}
\arguments{
\item{x}{A \code{\link{phyloseq}} object or a core matrix}

\item{detection.thresholds}{a vector of intensities around the data range}

\item{palette}{palette for the plot.type = 'heatmap'}
}
\value{
Used for its side effects
}
\description{
Core heatmap
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
A Salonen et al. The adult intestinal core microbiota is determined by
  analysis depth and health status. Clinical Microbiology and Infection
  18(S4):16 20, 2012.

  To cite the microbiome R package, see citation('microbiome')
}
\keyword{utilities}

