% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core.R
\name{core_heatmap}
\alias{core_heatmap}
\title{core_heatmap}
\usage{
core_heatmap(data, detection.thresholds = NULL, plot = TRUE,
  palette = "bw")
}
\arguments{
\item{data}{data matrix: phylotypes vs. samples}

\item{detection.thresholds}{Vector of detection thresholds}

\item{plot}{plot the figure}

\item{palette}{'bw' (grayscale) or 'spectral' (colourscale)}
}
\value{
List with the following elements:
        plot: ggplot figure
        data: prevalence data with the varying thresholds
}
\description{
Heatmap of core microbiota
}
\examples{
data(peerj32)
  core <- core_heatmap(t(peerj32$microbes))
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
A Salonen et al. The adult intestinal core microbiota is determined by
  analysis depth and health status. Clinical Microbiology and Infection
  18(S4):16 20, 2012.
  To cite the microbiome R package, see citation('microbiome')
}
\keyword{utilities}

