% Generated by roxygen2 (4.0.2): do not edit by hand
\name{core_heatmap}
\alias{core_heatmap}
\title{core_heatmap}
\usage{
core_heatmap(data, detection.thresholds = NULL, plot = TRUE,
  palette = "bw")
}
\arguments{
\item{data}{data matrix: phylotypes vs. samples}

\item{detection.thresholds}{Vector of detection thresholds}

\item{plot}{plot the figure}

\item{palette}{"bw" (grayscale) or "spectral" (colourscale)

Returns:}
}
\value{
List with the following elements:
  	     plot: ggplot figure
	     data: prevalence data with the varying thresholds
}
\description{
Description: Heatmap of core microbiota
}
\details{
Arguments:
}
\examples{
data(peerj32);
	     core <- core_heatmap(t(peerj32$microbes))
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

