% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_core.R
\name{core_heatmap}
\alias{core_heatmap}
\title{Core Heatmap}
\usage{
core_heatmap(x, dets, cols, min.prev, taxa.order)
}
\arguments{
\item{x}{OTU matrix}

\item{dets}{A vector or a scalar indicating the number of intervals
in (0, log10(max(data))). The dets are calculated for relative
abundancies.}

\item{cols}{colours for the heatmap}

\item{min.prev}{If minimum prevalence is set, then filter out those
rows (taxa) and columns (dets) that never exceed this prevalence.
This helps to zoom in on the actual core region of the heatmap.}

\item{taxa.order}{Ordering of the taxa.}
}
\value{
Used for its side effects
}
\description{
Core heatmap.
}
\references{
A Salonen et al. The adult intestinal core microbiota is determined by 
analysis depth and health status. Clinical Microbiology and Infection 
18(S4):16 20, 2012. 
To cite the microbiome R package, see citation('microbiome')
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
