% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{core_members}
\alias{core_members}
\alias{prevalent_taxa}
\title{Core Taxa}
\usage{
core_members(x, detection = 1, prevalence = 95, method = "standard",
  Nsample = NULL, bs.iter = 1000, I.max = NULL)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{detection}{Detection threshold (non-negative real)}

\item{prevalence}{Prevalence threshold (in [0, 100])}

\item{method}{Either "standard" or "bootstrap". The standard methods selects
the taxa that exceed the given detection and prevalence threshold.
The bootstrap method is more robust an described in Salonen et al. (2012).
Note that the results may depend on the random seed unless a sufficiently
large bootstrap sample size is used.}

\item{Nsample}{Only needed for method "bootstrap". Bootstrap sample size, default is the same size as data.}

\item{bs.iter}{Only needed for method "bootstrap". Bootstrap iterations.}

\item{I.max}{Only needed for method "bootstrap". Upper limit for intensity threshold. Later addition. Set to NULL (default) to replicate Salonen et al.}
}
\value{
Vector of core members
}
\description{
Determine members of the core microbiota with given abundance
and prevalences
}
\details{
For phyloseq object, lists taxa that are more prevalent with the
  given detection. For matrix, lists columns that satisfy
  these criteria.
}
\examples{
  data(dietswap)
  a <- core_members(dietswap, 1, 95)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
A Salonen et al. The adult intestinal core microbiota is determined by 
  analysis depth and health status. Clinical Microbiology and Infection 
  18(S4):16 20, 2012. 
  To cite the microbiome R package, see citation('microbiome')
}
\keyword{utilities}

