\name{count}
\alias{count}
\title{Description: count}
\usage{
  count(d)
}
\arguments{
  \item{d}{TBA

  Returns:}
}
\value{
  list
}
\description{
  For cross-hyb control
}
\details{
  Arguments:
}
\author{
  Contact: Jarkko Salojarvi
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

