% Generated by roxygen2 (4.0.0): do not edit by hand
\name{createCore}
\alias{createCore}
\title{createCore}
\usage{
createCore(data, I.thr = 0, verbose = FALSE, prevalence.intervals = NULL,
  intensity.intervals = 20)
}
\arguments{
\item{data}{data matrix; phylotypes vs. samples}

\item{I.thr}{minimum detection threshold}

\item{verbose}{verbose}

\item{prevalence.intervals}{number of bins on the prevalence grid}

\item{intensity.intervals}{number of bins on the intensity grid

Returns:}
}
\value{
TBA
}
\description{
Description: create coreMatrix
}
\details{
Arguments:
}
\examples{
data(peerj32); core <- createCore(t(peerj32$microbes))
}
\author{
Contact: Jarkko Salojarvi \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

