\name{createCore}
\alias{createCore}
\title{createCore}
\usage{
  createCore(data, I.thr = 0, verbose = FALSE,
    prevalence.intervals = NULL, intensity.intervals = 20)
}
\arguments{
  \item{data}{data matrix; phylotypes vs. samples}

  \item{I.thr}{core threshold}

  \item{verbose}{verbose}

  \item{prevalence.intervals}{number of bins on the
  prevalence grid}

  \item{intensity.intervals}{number of bins on the
  intensity grid

  Returns:}
}
\value{
  TBA
}
\description{
  Description: create coreMatrix
}
\details{
  Arguments:
}
\examples{
# core <- createCore(L2.matrix)
}
\author{
  Contact: Jarkko Salojarvi
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

