% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createCore}
\alias{createCore}
\title{createCore}
\usage{
createCore(data, verbose = FALSE, prevalence.intervals = seq(20, 100, 20),
  intensity.intervals = NULL)
}
\arguments{
\item{data}{data matrix; phylotypes vs. samples}

\item{verbose}{verbose}

\item{prevalence.intervals}{a vector of prevalence percentages in [0,100]}

\item{intensity.intervals}{a vector of intensities around the data range

Returns:}
}
\value{
TBA
}
\description{
Description: create coreMatrix
}
\details{
Arguments:
}
\examples{
data(peerj32);
	     core <- createCore(t(peerj32$microbes))
}
\author{
Contact: Jarkko Salojarvi \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

