% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cross.correlate.R
\name{cross.correlate}
\alias{cross.correlate}
\title{Cross-correlate columns of the input matrices}
\usage{
cross.correlate(x, y = NULL, method = "pearson", p.adj.threshold = Inf,
  cth = NULL, order = FALSE, n.signif = 0, mode = "table",
  p.adj.method = "fdr", verbose = FALSE, filter.self.correlations = FALSE)
}
\arguments{
\item{x}{matrix (samples x features if annotation matrix)}

\item{y}{matrix (samples x features if cross-correlated with annotations)}

\item{method}{association method ('pearson', 'spearman', or 'bicor'
for continuous; categorical for discrete)}

\item{p.adj.threshold}{q-value threshold to include features}

\item{cth}{correlation threshold to include features}

\item{order}{order the results}

\item{n.signif}{mininum number of significant correlations for each
element}

\item{mode}{Specify output format ('table' or 'matrix')}

\item{p.adj.method}{p-value multiple testing correction method.
One of the methods in p.adjust
   function ('BH' and others; see help(p.adjust)).
   Default: 'fdr'}

\item{verbose}{verbose}

\item{filter.self.correlations}{Filter out correlations between
                                  identical items.}
}
\value{
List with cor, pval, pval.adjusted
}
\description{
Cross-correlate columns of the input matrices
}
\details{
As the method=categorical (discrete) association measure
         for nominal (no order for levels) variables we use Goodman and
         Kruskal tau based on
         r-bloggers.com/measuring-associations-between-non-numeric-variables/
}
\examples{
data(peerj32)
  d1 <- peerj32$microbes[1:20, 1:10]
  d2 <- peerj32$lipids[1:20,1:10]
  cc <- cross.correlate(d1, d2)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

