\name{deconvolution.nonneg}
\alias{deconvolution.nonneg}
\title{Description: Deconvolution}
\usage{
  deconvolution.nonneg(oligo.data, oligomap, level,
    block.solution = T, verbose = FALSE)
}
\arguments{
  \item{oligo.data}{oligo.data in absolute domain}

  \item{oligomap}{oligo - phylotype mapping data frame}

  \item{level}{taxonomic level}

  \item{block.solution}{block.solution

  Returns:}
}
\value{
  list
}
\description{
  For cross-hyb control
}
\details{
  Arguments:
}
\author{
  Contact: Jarkko Salojarvi
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

