\name{density.plot}
\alias{density.plot}
\title{density.plot}
\usage{
<<<<<<< HEAD
  density.plot(mat, title = NULL)
=======
  density.plot(mat, title = NULL, x.ticks = 10,
    rounding = 0)
>>>>>>> 13227eee2b3f6b558c33b1a95c2e9482dc66cefd
}
\arguments{
  \item{mat}{Data matrix to plot. The first two columns
  will be visualized as a cross-plot.}

<<<<<<< HEAD
  \item{title}{title text
=======
  \item{title}{title text}

  \item{x.ticks}{Number of ticks on the X axis}

  \item{rounding}{Rounding for X axis tick values
>>>>>>> 13227eee2b3f6b558c33b1a95c2e9482dc66cefd

  Returns:}
}
\value{
  ggplot2 object
}
\description{
  Description: Plots densities of data points in addition
  to cross-plot points.
}
\details{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

