\name{detect.chip}
\alias{detect.chip}
\title{Description: Detect the chip type (H/M/PITChip) from database name}
\usage{
  detect.chip(dbname)
}
\arguments{
  \item{dbname}{MySQL database name

  Returns:}
}
\value{
  chip name
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

