\name{distance.matrix}
\alias{distance.matrix}
\title{Description: Calculate distance matrix between the _columns_ of the
input matrix. Can prduce correlation-based distance matrices, otherwise
uses the standard 'dist' function.}
\usage{
  distance.matrix(x, method = "pearson", ...)
}
\arguments{
  \item{x}{data matrix}

  \item{method}{distance method}

  \item{...}{other arguments to be passed

  Returns:}
}
\value{
  distance object
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

