% Generated by roxygen2 (4.0.0): do not edit by hand
\name{diversity.boxplot}
\alias{diversity.boxplot}
\title{diversity.boxplot}
\usage{
diversity.boxplot(dat, sample.groups, diversity.index = "shannon",
  det.th = NULL, title = NULL, col.list = NULL, ylim = NULL)
}
\arguments{
\item{dat}{data matrix in the original (non-log) scale: features x samples.}

\item{sample.groups}{specify the distinct sample groups}

\item{diversity.index}{Options: "shannon" / "invsimpson" / "richness" / "evenness"}

\item{det.th}{Detection threshold for richness and evenness calculations}

\item{title}{figure title}

\item{col.list}{Optional: colors for the sample groups}

\item{ylim}{y-axis limits

Returns:}
}
\value{
Sample group list corresponding to the boxplot groups.
}
\description{
Description: diversity.boxplot
}
\details{
Arguments:
}
\examples{
data(peerj32); div <- diversity.boxplot(peerj32$microbes, sample.groups = list(1:22, 23:44), det.th = 0)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

