% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{diversity}
\alias{diversity}
\title{Estimate Diversity}
\usage{
diversity(x, detection = 0, split = TRUE, measures = NULL)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{detection}{detection for observing taxa (absence / presence).
Used to determine the richness (Observed diversity) above this
abundance threshold. Zero by default.}

\item{split}{(Optional). Logical. Should a separate set of richness
estimates be performed for each sample? Or alternatively,
pool all samples and estimate richness of the entire set.}

\item{measures}{(Optional). Default is ‘NULL’, meaning that all available
         alpha-diversity measures will be included. Alternatively, you
         can specify one or more measures as a character vector of
         measure names. Values must be among those supported in the
         phyloseq::estimate_richness function. These include
‘c("Observed", "Chao1", "ACE", "Shannon", "Simpson", "InvSimpson", "Fisher")’.
In addition, the measure "Evenness" is provided (Pielou's index).}
}
\value{
A data.frame of samples x diversity indicators;
  except when split=FALSE, a vector of indices is returned.
}
\description{
Diversity estimation.
   Augments the estimate_richness function of the phyloseq package.
}
\examples{
  data(dietswap)
  d <- diversity(dietswap)
}
\references{
See citation('microbiome')
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
