% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{diversity}
\alias{diversity}
\title{Estimate Diversity}
\usage{
diversity(x, detection = 0.1/100, prevalence = 50/100, split = TRUE,
  measures = NULL)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{detection}{detection for observing taxa (absence / presence).
Used to determine the richness (Observed diversity) above this
abundance threshold. Zero by default.}

\item{prevalence}{Prevalence threshold (in [0, 1]; greater or equal)}

\item{split}{(Optional). Logical. Should a separate set of richness
estimates be performed for each sample? Or alternatively,
pool all samples and estimate richness of the entire set.}

\item{measures}{Default is ‘NULL’, meaning that all available
       alpha-diversity measures will be included. Alternatively, you
       can specify one or more measures as a character vector of
       measure names. Values include those supported in the
       phyloseq::estimate_richness function:
       "Observed", "Chao1", "ACE", "Shannon", "Simpson", "InvSimpson", "Fisher".
In addition, the following measures are provided:
  "Evenness" (Pielou's index);
  "Dominance" (Number of species needed to cover 50% of the ecosystem);
  "Top" (Relative proportion of the most dominant species in [0,1]);
  "Rarity" (Relative proportion of the rare (non-core) species in [0,1]);
  "Core" (Relative proportion of the core species in [0,1]);
  "Gini" (Gini index).}
}
\value{
A data.frame of samples x diversity indicators;
  except when split=FALSE, a vector of indices is returned.
}
\description{
Diversity estimation.
   Augments the estimate_richness function of the phyloseq package.
}
\details{
This function returns the indices with the default choices for simplicity. See the individual functions for more options.
}
\examples{
  data(dietswap)
  d <- diversity(dietswap)
}
\references{
See citation('microbiome')
}
\seealso{
rarity, core_abundance, top_abundance, dominance, gini
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
