% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{diversity}
\alias{diversity}
\title{Estimate Global Indices}
\usage{
diversity(x)
}
\arguments{
\item{x}{A species abundance vector, or matrix (taxa/features x samples) with the absolute count data (no relative abundances), or \code{\link{phyloseq-class}} object}
}
\value{
A data.frame of samples x global indicators
}
\description{
Estimate global indicators of the ecoystem state (richness, diversity, and other indicators).
}
\details{
This function returns the indices with the default choices for detection, prevalence and other parameters for simplicity and standardization. See the individual functions for more options. This function extends the functionality of the phyloseq::estimate_richness function.
}
\references{
See citation('microbiome')
}
\seealso{
rarity, core_abundance, dominance, low_abundance, dominance, gini, phyloseq::estimate_richness
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
