% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominance.R
\name{dominance}
\alias{dominance}
\title{Dominance Index}
\usage{
dominance(x, threshold = 0.5, split = TRUE)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{threshold}{Indicates the fraction of the ecosystem that has to be covered by the dominant species.}

\item{split}{(Optional). Logical. Should a separate set of richness
estimates be performed for each sample? Or alternatively,
pool all samples and estimate richness of the entire set.}
}
\value{
A vector of dominance indices
}
\description{
Calculates the community dominance index.
}
\details{
The dominance index gives the number of groups needed to have a given proportion of the ecosystem occupied (by default 0.5 ie 50%).
}
\examples{
  data(dietswap)
  d <- dominance(dietswap, threshold = 0.5)
}
\seealso{
top_abundance, diversity
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
