% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{download_microbiome}
\alias{download_microbiome}
\title{Download microbiome data sets}
\usage{
download_microbiome(id = "datasets")
}
\arguments{
\item{id}{Data set name. For options, see download_microbiome()}
}
\value{
Data set
}
\description{
Download microbiome data sets
}
\examples{
download_microbiome()
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
To cite the microbiome R package, see citation('microbiome')
}
\keyword{utilities}

