\name{esort}
\alias{esort}
\title{Description: Sort data frame dd by columns like: esort(dd, -z, b)}
\usage{
  esort(x, sortvar, ...)
}
\arguments{
  \item{x}{data frame to sort}

  \item{sortvar}{sorted variable/s}

  \item{...}{further parameters to pass

  Returns:}
}
\value{
  sorted data frame
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

