\name{estimateSpeciesPresence}
\alias{estimateSpeciesPresence}
\title{Description: Returns the difference between the mean log10 signals of specific
             and unspecific probes per species}
\usage{
  estimateSpeciesPresence(data, pi)
}
\arguments{
  \item{data}{TBA}

  \item{pi}{TBA

  Returns:}
}
\value{
  TBA
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

