% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential_analysis.R
\name{find_optima}
\alias{find_optima}
\title{Find Optima}
\usage{
find_optima(f, peak.threshold = 0, bw = 1, min.density = 1)
}
\arguments{
\item{f}{density}

\item{peak.threshold}{Mode detection threshold}

\item{bw}{bandwidth}

\item{min.density}{Minimun accepted density for a maximum; 
as a multiple of kernel height}
}
\value{
A list with min (minima), max (maxima), and
       peak.threshold (minimum detection density)
}
\description{
Detect optima, excluding local optima below peak.threshold.
}
\examples{
   # Not exported
   # o <- find_optima(rnorm(100), bw=1)
}
\references{
See citation('microbiome')
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
