% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_phyloseq.R
\name{format_phyloseq}
\alias{format_phyloseq}
\title{Formatting the Phyloseq Object}
\usage{
format_phyloseq(x)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}
}
\value{
\code{\link{phyloseq-class}} object.
}
\description{
Format the phyloseq object to correct the taxonomy in phyloseq object (tax_table).
}
\details{
Most commonly it is observed that the taxonomy file has classification until a given taxonomic level.
         Hence, to avoid loss of OTU information while using the function tax_glom() for merging at a specific taxonomic level.
         we will fill the empty cells with the maximum classification available along with the OTU number. This code is a slight modification.
         the code from  \pkg{ampvis} \code{\link{phyloseq-class}}. Here, we directly take the phyloseq object as input and make the necessary formatting.
}
\examples{
\dontrun{
  # Example data
    library(microbiome)
    data(DynamicsIBD)
    p0 <- DynamicsIBD
    p0.f <- format_phyloseq(p0)
          }
}
\keyword{utilities}
