\name{GenerateSimulatedData}
\alias{GenerateSimulatedData}
\title{Description: Generate toydata for the package}
\usage{
  GenerateSimulatedData(output.dir)
}
\arguments{
  \item{output.dir}{output directory name

  Returns:}
}
\value{
  output file name
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

