\name{get.bkg.params}
\alias{get.bkg.params}
\title{Description: get background parameters}
\usage{
  get.bkg.params()
}
\value{
  TBA
}
\description{
  Arguments:
}
\details{
  Returns:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

