% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.probeset.R
\name{get.probeset}
\alias{get.probeset}
\title{get probeset data matrix}
\usage{
get.probeset(name, level, taxonomy, probedata, log10 = TRUE)
}
\arguments{
\item{name}{name}

\item{level}{taxonomic level}

\item{taxonomy}{taxonomy}

\item{probedata}{oligos vs. samples preprocessed data matrix;
                 absolute scale}

\item{log10}{Logical. Logarithmize the data TRUE/FALSE}
}
\value{
probeset data matrix
}
\description{
get probeset data matrix
}
\examples{
taxonomy <- GetPhylogeny('HITChip', 'filtered')
  data.dir <- system.file("extdata", package = "microbiome")
  probedata <- read_hitchip(data.dir, "rpa")$probedata
  ps <- get.probeset('Akkermansia', 'L2', taxonomy, probedata)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

