% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.probeset.R
\name{get.probeset}
\alias{get.probeset}
\title{get probeset data matrix}
\usage{
get.probeset(name, level, phylogeny.info, probedata, log10 = TRUE)
}
\arguments{
\item{name}{name}

\item{level}{taxonomic level}

\item{phylogeny.info}{phylogeny.info}

\item{probedata}{oligos vs. samples preprocessed data matrix;
                 absolute scale}

\item{log10}{Logical. Logarithmize the data TRUE/FALSE}
}
\value{
probeset data matrix
}
\description{
get probeset data matrix
}
\examples{
phylogeny.info <- GetPhylogeny('HITChip', 'filtered')
  data.dir <- system.file("extdata", package = "microbiome")
  probedata <- read.profiling("frpa", data.dir = data.dir)$oligo
  ps <- get.probeset('Akkermansia', 'L2', phylogeny.info, probedata)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

