\name{GetPhylogeny}
\alias{GetPhylogeny}
\title{GetPhylogeny}
\usage{
  GetPhylogeny(chip)
}
\arguments{
  \item{chip}{chip type (e.g. "HITChip")

  Returns:}
}
\value{
  phylogeny mapping table
}
\description{
  Description: Get Chip phylogeny
}
\details{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

