% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetPhylogeny}
\alias{GetPhylogeny}
\title{GetPhylogeny}
\usage{
GetPhylogeny(chip, phylogeny.version = "full")
}
\arguments{
\item{chip}{chip type (e.g. "HITChip")}

\item{phylogeny.version}{"full" or "filtered" (latter is the basis for species/L1/L2 summarization)

Returns:}
}
\value{
phylogeny mapping table
}
\description{
Description: Get Chip phylogeny
}
\details{
Arguments:
}
\examples{
# NOT RUN: phylogeny.info <- GetPhylogeny("HITChip", "full")
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

