\name{getPvalsFromLmMatrix}
\alias{getPvalsFromLmMatrix}
\title{Description: Get Pvals From Lm Matrix}
\usage{
  getPvalsFromLmMatrix(lmMatrix)
}
\arguments{
  \item{lmMatrix}{TBA

  Returns:}
}
\value{
  anova.pvals
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

