% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{harmonize_fields}
\alias{harmonize_fields}
\title{harmonize_fields}
\usage{
harmonize_fields(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
data.frame with harmonized fields
}
\description{
harmonize fields
}
\examples{
  # m <- harmonize_fields(x)

}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

