% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hit2pseq.R
\name{hitchip2physeq}
\alias{hitchip2physeq}
\title{hitchip2physeq}
\usage{
hitchip2physeq(otu, meta, taxonomy = NULL, detection.limit = 1.8)
}
\arguments{
\item{otu}{Sample x OTU absolute HITChip signal}

\item{meta}{Sample x features metadata data.frame}

\item{taxonomy}{OTU x Taxonomy data.frame (HITChip taxonomy used by default)}

\item{detection.limit}{HITChip signal detection limit (absence / presence)}
}
\value{
phyloseq object
}
\description{
Convert HITChip data into phyloseq format
}
\examples{
library(microbiome)
  data(peerj32)
  otu <- peerj32$microbes
  meta <- peerj32$meta
  physeq <- hitchip2physeq(otu, meta)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
Utilizes the phyloseq package, see citation("phyloseq").
            For this function, see citation('microbiome').
}
\keyword{utilities}

