% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/htree.plot.R
\name{htree.plot}
\alias{htree.plot}
\title{htree.plot}
\usage{
htree.plot(dat, method = "complete", metric = "pearson")
}
\arguments{
\item{dat}{oligoprofile data in original (non-log) domain}

\item{method}{hierarchical clustering method}

\item{metric}{clustering similarity measure}
}
\value{
Used for its side effects; returns the arguments
}
\description{
Plot hierarchical clustering for the input data in absolute
and log10 scale using euclidean and pearson correlation similarities.
Intended for internal use in the run.profiling.script function.
}
\examples{
data(peerj32)
  tmp <- htree.plot(peerj32$microbes[,1:5])
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

