% Generated by roxygen2 (4.0.2): do not edit by hand
\name{impute}
\alias{impute}
\title{Impute missing values from a Gaussian.}
\usage{
impute(X)
}
\arguments{
\item{X}{data matrix (features x samples)}
}
\value{
imputed data matrix
}
\description{
Impute missing values from a Gaussian.
}
\examples{
data(peerj32)
  x <- peerj32$microbes
  xi <- impute(x)
}
\author{
Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

