\name{impute}
\alias{impute}
\title{Impute missing values from a Gaussian.}
\usage{
  impute(X)
}
\arguments{
  \item{X}{data matrix (features x samples)}
}
\value{
  imputed data matrix
}
\description{
  Impute missing values from a Gaussian.
}
\examples{
#X2 <- impute(X)
}
\author{
  Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

