\name{level1TOlevel0}
\alias{level1TOlevel0}
\title{Description: L1-L0 mappings}
\usage{
  level1TOlevel0(l1, phylogeny.info)
}
\arguments{
  \item{l1}{level1 phylotypes}

  \item{phylogeny.info}{phylogeny.info

  Returns:}
}
\value{
  mappings
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

