\name{levelmap}
\alias{levelmap}
\title{levelmap}
\usage{
  levelmap(phylotypes = NULL, level.from, level.to,
    oligomap)
}
\arguments{
  \item{phylotypes}{phylotypes to convert; if NULL then
  considering all phylotypes in the oligomap}

  \item{level.from}{level.from}

  \item{level.to}{level.to}

  \item{oligomap}{oligomap

  Returns:}
}
\value{
  mappings
}
\description{
  Description: map phylotypes between hierarchy levels
}
\details{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

