% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelmap.R
\name{levelmap}
\alias{levelmap}
\title{Map taxonomic levels}
\usage{
levelmap(taxa = NULL, from, to, data)
}
\arguments{
\item{taxa}{taxa to convert; 
if NULL then considering all taxa in the tax.table}

\item{from}{convert from taxonomic level}

\item{to}{convert to taxonomic level}

\item{data}{Either a \code{\link{phyloseq}} object or its code{\link{taxonomyTable-class}} , see the \pkg{phyloseq} package.}
}
\value{
mappings
}
\description{
Map taxa between hierarchy levels.
}
\examples{
  tax.table <- GetPhylogeny('HITChip', 'filtered')
  levelmap('Akkermansia', 'L2', 'L1', tax.table)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

