% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phylo.R
\name{levelmap}
\alias{levelmap}
\title{levelmap}
\usage{
levelmap(phylotypes = NULL, from, to, tax.table)
}
\arguments{
\item{phylotypes}{phylotypes to convert;
if NULL then considering all phylotypes in the tax.table}

\item{from}{convert from taxonomic level}

\item{to}{convert to taxonomic level}

\item{tax.table}{tax.table}
}
\value{
mappings
}
\description{
Map phylotypes between hierarchy levels
}
\examples{
tax.table <- GetPhylogeny('HITChip', 'filtered')
  levelmap(phylotypes = 'Akkermansia', 'L2', 'L1', tax.table)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

