% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phylo.R
\name{levelmap}
\alias{levelmap}
\title{levelmap}
\usage{
levelmap(phylotypes = NULL, level.from, level.to, phylogeny.info)
}
\arguments{
\item{phylotypes}{phylotypes to convert;
if NULL then considering all phylotypes in the phylogeny.info}

\item{level.from}{convert from
Options: 'L0', 'L1', 'L2', 'species', 'oligo'}

\item{level.to}{conver to Options: 'L0', 'L1', 'L2', 'species', 'oligo'}

\item{phylogeny.info}{phylogeny.info}
}
\value{
mappings
}
\description{
Map phylotypes between hierarchy levels
}
\examples{
phylogeny.info <- GetPhylogeny('HITChip', 'filtered')
  levelmap(phylotypes = 'Akkermansia', 'L2', 'L1', phylogeny.info)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

