\name{list.clustering.metrics}
\alias{list.clustering.metrics}
\title{Description: List clustering metrics}
\usage{
  list.clustering.metrics()
}
\value{
  list of clustering metrics
}
\description{
  Arguments:
}
\details{
  Returns:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

