% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{list.color.scales}
\alias{list.color.scales}
\title{List color scales}
\usage{
list.color.scales()
}
\value{
list of color scales
}
\description{
List color scales
}
\examples{
list.color.scales()
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

