\name{list.color.scales}
\alias{list.color.scales}
\title{Description: List color scales}
\usage{
list.color.scales()
}
\value{
list of color scales
}
\description{
Arguments:
}
\details{
Returns:
}
\author{
Contact: Leo Lahti
\email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{internal}

