\name{list.entities}
\alias{list.entities}
\title{Description: list corresponding entities}
\usage{
  list.entities(level, phylo)
}
\arguments{
  \item{level}{level}

  \item{phylo}{phylo

  Returns:}
}
\value{
  entities
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

