\name{list.mysql.projects}
\alias{list.mysql.projects}
\title{list.mysql.projects}
\usage{
  list.mysql.projects(dbuser, dbpwd, dbname)
}
\arguments{
  \item{dbuser}{MySQL user}

  \item{dbpwd}{MySQL password}

  \item{dbname}{MySqL database name Returns:}
}
\value{
  project names vector
}
\description{
  Description: List projects in MySQL database
}
\details{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

