\name{list.mysql.projects}
\alias{list.mysql.projects}
\title{list.mysql.projects}
\usage{
  list.mysql.projects(dbuser, dbpwd, dbname, host = NULL,
    port = NULL)
}
\arguments{
  \item{dbuser}{MySQL user}

  \item{dbpwd}{MySQL password}

  \item{dbname}{MySqL database name}

  \item{host}{host; needed with FTP connections}

  \item{port}{port; needed with FTP connections Returns:}
}
\value{
  project names vector
}
\description{
  Description: List projects in MySQL database
}
\details{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

