\name{list.probes}
\alias{list.probes}
\title{list.probes}
\usage{
  list.probes(name, level, phylo)
}
\arguments{
  \item{name}{name}

  \item{level}{level}

  \item{phylo}{phylo

  Returns:}
}
\value{
  probes
}
\description{
  FIXME: merge with retrieve probesets
}
\details{
  Description: list probes corresponding to a given
  phylogenetic level

  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

