% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prevalent.R
\name{list_prevalent_groups}
\alias{list_prevalent_groups}
\title{List prevalent groups}
\usage{
list_prevalent_groups(x, detection.threshold, prevalence.threshold)
}
\arguments{
\item{x}{Abundance data matrix: samples x features (microbes)}

\item{detection.threshold}{Detection threshold for absence/presence.}

\item{prevalence.threshold}{Detection threshold for prevalence}
}
\value{
Vector of prevalent groups
}
\description{
List prevalent groups
}
\details{
Lists groups that are more prevalent above the detection
         threshold as specified by the detection and prevalence threshold
	    arguments
}
\examples{
data(peerj32)
  list_prevalent_groups(peerj32$microbes, 2, 0.2)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
A Salonen et al. The adult intestinal core microbiota is determined by
  analysis depth and health status. Clinical Microbiology and Infection
  18(S4):16 20, 2012.
  To cite the microbiome R package, see citation('microbiome')
}
\keyword{utilities}

