% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lower.triangle}
\alias{lower.triangle}
\title{Description:
Get lower triangle of a square matrix
as a numeric vector such that
row-by-row, picking elements in the order
2,1;3,1;3,2;4,1,...}
\usage{
lower.triangle(mat)
}
\arguments{
\item{mat}{data matrix
Returns:}
}
\value{
lower triangle as vector
}
\description{
Arguments:
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

