% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{lower.triangle}
\alias{lower.triangle}
\title{Get lower triangle of a square matrix
as a numeric vector such that
row-by-row, picking elements in the order
2,1;3,1;3,2;4,1,...}
\usage{
lower.triangle(mat)
}
\arguments{
\item{mat}{data matrix}
}
\value{
lower triangle as vector
}
\description{
Get lower triangle of a square matrix
as a numeric vector such that
row-by-row, picking elements in the order
2,1;3,1;3,2;4,1,...
}
\examples{
mat <- rbind(c(1,2,3), c(4,5,6), c(7,8,9))
  vec <- lower.triangle(mat)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

