% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.abundancy.table.R
\name{make.abundancy.table}
\alias{make.abundancy.table}
\title{Make abundance table}
\usage{
make.abundancy.table(dat, det.th, discretization.resolution = 1)
}
\arguments{
\item{dat}{data matrix}

\item{det.th}{detection threshold}

\item{discretization.resolution}{discretization resolution}
}
\value{
abundancy table
}
\description{
Calculate abundancies
Discretize Hitchip matrix to form abundancy table
of form j, nj where j is number of counts and nj is number
of phylotypes with the corresponding counts
this format is often required by richness estimation
}
\examples{
  data(peerj32)
  abtab <- make.abundancy.table(10^t(peerj32$microbes), det.th = 0)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
To cite the microbiome R package, see citation('microbiome')
}
\keyword{utilities}

