% Generated by roxygen2 (4.0.2): do not edit by hand
\name{make.abundancy.table}
\alias{make.abundancy.table}
\title{make.abundancy.table}
\usage{
make.abundancy.table(dat, det.th, discretization.resolution = 1)
}
\arguments{
\item{dat}{data matrix}

\item{det.th}{detection threshold}

\item{discretization.resolution}{discretization resolution

Returns:}
}
\value{
abundancy table
}
\description{
Description: Calculate abundancies
Discretize Hitchip matrix to form abundancy table
of form j, nj where j is number of counts and nj is number
of phylotypes with the corresponding counts
this format is often required by richness estimation
}
\details{
Arguments:
}
\examples{
data(peerj32);
	     abtab <- make.abundancy.table(10^t(peerj32$microbes),
	     	      			det.th = 0)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

