% Generated by roxygen2 (4.0.2): do not edit by hand
\name{matrix.qvalue}
\alias{matrix.qvalue}
\title{matrix.qvalue}
\usage{
matrix.qvalue(pvals)
}
\arguments{
\item{pvals}{p-value matrix}
}
\value{
q-value matrix
}
\description{
Calculate qvalues for a matrix of pvalues which may contain missing values.
}
\examples{
qvals <- matrix.qvalue(matrix(runif(1000), nrow = 100))
}
\author{
Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

