% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{microbiome-package}
\alias{microbiome}
\alias{microbiome-package}
\title{R package for microbiome studies}
\description{
Brief summary of the microbiome package
}
\details{
\tabular{ll}{
Package: \tab microbiome\cr
Type: \tab Package\cr
Version: \tab See sessionInfo() or DESCRIPTION file\cr
Date: \tab 2014\cr
License: \tab FreeBSD\cr
LazyLoad: \tab yes\cr
}

R package for microbiome studies
}
\examples{
citation("microbiome")
}
\author{
Leo Lahti and Jarkko Salojarvi (University of Helsinki, Finland and Wageningen University, The Netherlands). \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
\url{http://microbiome.github.io}
}
\keyword{package}

