\docType{package}
\name{microbiome-package}
\alias{microbiome}
\alias{microbiome-package}
\title{R package for microbiome studies}
\description{
  Brief summary of the microbiome package
}
\details{
  \tabular{ll}{ Package: \tab microbiome\cr Type: \tab
  Package\cr Version: \tab See sessionInfo() or DESCRIPTION
  file\cr Date: \tab 2013\cr License: \tab FreeBSD\cr
  LazyLoad: \tab yes\cr }

  R package for microbiome studies
}
\examples{
library(microbiome)
}
\author{
  Leo Lahti and Jarkko Salojarvi (UH/WUR). Contact:
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
  \url{http://microbiome.github.com}
}
\keyword{package}

