% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome-package.R
\docType{package}
\name{microbiome-package}
\alias{microbiome-package}
\alias{microbiome}
\title{R package for microbiome studies}
\description{
Brief summary of the microbiome package
}
\details{
\tabular{ll}{
Package: \tab microbiome\cr
Type: \tab Package\cr
Version: \tab See sessionInfo() or DESCRIPTION file\cr
Date: \tab 2014-2017\cr
License: \tab FreeBSD\cr
LazyLoad: \tab yes\cr
}

R package for microbiome studies
}
\examples{
citation('microbiome')
}
\references{
See citation('microbiome')
\url{http://microbiome.github.io}
}
\author{
Leo Lahti et al.
  \email{microbiome-admin@googlegroups.com}
}
\keyword{package}
