\name{mixingMatrix}
\alias{mixingMatrix}
\title{Description: mixingMatrix}
\usage{
  mixingMatrix(oligo.map, level)
}
\arguments{
  \item{oligo.map}{oligo.map}

  \item{level}{taxonomic level

  Returns:}
}
\value{
  oligos x phylotypes mixing matrix
}
\description{
  For cross-hyb control
}
\details{
  Arguments:
}
\author{
  Contact: Jarkko Salojarvi
  \email{jarkko.salojarvi@helsinki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

