% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multimodality.R
\name{multimodality}
\alias{multimodality}
\title{multimodality}
\usage{
multimodality(x, method = "potential.bootstrap", detection.threshold = 1,
  bw.adjust = 1, bs.iterations = 100, detection.limit = 1,
  verbose = TRUE)
}
\arguments{
\item{x}{A vector, matrix, or a phyloseq object}

\item{method}{multimodality quantification method ('potential.bootstrap'
or 'coefficient_of_bimodality')}

\item{detection.threshold}{Mode detection threshold}

\item{bw.adjust}{Bandwidth adjustment}

\item{bs.iterations}{Bootstrap iterations}

\item{detection.limit}{minimum accepted density for a maximum;
as a multiple of kernel height}

\item{verbose}{Verbose}
}
\value{
A list with following elements:
  \itemize{
    \item{score}{Fraction of bootstrap samples where multiple modes are observed}
    \item{nmodes}{The most frequently observed number of modes in bootrstrap sampling results}
    \item{results}{Full results of potential_analysis_bootstrap for each row of the input matrix.}
  }
}
\description{
Wrapper to calculate multimodality score based on various methods
}
\details{
\itemize{
	    \item{coefficient_of_bimodality}{Coefficient of bimodality, used and described in Shade et al. (2014) and Ellison AM (1987).}
 	    \item{potential.bootstrap}{Repeats potential analysis (Livina et al. 2010) multiple times with bootstrap sampling for each row of the input data (as in Lahti et al. 2014) and returns the bootstrap score.}
      }
}
\examples{
data(peerj32)
  multimodality(peerj32$microbes[, "Akkermansia"])
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
Livina et al. (2010). Potential analysis
reveals changing number of climate states during the last 60
kyr. \emph{Climate of the Past}, 6, 77-82.

Lahti et al. (2014). Tipping elements of the human intestinal
ecosystem. \emph{Nature Communications} 5:4344.

Shade et al. (2014). Conditionally Rare Taxa Disproportionately Contribute to Temporal Changes in Microbial Diversity. mBio 5(4):e01371-14. doi: 10.1128/mBio.01371-14

Ellison AM (1987). Effect of seed dimorphism on the density-dependent dynamics of experimental populations of Atriplex triangularis (Chenopodiaceae). Am. J. Bot. 74:1280–1288. doi:10.2307/2444163.
}
\keyword{utilities}

