\name{n.phylotypes.per.oligo}
\alias{n.phylotypes.per.oligo}
\title{Description: Check number of matching phylotypes for each probe}
\usage{
  n.phylotypes.per.oligo(phylogeny.info, level)
}
\arguments{
  \item{phylogeny.info}{oligo - phylotype matching
  data.frame}

  \item{level}{phylotype level

  Returns:}
}
\value{
  number of matching phylotypes for each probe
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

