% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.phylotypes.per.oligo.R
\name{n.phylotypes.per.oligo}
\alias{n.phylotypes.per.oligo}
\title{n.phylotypes.per.oligo}
\usage{
n.phylotypes.per.oligo(taxonomy, level)
}
\arguments{
\item{taxonomy}{oligo - phylotype matching data.frame}

\item{level}{phylotype level}
}
\value{
number of matching phylotypes for each probe
}
\description{
Check number of matching phylotypes for each probe
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

