% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neatsort.R
\name{neatsort}
\alias{neatsort}
\title{Neatmap ordering for matrix rows or columns}
\usage{
neatsort(x, arrange, method = "NMDS", distance = "bray", first = NULL,
  ...)
}
\arguments{
\item{x}{A matrix.}

\item{arrange}{Order "rows" or "cols"}

\item{method}{Ordination method. Only NMDS implemented for now.}

\item{distance}{Distance method. See \code{\link{vegdist}} function from the \pkg{vegan} package.}

\item{first}{Optionally provide the name of the first sample/taxon to start the ordering (the ordering is cyclic so we can start at any point). The choice of the first sample may somewhat affect the overall ordering.}

\item{...}{Arguments to pass.}
}
\value{
Vector of ordered elements
}
\description{
Order rows or columns based on the neatmap approach.
}
\details{
This function borrows elements from the heatmap implementation in the \pkg{phyloseq} package. The row/column sorting is there
not available as a separate function at present, however, hindering reuse in other tools. This function provides an independent
method for easy row/column reordering for matrices. This a quick hack and using the ordination could be expanded further
(now only NMDS is available, and the sorting is done independently for rows and columns).
}
\examples{
\dontrun{
  data(peerj32)
  x <- peerj32$microbes
  x <- neatsort(x, "rows", method = "NMDS", distance = "bray")
                  }
}
\references{
This function is partially based on code derived from the \pkg{phyloseq} package. However for the original
  neatmap approach for heatmap sorting, see (and cite):
  Rajaram, S., & Oono, Y. (2010). NeatMap--non-clustering heat map alternatives in R. BMC Bioinformatics, 11, 45.
}
\keyword{utilities}

