% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{oligoTOhigher}
\alias{oligoTOhigher}
\title{oligo-species mappings}
\usage{
oligoTOhigher(oligo, phylogeny.info, level.to = "species")
}
\arguments{
\item{oligo}{oligos}

\item{phylogeny.info}{phylogeny.info}
}
\value{
mappings
}
\description{
oligo-species mappings
}
\author{
Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation('microbiome')
}
\keyword{internal}

