% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bimodal.R
\name{plot_bimodal}
\alias{plot_bimodal}
\title{Univariate bimodality plot}
\usage{
plot_bimodal(x, taxon, tipping.point = NULL, lims = NULL, shift = 0.001)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{taxon}{Taxonomic group to visualize.}

\item{tipping.point}{Optional. Indicate critical point for abundance variations to be highlighted.}

\item{lims}{Optional. Figure X axis limits.}

\item{shift}{Small constant to avoid problems with zeroes in log10}
}
\value{
\code{\link{ggplot}} object
}
\description{
Coloured bimodality plot.
}
\examples{
  data("atlas1006")
  pseq <- atlas1006
  pseq <- subset_samples(pseq, DNA_extraction_method == "r")
  pseq <- transform_phyloseq(pseq, "relative.abundance")
  p <- plot_bimodal(pseq, "Dialister", tipping.point = 1)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

