% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core.R
\name{plot_core}
\alias{plot_core}
\title{plot_core}
\usage{
plot_core(x, title = "Core", plot = TRUE, prevalence.intervals = NULL,
  detection.thresholds = NULL, plot.type = "lineplot", palette = "bw")
}
\arguments{
\item{x}{A \code{\link{phyloseq}} object or a core matrix}

\item{title}{title}

\item{plot}{plot the figure}

\item{prevalence.intervals}{a vector of prevalence percentages in [0,100]}

\item{detection.thresholds}{a vector of intensities around the data range}

\item{plot.type}{Plot type ('lineplot' or 'heatmap')}

\item{palette}{palette for the plot.type = 'heatmap'}
}
\value{
Used for its side effects
}
\description{
Core visualization 2D
}
\examples{
pseq <- download_microbiome("atlas1006")
p <- plot_core(pseq, prevalence.intervals = seq(10, 100, 10), detection.thresholds = c(0, 10^(0:4)))
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
A Salonen et al. The adult intestinal core microbiota is determined by
  analysis depth and health status. Clinical Microbiology and Infection
  18(S4):16 20, 2012.

  To cite the microbiome R package, see citation('microbiome')
}
\keyword{utilities}

