% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_cumulative}
\alias{plot_cumulative}
\title{plot_cumulative}
\usage{
plot_cumulative(d.sub, i.set = NULL, type = "cumulative", ylim = NULL,
  phylogeny.info)
}
\arguments{
\item{d.sub}{d.sub}

\item{i.set}{i.set}

\item{type}{plot type}

\item{ylim}{y axis limits}

\item{phylogeny.info}{phylogeny.info matrix}
}
\value{
Used for side-effects (plot)
}
\description{
Plot cumulative core microbiota.
}
\examples{
\dontrun{
  data(peerj32)
  bs <- bootstrap.microbes(t(peerj32$microbes), Nboot = 5);
  phylogeny.info <- GetPhylogeny("HITChip")
  plot_cumulative(bs, phylogeny.info = phylogeny.info)
}
}
\author{
Contact: Jarkko Salojarvi \email{microbiome-admin@googlegroups.com}
}
\references{
The core microbiota bootstrap method implemented with this function:
Salonen A, Salojarvi J, Lahti L, de Vos WM. The adult intestinal
core microbiota is determined by analysis depth and health
status. Clinical Microbiology and Infection 18(S4):16-20, 2012

To cite this R package, see citation("microbiome")
}
\keyword{utilities}

