% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_cumulative}
\alias{plot_cumulative}
\title{plot_cumulative}
\usage{
plot_cumulative(d.sub, writedir, fname, i.set = NULL, type = "cumulative",
  ylim = NULL, phylogeny.info)
}
\arguments{
\item{d.sub}{d.sub}

\item{writedir}{output directory}

\item{fname}{output file name}

\item{i.set}{i.set}

\item{type}{plot type}

\item{ylim}{y axis limits}

\item{phylogeny.info}{phylogeny.info matrix

Returns:}
}
\value{
TBA
}
\description{
Description: plot_cumulative
}
\details{
Arguments:
}
\examples{
\dontrun{bs <- bootstrap.microbes(t(peerj32$microbes), Nboot = 5)
	     	      plot_cumulative(bs, writedir = "./", "tmp-",
		      			phylogeny.info = phylogeny.info)}
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
A Salonen et al. The adult intestinal core microbiota is determined by analysis depth and health status. Clinical Microbiology and Infection 18(S4):16 20, 2012.
  To cite the microbiome R package, see citation("microbiome")
}
\keyword{utilities}

