% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_frequencies.R
\name{plot_frequencies}
\alias{plot_frequencies}
\title{Plot Frequencies}
\usage{
plot_frequencies(x, Groups, Factor)
}
\arguments{
\item{x}{\code{\link{data.frame}}}

\item{Groups}{Name of the grouping variable}

\item{Factor}{Name of the frequency variable}
}
\value{
\code{\link{ggplot}} plot object.
}
\description{
Plot relative frequencies within each Group for the
levels of the given factor.
}
\details{
For table with the indicated frequencies, see the returned
phyloseq object.
}
\examples{
data(dietswap)
p <- plot_frequencies(meta(dietswap), 'group', 'sex')
}
\keyword{utilities}
