% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_landscape.R
\name{plot_landscape}
\alias{plot_landscape}
\title{Landscape Plot}
\usage{
plot_landscape(x, method = "NMDS", distance = "bray", col = NULL, ...)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object or a data matrix 
(features x samples; eg. HITChip taxa vs. samples)}

\item{method}{Ordination method, see phyloseq::plot_ordination}

\item{distance}{Ordination distance, see phyloseq::plot_ordination}

\item{col}{Variable name to highlight samples (points) with colors}

\item{...}{Further arguments to pass}
}
\value{
A \code{\link{ggplot}} plot object.
}
\description{
Plot abundance landscape ie. sample density in 2D projection landscape
}
\details{
For consistent results, set random seet (set.seed) before function call
}
\examples{
\dontrun{
  data(dietswap)
  p <- plot_landscape(dietswap)
}
}
\keyword{utilities}

