% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_potential.R
\name{plot_potential}
\alias{plot_potential}
\title{Plot Potential}
\usage{
plot_potential(res, cutoff = 0.5, plot.contours = TRUE, binwidth = 0.2,
  bins = NULL)
}
\arguments{
\item{res}{output from potential_slidingaverage function}

\item{cutoff}{parameter determining the upper limit of potential for
visualizations}

\item{plot.contours}{Plot contour lines.}

\item{binwidth}{binwidth for contour plot}

\item{bins}{bins for contour plot. Overrides binwidth if given}
}
\value{
A ggplot2 visualization of the potential landscape.
}
\description{
Visualization of the potential function.
}
\details{
Applied on the output of the
\code{\link{potential_slidingaverage}} function.
}
\examples{
# Load gut microbiota data on 1006 western adults
# From http://doi.org/10.5061/dryad.pk75d
# (see help(atlas1006) for references and details)
data(atlas1006)
pseq <- subset_samples(atlas1006, DNA_extraction_method == 'r')

# Pick diversity and age
diversity <- exp(global(pseq, index = 'shannon'))
age <- meta(pseq)$age
  
# Run potential analysis
# library(earlywarnings)
# library(ggplot2)
# suppressWarnings(res <- movpotential_ews(diversity, age)$res)
# suppressWarnings(plot_potential(res, cutoff = 0.5))
}
\references{
Lahti et al. Tipping elements of the human intestinal ecosystem.
\emph{Nature Communications} 5:4344, 2014.
}
\seealso{
potential_analysis
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utils}
