% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_potential.R
\name{plot_potential}
\alias{plot_potential}
\title{Plot Potential}
\usage{
plot_potential(res, cutoff = 0.5, plot.contours = TRUE, binwidth = 0.2,
  bins = NULL)
}
\arguments{
\item{res}{output from potential_slidingaverage function}

\item{cutoff}{parameter determining the upper limit of potential for
visualizations}

\item{plot.contours}{Plot contour lines.}

\item{binwidth}{binwidth for contour plot}

\item{bins}{bins for contour plot. Overrides binwidth if given}
}
\value{
A ggplot2 visualization of the potential landscape.
}
\description{
Visualization of the potential function.
}
\details{
Applied on the output of the
         \code{\link{potential_slidingaverage}} function.
}
\examples{
  X <- c(rnorm(1000, mean = 0), rnorm(1000, mean = -2), 
  	    rnorm(1000, mean = 2))
  param <- seq(0,5,length=3000); 
  res <- potential_slidingaverage(X, param); 
  plot_potential(res$res, title = '', 
       	    xlab.text = '', ylab.text = '', 
	    cutoff = 0.5, plot.contours = TRUE,
	    binwidth = 0.2)
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
Lahti et al. Tipping elements of the human intestinal ecosystem.
            \emph{Nature Communications} 5:4344, 2014.
}
\keyword{utils}

