% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rda_bagged.R
\name{plot_rda_bagged}
\alias{plot_rda_bagged}
\title{Plot RDA}
\usage{
plot_rda_bagged(x, which.bac = 1:nrow(x$loadings), ptype = "spider",
  comp = 1:2, cex.bac = 0.5, plot.names = TRUE,
  group.cols = as.numeric(unique(Y)), ...)
}
\arguments{
\item{x}{Output from rda_bagged}

\item{which.bac}{TBA}

\item{ptype}{Plot type. 'spider' or 'hull'}

\item{comp}{TBA}

\item{cex.bac}{Plot size.}

\item{plot.names}{Plot names}

\item{group.cols}{Group colors.}

\item{...}{Other arguments to be passed}
}
\value{
TBA
}
\description{
rda_bagged output visualization.
}
\examples{
  library(microbiome)
  data(peerj32)
  x <- t(peerj32$microbes)
  y <- factor(peerj32$meta$time); names(y) <- rownames(peerj32$meta)
  # Use more iterations in practiec
  res <- rda_bagged(x, y, detection=0.05, bs.iter=5)
  plot_rda_bagged(res)
}
\references{
See citation('microbiome')
}
\author{
Contact: Jarkko Salojarvi \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
