% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa_prevalence.R
\name{plot_taxa_prevalence}
\alias{plot_taxa_prevalence}
\title{Plotting the Taxa Prevalence}
\usage{
plot_taxa_prevalence(x, level)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{level}{Phylum/Order/Class/Family}
}
\value{
A \code{\link{ggplot}} plot object.
}
\description{
A simplified wrapper for creating taxa prevalence plots at Phylum/Order/Class/Family level.
}
\details{
This plot is useful to have a first insights into how is the taxa distribution in your data.
         This also gives an idea about the taxonomic affiliation of rare and abundant taxa in your dataset. This may be helpful in some data filtering or other downstream analysis.
}
\examples{
\dontrun{
 library(microbiome)
 data(DynamicsIBD)
 p0 <- DynamicsIBD
 p0.f <- format_phyloseq(p0)
 plot <- plot_taxa_prevalence(p0.f, 'Phylum')
    }
}
\author{
Sudarshan A. Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{utilities}
