\name{plotMeanProfile}
\alias{plotMeanProfile}
\title{Description: Plot Mean Profile}
\usage{
  plotMeanProfile(data)
}
\arguments{
  \item{data}{TBA

  Returns:}
}
\value{
  barplot
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

