% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotPhylochipHeatmap}
\alias{PlotPhylochipHeatmap}
\title{PlotPhylochipHeatmap}
\usage{
PlotPhylochipHeatmap(data, phylogeny.info, metric = "pearson", level = "L1",
  tree.display = TRUE, palette = "white/black", fontsize = 40,
  figureratio = 10, hclust.method = "complete")
}
\arguments{
\item{data}{oligoprofile data in original (non-log) domain}

\item{phylogeny.info}{oligo-phylotype mappings}

\item{metric}{clustering metric}

\item{level}{taxonomic level to show (L0 / L1 / L2 / species)}

\item{tree.display}{tree.display}

\item{palette}{color palette ("white/black" / "white/blue" / "black/yellow/white")}

\item{fontsize}{font size}

\item{figureratio}{figure ratio}

\item{hclust.method}{hierarchical clustering method. See help(hclust) for details. To prevent ordering of the rows, use hclust.method = NULL.

Returns:}
}
\value{
parameters
}
\description{
Description: Plots heatmap of the oligo profiles together with phylotype grouping and sample clustering
}
\details{
Arguments:
}
\examples{
# TODO
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

