% Generated by roxygen2 (4.0.2): do not edit by hand
\name{polish.phylogeny.info}
\alias{polish.phylogeny.info}
\title{polish.phylogeny.info}
\usage{
polish.phylogeny.info(phylogeny.info)
}
\arguments{
\item{phylogeny.info}{phylogeny.info data frame}
}
\value{
polished phylogeny.info
}
\description{
Ensure phylogeny.info is in correct format
}
\examples{
phylogeny.info <- GetPhylogeny('HITChip', 'filtered')
  phylogeny.info <- polish.phylogeny.info(phylogeny.info)
}
\author{
Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

