\name{preproc.Hitchip}
\alias{preproc.Hitchip}
\title{Description: HITChip preprocessing function}
\usage{
  preproc.Hitchip(x)
}
\arguments{
  \item{x}{TBA

  Returns:}
}
\value{
  x
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

