% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{prevalence}
\alias{prevalence}
\title{Prevalence for phyloseq OTUs}
\usage{
prevalence(x, detection.threshold, sort = FALSE, relative = TRUE)
}
\arguments{
\item{x}{A vector, data matrix or phyloseq object}

\item{detection.threshold}{Detection threshold for absence/presence.}

\item{sort}{Sort the groups by prevalence}

\item{relative}{Logical. Indicate prevalence as fraction of samples (in percentage [0, 100]; default); or in absolute counts indicating the number of samples where the OTU is detected above the given abundance threshold.}
}
\value{
For each OTU, the fraction of samples where a given OTU is detected. The output is readily given as a percentage.
}
\description{
Simple prevalence measure.
}
\details{
For vectors, calculates the fraction (mode relative) or number (mode absolute) of samples that exceed the
detection threshold. For matrices, calculates this for each matrix column. For phyloseq objects, calculates this for each OTU. The relative prevalence (relative = TRUE) is simply the absolute prevalence (relative = FALSE) divided by the number of samples.
}
\examples{
  data(peerj32)     
  ## With matrix
  prevalence(peerj32$data$microbes, detection.threshold = 200, sort = TRUE)
  ## With phyloseq
  prevalence(peerj32$phyloseq, detection.threshold = 200, sort = TRUE)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
A Salonen et al. The adult intestinal core microbiota is determined by 
  analysis depth and health status. Clinical Microbiology and Infection 
  18(S4):16 20, 2012. 
  To cite the microbiome R package, see citation('microbiome')
}
\keyword{utilities}

