% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prevalence}
\alias{prevalence}
\title{Description: Simple prevalence measure}
\usage{
prevalence(x, detection.threshold, sort = F)
}
\arguments{
\item{x}{Abundance data matrix: samples x features (microbes)}

\item{detection.threshold}{Detection threshold for absence/presence.}
}
\value{
Number of OTUs.
}
\description{
Arguments:
}
\details{
Calculates for each sample the fraction of samples that
         exceed the detection threshold


Returns:
}
\examples{
data(peerj32)
prevalence(peerj32$microbes, detection.threshold, sort = T)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
A Salonen et al. The adult intestinal core microbiota is determined by
  analysis depth and health status. Clinical Microbiology and Infection
  18(S4):16 20, 2012.
  To cite the microbiome R package, see citation('microbiome')
}
\keyword{utilities}

