% Generated by roxygen2 (4.0.2): do not edit by hand
\name{project.data}
\alias{project.data}
\title{Project high-dimensional data on two-dimensional plane by various methods}
\usage{
project.data(amat, type = "PCA")
}
\arguments{
\item{amat}{data matrix (samples x features)}

\item{type}{projection type
          (options: PCA, MDS.classical, MDS.nonmetric, Sammon)}
}
\value{
projected data matrix
}
\description{
Project high-dimensional data on two-dimensional plane by various methods
}
\examples{
data(peerj32)
  xy <- project.data(peerj32$microbes[,1:3])
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
D. Hand and H. Mannila and P. Smyth:
   Principles of Data Mining. MIT Press. Cambridge, MA, US (2001).

   To cite microbiome R package, see citation('microbiome')
}
\keyword{utilities}

