% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarity.R
\name{rarity}
\alias{rarity}
\title{Rarity Index}
\usage{
rarity(x, detection = 0.1/100, prevalence = 50/100, split = TRUE)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{detection}{Detection threshold for absence/presence (greater or equal).}

\item{prevalence}{Prevalence threshold (in [0, 1]; greater or equal)}

\item{split}{(Optional). Logical. Pool all samples and estimate index for the entire set.}
}
\value{
A vector of rarity indices
}
\description{
Calculates the community rarity index.
}
\details{
The rarity index gives the relative proportion of rare species (ie. those that are not part of the core microbiota) in the interval [0,1]. This is the complement (1-x) of the core abundance.
}
\examples{
  data(dietswap)
  d <- rarity(dietswap, detection = 0.1/100, prevalence = 50/100)
}
\seealso{
core_abundance, rarity, diversity
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
