% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarity.R
\name{rarity}
\alias{rarity}
\title{Rarity Index}
\usage{
rarity(x, q = 0.5, n = 50)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{q}{Arithmetic abundance classes are evenly cut up to to this quantile of the data. The assumption is that abundances higher than this are not common, and they are classified in their own group.}

\item{n}{The number of arithmetic abundance classes from zero to the quantile cutoff indicated by q.}
}
\value{
A vector of rarity indices
}
\description{
Calculates the community rarity index.
}
\details{
The rarity index characterizes the concentration of species at low abundance. Here, we use the skewness of the frequency distribution of arithmetic abundance classes (see Magurran & McGill 2011). These are typically right-skewed; to avoid taking log of occasional negative skews, we follow Locey & Lennon (2016) and use the log-modulo transformation that adds a value of one to each measure of skewness to allow logarithmization.
}
\examples{
  data(dietswap)
  d <- rarity(dietswap)
}
\references{
Kenneth J. Locey and Jay T. Lennon. Scaling laws predict global microbial diversity. PNAS 2016 113 (21) 5970-5975; doi:10.1073/pnas.1521291113.

  Magurran AE, McGill BJ, eds (2011) Biological Diversity: Frontiers in Measurement and Assessment (Oxford Univ Press, Oxford), Vol 12
}
\seealso{
core_abundance, low_abundance, global
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
