% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarity.R
\name{rarity}
\alias{rarity}
\title{Rarity Index}
\usage{
rarity(x, detection, prevalence, split = TRUE)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{detection}{Detection threshold for absence/presence (greater or equal).}

\item{prevalence}{Prevalence threshold (in [0, 1]; greater or equal)}

\item{split}{(Optional). Logical. Should a separate set of richness
estimates be performed for each sample? Or alternatively,
pool all samples and estimate richness of the entire set.}
}
\value{
A vector of rarity indices
}
\description{
Calculates the community rarity index.
}
\details{
The rarity index gives the relative proportion of rare species in [0,1].
}
\examples{
  data(dietswap)
  d <- rarity(dietswap, detection = 0.1/100, prevalence = 50/100)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
