% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda_bagged.R
\name{rda_bagged}
\alias{rda_bagged}
\title{Bagged RDA}
\usage{
rda_bagged(x, y, bs.iter = 100, verbose = TRUE)
}
\arguments{
\item{x}{a matrix, samples on columns, variables (bacteria) on rows. 
Or a \code{\link{phyloseq-class}} object}

\item{y}{vector or factor with names(y)=rownames(X). 
Or name of phyloseq sample data variable name
(one of sample_variables(x)).}

\item{bs.iter}{Number of bootstrap iterations}

\item{verbose}{verbose}
}
\value{
List with items:
   \itemize{
       \item{loadings}{bagged loadings}
       \item{significance}{significances of X variables}
       \item{scores}{bagged scores}
       \item{group.centers}{group centers on latent space}
       \item{bootstrapped}{bootstrapped loadings}
       \item{data}{data set with non-significant components dropped out}
   }
}
\description{
Bagged (or Bootstrap Aggregated) RDA feature selection
}
\details{
Bootstrap aggregation (bagging) is expected to improve the
   stability of the results. Aggregating results over several modeling runs
   with different boostrap samples of the data are averaged to produce the
   final summary.
}
\examples{
  # RDA with phyloseq object
  data(peerj32)
  res <- rda_bagged(peerj32$phyloseq, 'gender', bs.iter=2)
}
\references{
See citation('microbiome')
}
\seealso{
plot_rda_bagged, vegan::rda, phyloseq::ordinate
}
\author{
Jarkko Salojarvi \email{microbiome-admin@googlegroups.com}
}
\keyword{utilities}
