% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda_physeq.R
\name{rda_physeq}
\alias{rda_physeq}
\title{RDA for phyloseq objects}
\usage{
rda_physeq(x, varname, scale = FALSE, na.action = na.fail, ...)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{varname}{Variable to apply in RDA visualization.}

\item{scale}{See help(rda)}

\item{na.action}{See help(rda)}

\item{...}{Other arguments to be passed}
}
\value{
rda result. See help(vegan::rda)
}
\description{
RDA for phyloseq objects based on the \code{\link{rda}} function from the \pkg{vegan} package.
}
\examples{
#
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation("microbiome")
}
\keyword{utilities}

