\name{read.hitchip.prof008v3.4}
\alias{read.hitchip.prof008v3.4}
\title{Description: read hitchip prof 008 v3.4}
\usage{
  read.hitchip.prof008v3.4(path = "../data/", dataID = "")
}
\arguments{
  \item{path}{TBA}

  \item{dataID}{TBA

  Returns:}
}
\value{
  d
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

