\name{read.hitchip2}
\alias{read.hitchip2}
\title{Description: Read HITChip file}
\usage{
  read.hitchip2(hitchip.file)
}
\arguments{
  \item{hitchip.file}{hitchip file name

  Returns:}
}
\value{
  matrix
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

