\name{read.in.data}
\alias{read.in.data}
\title{Description: Read data}
\usage{
  read.in.data(...)
}
\arguments{
  \item{...}{No input by default

  Returns:}
}
\value{
  Used for side effects
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

