% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\name{read.profiling}
\alias{read.profiling}
\title{read.profiling}
\usage{
read.profiling(data.dir)
}
\arguments{
\item{data.dir}{Profiling script output directory for reading the data.
                If not given, GUI will ask to specify the file and
                   overruns the possible level / method arguments in the
                   function call.}
}
\value{
data matrix (phylo x samples)
}
\description{
Read run.profiling.script output into R
}
\examples{
# data.dir <- system.file("extdata", package = "microbiome")
  # dat <- read.profiling(data.dir)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

