% Generated by roxygen2 (4.0.2): do not edit by hand
\name{relative.abundance}
\alias{relative.abundance}
\title{relative.abundance}
\usage{
relative.abundance(dat, det.th = 0)
}
\arguments{
\item{dat}{data matrix (phylotypes x samples) in original (non-log) scale}

\item{det.th}{detection threshold.

Returns:}
}
\value{
Vector containing relative proportions for each phylotype in
           each sample
}
\description{
Description: Estimate relative abundance for each phylotype in each
sample with a given threshold '
}
\details{
Arguments:
}
\examples{
data(peerj32);
         relab <- relative.abundance(10^t(peerj32$microbes),
                          det.th = 0)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

