% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative.abundance.R
\name{relative.abundance}
\alias{relative.abundance}
\title{Relative abundance}
\usage{
relative.abundance(dat, det.th = 0)
}
\arguments{
\item{dat}{data matrix (phylotypes x samples) in original (non-log) scale}

\item{det.th}{detection threshold.}
}
\value{
Vector containing relative proportions for each phylotype in 
        each sample
}
\description{
Estimate relative abundance for each phylotype in each sample 
with a given threshold.
}
\examples{
  data(peerj32)
  relab <- relative.abundance(10^t(peerj32$microbes), det.th = 0)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

