% Generated by roxygen2 (4.0.0): do not edit by hand
\name{remove.nested.groups}
\alias{remove.nested.groups}
\title{Description: remove nested groups from hierarchical clustering results;
             keep only the largest group when nested groups are detected}
\usage{
remove.nested.groups(groups)
}
\arguments{
\item{groups}{list of groups}
}
\value{
list of groups with nested groups filtered out
}
\description{
Description: remove nested groups from hierarchical clustering results;
             keep only the largest group when nested groups are detected
}
\examples{
#
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("microbiome")
}
\keyword{methods}

