\name{retrieve.probesets}
\alias{retrieve.probesets}
\title{retrieve.probesets}
\usage{
  retrieve.probesets(phylo, level = "species")
}
\arguments{
  \item{phylo}{data.frame with oligo - phylotype mapping
  information}

  \item{level}{phylotype level for probesets Returns:}
}
\value{
  A list. Probes for each phylotype.
}
\description{
  Description: List probes for each probeset
}
\details{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

