% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/phylo.R
\name{retrieve.probesets}
\alias{retrieve.probesets}
\title{retrieve.probesets}
\usage{
retrieve.probesets(phylogeny.info, level = "species", name = NULL)
}
\arguments{
\item{phylogeny.info}{data.frame with oligo - phylotype
mapping information}

\item{level}{phylotype level for probesets}

\item{name}{specify phylotypes to check (optional)}
}
\value{
A list. Probes for each phylotype.
}
\description{
List probes for each probeset
}
\examples{
phylogeny.info <- GetPhylogeny('HITChip')
  sets <- retrieve.probesets(phylogeny.info, 'species', 'Weissella confusa')
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

