\name{retrieve.probesets}
\alias{retrieve.probesets}
\title{Description: List probes for each probeset}
\usage{
  retrieve.probesets(phylo, level = "species")
}
\arguments{
  \item{phylo}{data.frame with oligo - phylotype mapping
  information}

  \item{level}{phylotype level for probesets Returns:}
}
\value{
  A list. Probes for each phylotype.
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

