% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ROC.R
\name{roc.plot}
\alias{roc.plot}
\title{Plot ROC curve}
\usage{
roc.plot(ordered.results, true.positives, line = FALSE, title = "")
}
\arguments{
\item{ordered.results}{Items ordered from best to worst according
to the test score.}

\item{true.positives}{known true positives}

\item{line}{Draw 45 angle line}

\item{title}{Title text}
}
\value{
Used for its side effects (plot)
}
\description{
Plot ROC curve
}
\examples{
data(peerj32)
  x <- unlist(peerj32$microbes[1,])
  res <- roc.plot(names(x), sample(names(x), 10))
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

