% Generated by roxygen2 (4.0.2): do not edit by hand
\name{roc}
\alias{roc}
\title{Description: ROC analysis returning true and false positive rates
along an ordered list}
\usage{
roc(ordered.results, true.positives)
}
\arguments{
\item{ordered.results}{Items ordered from best to worst according
to the test score.}

\item{true.positives}{known true positives

Returns:}
}
\value{
List: true positive rate (tpr) and false positive rate (fpr)
}
\description{
Arguments:
}
\examples{
data(peerj32);
          x <- unlist(peerj32$microbes[1,]);
          res <- roc(names(x), sample(names(x), 10))
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
To cite microbiome R package, see citation('microbiome')
}
\keyword{utilities}

