\name{simulate.hitchip}
\alias{simulate.hitchip}
\title{Description: simulate.hitchip}
\usage{
  simulate.hitchip(PH.i, Ns, level = "L2", N = 5000)
}
\arguments{
  \item{PH.i}{oligomap}

  \item{Ns}{Ns}

  \item{level}{taxonomic level}

  \item{N}{N

  Returns:}
}
\value{
  list
}
\description{
  For cross-hyb control
}
\details{
  Arguments:
}
\author{
  Contact: Jarkko Salojarvi
  \email{jarkko.salojarvi@helsinki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

