\name{species2levels}
\alias{species2levels}
\title{Description: species-levels mappings}
\usage{
  species2levels(spec, phylogeny.info)
}
\arguments{
  \item{spec}{species}

  \item{phylogeny.info}{phylogeny.info

  Returns:}
}
\value{
  species-levels mappings
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

