% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{species2levels}
\alias{species2levels}
\title{Species-levels mappings}
\usage{
species2levels(spec, phylogeny.info)
}
\arguments{
\item{spec}{species}

\item{phylogeny.info}{phylogeny.info}
}
\value{
species-levels mappings
}
\description{
Species-levels mappings
}
\author{
Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation('microbiome')
}
\keyword{internal}

