\name{Strip}
\alias{Strip}
\title{Strip string i.e. remove spaces from the beginning and end}
\usage{
  Strip(s)
}
\arguments{
  \item{s}{string or character vector}
}
\value{
  Stripped string
}
\description{
  Strip string i.e. remove spaces from the beginning and
  end
}
\examples{
#s2 <- Strip(s)
}
\author{
  Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

