\name{submitDB.sampleOutliers}
\alias{submitDB.sampleOutliers}
\title{Delete previous outlier detection results for these samples. Also set
the reproducibility-check flag to 0 for all featureextractions belonging
to these samples.}
\usage{
  submitDB.sampleOutliers(x)
}
\arguments{
  \item{x}{TBA}
}
\value{
  TBA
}
\description{
  Delete previous outlier detection results for these
  samples. Also set the reproducibility-check flag to 0 for
  all featureextractions belonging to these samples.
}
\examples{
# TBA
}
\author{
  Douwe Molenaar. Maintainer: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

