\name{summarize.oligos}
\alias{summarize.oligos}
\title{Description: summarize.oligos}
\usage{
  summarize.oligos(oligo.matrix, oligomap, level = "L2")
}
\arguments{
  \item{oligo.matrix}{oligo.matrix}

  \item{oligomap}{oligomap}

  \item{level}{taxonomic level

  Returns:}
}
\value{
  list
}
\description{
  Arguments:
}
\author{
  Contact: Jarkko Salojarvi
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

