\name{threshold.data}
\alias{threshold.data}
\title{Description: determine detection threshold for the data}
\usage{
  threshold.data(dat, sd.times = 6)
}
\arguments{
  \item{dat}{data}

  \item{sd.times}{standard deviation threshold

  Returns:}
}
\value{
  thresholded data matrix
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

