% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesplit.R
\name{timesplit}
\alias{timesplit}
\title{Time Split}
\usage{
timesplit(x)
}
\arguments{
\item{x}{\pkg{phyloseq} object.}
}
\value{
data.frame
}
\description{
For each subject, return temporally consecutive sample pairs
  together with the corresponding time difference.
}
\examples{
data(atlas1006)
x <- timesplit(subset_samples(atlas1006,
  DNA_extraction_method == 'r' & sex == "male"))
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
