\name{top.barplots}
\alias{top.barplots}
\title{Description: Barplot of top findings from pairwise.comparisons function}
\usage{
  top.barplots(top.findings, topN = 5, annot)
}
\arguments{
  \item{top.findings}{Ouput from pairwise.comparisons
  function}

  \item{topN}{number of top findings to plot}

  \item{annot}{annotation matrix

  Returns:}
}
\value{
  List with top findings from pairwise comparisons and
  their q-values
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

