\docType{data}
\name{toydata-hitchip}
\alias{toydata-hitchip}
\title{toydata-hitchip}
\description{
  The toydata-hitchip contains the following simulated data
  matrices:
}
\details{
  metadata.simulated sample metadata
  oligo.matrix.nolog.simulated: oligo-level data matrix
  (oligos vs. samples) genus.matrix.log10.simulated: genus
  (L2) level data matrix (phylotypes vs. samples)

  Generate with GenerateSimulatedData function
}
\author{
  Leo Lahti \email{microbiome-devel@googlegroups.com}
}
\references{
  \url{microbiome.github.com}
}
\keyword{data}
\keyword{datasets}

