% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform}
\alias{transform}
\title{Data Transformations for phyloseq Objects}
\usage{
transform(x, transform = "identity", target = "OTU")
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{transform}{Transformation to apply. The options include:
'compositional' (ie relative abundance), 'Z', 'log10', 'hellinger',
'identity', 'clr', 'ilr',
 or any method from the vegan::decostand function.}

\item{target}{Apply the transform for 'sample' or 'OTU'.
Does not affect the log transform.}
}
\value{
Transformed \code{\link{phyloseq}} object
}
\description{
Standard transforms for \code{\link{phyloseq-class}}.
}
\details{
The relative abunance are returned as percentages in [0,
  1]. The Hellinger transform is square root of the relative
  abundance but instead given at the scale [0,1].
}
\examples{
\dontrun{

  # OTU relative abundances
  xt <- transform(x, "relative.abundance", "OTU")

  # Z-transform for OTUs
  xt <- transform(x, "Z", "OTU")

  # Z-transform for samples
  xt <- transform(x, "Z", "sample")

  # Log10 transform (log(1+x) if the data contains zeroes)
  xt <- transform(x, "log10")

}
}
\keyword{utilities}
