% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_phyloseq}
\alias{transform_phyloseq}
\title{Standard phyloseq transformations}
\usage{
transform_phyloseq(x, transformation = "relative.abundance", target = "OTU")
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{transformation}{Transformation to apply: 'relative.abundance', 'Z', or 'log10'.}

\item{target}{Apply the transformation for 'sample' or 'OTU'. Does not affect the log transformation.}
}
\value{
Transformed \code{\link{phyloseq}} object
}
\description{
Transform phyloseq objects with standard transformations including log10(x), log10(1+x), z, relative abundance.
}
\details{
The relative abunance are returned as percentages in [0, 100].
}
\examples{
\dontrun{

  # OTU relative abundances
  xt <- transform_phyloseq(x, "relative.abundance", "OTU")

  # Z-transformation for OTUs
  xt <- transform_phyloseq(x, "Z", "OTU")

  # Z-transformation for samples
  xt <- transform_phyloseq(x, "Z", "sample")

  # Log10 transformation (log(1+x) if the data contains zeroes)
  xt <- transform_phyloseq(x, "log10")

}
}
\keyword{utilities}

