% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stability.R
\name{variability_individual}
\alias{variability_individual}
\title{variability_individual}
\usage{
variability_individual(dat, meta, method = "timevar")
}
\arguments{
\item{dat}{Data matrix (samples x features)}

\item{meta}{Metadata (samples x variables) with the following
fields: subject, time}

\item{method}{Method to calculate the variability within
             individuals: sd (ignores time) or timevar (normalizes by time)}
}
\value{
Vector. Average variability within individuals for each feature.
}
\description{
Average CoV within individuals for each feature in the data matrix
}
\examples{
#library(microbiome)
  #data(peerj32)
  #cv <- variability_individual(peerj32$microbes, peerj32$meta)
  #print(head(cv))
}

