% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_phyloseq.R
\name{write_phyloseq}
\alias{write_phyloseq}
\title{Exporting phyloseq Data in CSV Files}
\usage{
write_phyloseq(x, type = "all", path = getwd())
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{type}{'OTU' or 'TAXONOMY' or 'METADATA'}

\item{path}{Path to the directory/folder where the data will be written.
Uses the working directory by default.}
}
\value{
Output file path (a string)
}
\description{
Writes the otu, taxonomy and metadata in csv files.
}
\examples{
\dontrun{
#data(dietswap)
#pseq <- dietswap
## By default writes all info at once (ie OTU/TAXONOMY/METADATA)
#write_phyloseq(pseq) 
#write_phyloseq(pseq, 'OTU')
#write_phyloseq(pseq, 'TAXONOMY')
#write_phyloseq(pseq, 'METADATA')
}
}
\seealso{
read_phyloseq
}
\keyword{utilities}
