\name{WriteChipData}
\alias{WriteChipData}
\title{Description: Write preprocessed data into the output directory}
\usage{
  WriteChipData(finaldata, output.dir, oligomap,
    verbose = TRUE)
}
\arguments{
  \item{finaldata}{preprocessed data matrices in absolute
  scale (from the preprocess.chipdata function)}

  \item{output.dir}{output directory}

  \item{oligomap}{oligomap}

  \item{verbose}{verbose

  Returns:}
}
\value{
  Preprocessed data in absolute scale, oligomap, and
  parameters
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

