\name{WriteChipData}
\alias{WriteChipData}
\title{Description: Writed data into the output directory}
\usage{
  WriteChipData(finaldata, output.dir, phylogeny.info,
    verbose = TRUE)
}
\arguments{
  \item{finaldata}{preprocessed data matrices in absolute
  scale (from the chipdata function)}

  \item{output.dir}{output directory}

  \item{phylogeny.info}{phylogeny.info}

  \item{verbose}{verbose

  Returns:}
}
\value{
  Preprocessed data in absolute scale, phylogeny.info, and
  parameters
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

