\name{WriteLog}
\alias{WriteLog}
\title{Description: Write log file}
\usage{
  WriteLog(naHybs, params)
}
\arguments{
  \item{naHybs}{hybridisation that were removed due to NAs}

  \item{params}{parameters

  Returns:}
}
\value{
  List of scaling methods
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

