\name{WriteMatrix}
\alias{WriteMatrix}
\title{Description: Write matrix in tab file}
\usage{
  WriteMatrix(dat, filename, verbose = FALSE)
}
\arguments{
  \item{dat}{data matrix}

  \item{filename}{output file}

  \item{verbose}{verbose Returns:}
}
\value{
  output file location
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti
  \email{microbiome-admin@googlegroups.com}
}
\references{
  See citation("microbiome")
}
\keyword{utilities}

