% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{ztransform}
\alias{ztransform}
\title{Z Transformation}
\usage{
ztransform(x, which, log10 = TRUE)
}
\arguments{
\item{x}{a matrix}

\item{which}{margin}

\item{log10}{apply log10 transformation before Z}
}
\value{
Z-transformed matrix
}
\description{
Z transform for matrices
}
\details{
Performs centering (to zero) and scaling (to unit
  variance) across samples for each taxa.
}
\examples{
#data(peerj32)
#pseqz <- ztransform(abundances(peerj32$phyloseq))
}
\references{
See citation('microbiome')
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\keyword{internal}
