% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztransform_phyloseq.R
\name{ztransform_phyloseq}
\alias{ztransform_phyloseq}
\title{Phyloseq z transformation}
\usage{
ztransform_phyloseq(x, which)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{which}{Specify Z transformation for "sample" or "OTU"}
}
\value{
Z-transformed phyloseq object
}
\description{
Z transform phyloseq objects.
}
\details{
Performs centering (to zero) and scaling (to unit
variance) across samples for each taxa.
}
\examples{
  data(peerj32)
  pseqz <- ztransform_phyloseq(peerj32$physeq)
}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

