% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztrans.R
\name{ztransform_phyloseq}
\alias{ztransform_phyloseq}
\title{ztransform_phyloseq}
\usage{
ztransform_phyloseq(x, which)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{which}{Specify Z transformation for "sample" or "OTU"}
}
\value{
Z-transformed phyloseq object
}
\description{
Z transform phyloseq objects
}
\examples{
  pseq <- download_microbiome("peerj32")$physeq
  pseqz <- ztransform_phyloseq(pseq, "OTU")

}
\author{
Contact: Leo Lahti \email{microbiome-admin@googlegroups.com}
}
\references{
See citation('microbiome')
}
\keyword{utilities}

